Datasets used in examples

Dataset : Bauges

Steps :

1. find dominant species and build PFG

2. prepare parameter files for a FATE simulation

Base de données du Conservatoire Botanique National Alpin - 2018

data(DATASET_Bauges_PFG)
data(DATASET_Bauges_parameters)

Format

>> DATASET_Bauges_PFG <<

A list object with 7 elements to help finding dominant species and building Plant Functional Groups in Bauges area :

sp.names

a data.frame of dimension 2913 x 3
containing information about names of plant species in Bauges

species : species ID

GENUS : species' genus

NAME : species' full name

sp.observations

a data.frame of dimension 288654 x 6
containing releves data about plant species in Bauges
to be used with the PRE_FATE.selectDominant function

sites : sites ID

species : species ID

abund : relative abundance obtained with the PRE_FATE.abundBraunBlanquet function

X : x-axis coordinates in Lambers (lcc)

Y : y-axis coordinates in Lambers (lcc)

habitat : habitat ID

1 : open areas (fields, lawns, meadows...)

2 : closed areas (forests...)

3 : urban areas

dom.traits

data.frame of dimension 284 x 7
containing traits for dominant species
to be used with the PRE_FATE.speciesDistance function

species : species ID

GROUP : rough generalization of Raunkier life-forms :

Herbaceous

Chamaephyte

Phanerophyte

HEIGHT : mean plant height (cm)

DISPERSAL : classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

LIGHT : Flora Indicativa value for light preference (from 1 to 5)

NITROGEN : Flora Indicativa value for soil fertility preference (from 1 to 5)

MOISTURE : Flora Indicativa value for soil moisture preference (from 1 to 5)

dom.dist_overlap

matrix of dimension 275 x 275
containing niche overlap distance for dominant species
to be used with the PRE_FATE.speciesDistance function

dom.dist_total

list of three matrix containing (traits + niche overlap) distance for dominant species :

Herbaceous : dimension 204 x 204

Chamaephyte : dimension 28 x 28

Phanerophyte : dimension 30 x 30

to be used with the PRE_FATE.speciesClustering_step1 function

dom.determ

data.frame of dimension 262 x 10
containing dominant species information relative to PFG
obtained from the PRE_FATE.speciesClustering_step2 function

PFG : Plant Functional Group name

GROUP : rough generalization of Raunkier life-forms :

Herbaceous

Chamaephyte

Phanerophyte

ID.cluster : Plant Functional Group ID

species : species ID

ID : species ID within its PFG

sp.mean.dist :

allSp.mean :

allSp.min :

allSp.max :

DETERMINANT : is the species kept as determinant species within the PFG ? (see PRE_FATE.speciesClustering_step2 function for details)

PFG.observations

a data.frame of dimension 31666 x 20
containing gathered releves data by PFG for each releve site

PFG.traits

data.frame of dimension 20 x 8
containing traits for plant functional groups
obtained from the PRE_FATE.speciesClustering_step3 function

PFG : Plant Functional Group name

no.species : number of dominant species within each Plant Functional Group

GROUP : rough generalization of Raunkier life-forms :

Herbaceous

Chamaephyte

Phanerophyte

HEIGHT : mean value

DISPERSAL : median value

LIGHT : median value

NITROGEN : median value

MOISTURE : median value

>> DATASET_Bauges_parameters <<

A list object with 7 elements to help building parameter files in order to run a FATE simulation :

dom.traits

a data.frame of dimension 125 x 12
containing traits for dominant species

species : species ID

NAME : species name

PFG : Plant Functional Group name

type : rough generalization of Raunkier life-forms :

Herbaceous

Chamaephyte

Phanerophyte

maturity

longevity

height

dispersal :classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

light : Flora Indicativa value for light preference (from 1 to 5)

soil_contrib : Flora Indicativa value for soil fertility (from 1 to 5)

soil_tolerance : Flora Indicativa value for soil tolerance (either 0.2 for small or 1.0 for high)

dist_tolerance : tolerance to disturbance (grazing, mowing) (from 1 to 5)

PFG.traits

a data.frame of dimension 16 x 12
containing traits for Plant Functional Groups

PFG : Plant Functional Group name

no.species : number of dominant species within each Plant Functional Group

type : rough generalization of Raunkier life-forms :

Herbaceous

Chamaephyte

Phanerophyte

... (same traits as dom.traits)

tab.succ

a data.frame of dimension 16 x 5 containing data from PFG.traits (PFG, type, height, maturity, longevity)
to be used with the function PRE_FATE.params_PFGsuccession

tab.light

a data.frame of dimension 16 x 3 containing data from PFG.traits (PFG, type, light)
to be used with the function PRE_FATE.params_PFGlight

tab.soil

a data.frame of dimension 16 x 5 containing data from PFG.traits (PFG, type, soil_contrib, soil_tol_min, soil_tol_max)
to be used with the function PRE_FATE.params_PFGsoil

tab.disp

a data.frame of dimension 16 x 3 containing data from PFG.traits (PFG, type, dispersal)
to be used with the function PRE_FATE.params_PFGdispersal

tab.dist

a data.frame of dimension 16 x 3 containing data from PFG.traits (PFG, type, dist_tolerance)
to be used with the function PRE_FATE.params_PFGdisturbance

Author

Maya Guéguen