FATE Wrapper

This function runs a FATE dynamical landscape vegetation simulation.

FATE(simulParam, no_CPU = 1L, verboseLevel = 0L)

Arguments

simulParam

a string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

no_CPU

(optional) default 1.
The number of resources that can be used to parallelize the simulation

verboseLevel

(optional) default 0.
The logger verbose level : a FATE simulation can render different levels of information (from 0 to 4, see Details).

Value

None

Details

This function allows to run a vegetation simulation with the FATE model, based on a simulation folder and a species simulation parameter file.

A FATE simulation can be parallelized, using the no_CPU parameter, given that the user machine is multi-core !

Quantity of informations are rendered by the software into the R console, and the verboseLevel parameter allows to filter which level of information is printed :

0.

shows any message

1.

shows any message, except debug

2.

shows only warning and error messages

3.

shows only error messages

4.

mute

Author

Damien Georges, Isabelle Boulangeat, Maya Guéguen

Examples

if (FALSE) FATE()