Ecrins National Park (PNE) data : building of Plant Functional Groups

This dataset contains elements to create the Plant Functional Groups (PFG) over the Ecrins National Park (PNE) (Boulangeat et al. 2012 GCB).

Format

A list object with 6 elements to help building the Plant Functional Group :

sp.observations

a data.frame of dimension 168313 x 7
containing releves data about plant species in the PNE
to be used with the PRE_FATE.selectDominant function

sites : sites ID

X : x-axis coordinates in Lambers (lcc)

Y : y-axis coordinates in Lambers (lcc)

habitat : habitat ID from DELPHINE :

0 : glaciers and eternal snows

31 : uncolonized rocks

40 : lawns and meadows

50 : low moors

60 : open areas, brush

70 : semi-closed areas

81 : closed areas

83 : forests

species : species ID

abund_BB : Braun-Blanquet abundance
(see PRE_FATE.abundBraunBlanquet function for details)

abund : relative abundance obtained from the abund_BB column with the PRE_FATE.abundBraunBlanquet function

dom.traits

data.frame of dimension 359 x 6
containing traits for dominant species
to be used with the PRE_FATE.speciesDistance function

species : species ID

GROUP : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

height : mean plant height (cm)

dispersal : classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

palatability : classes (from 0 to 3) (CBNA)

light : Ellenberg indicator value for light (from 1 to 9)

dom.dist_overlap

matrix of dimension 358 x 358
containing niche overlap distance for dominant species
to be used with the PRE_FATE.speciesDistance function

nb.clusters

vector of length 3
number of groups kept for each life-form class, obtained by cutting the hierarchical tree obtained from species distances

dom.determ

data.frame of dimension 359 x 5
containing dominant species information relative to PFG
obtained from the PRE_FATE.speciesClustering_step2 function

species : species ID

name : species name (taxonomic)

GROUP : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

PFG : name of assigned Plant Functional Group

determinant : is the species kept as determinant species within the PFG ? (see PRE_FATE.speciesClustering_step2 function for details)

PFG.traits

data.frame of dimension 24 x 14
containing traits for plant functional groups
obtained from the PRE_FATE.speciesClustering_step3 function

PFG_name : full descriptive Plant Functional Group name

PFG : Plant Functional Group short name

type : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

strata : maximum height stratum that can be reached by the PFG (from 1 to 5)

disp : MEDIAN classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

light : MEDIAN Ellenberg indicator value for light (from 1 to 9)

height : MEAN PFG height (cm)

palatability : MEDIAN classes (from 0 to 3) (CBNA)

longevity : MEAN age of lifespan

maturity : MEAN age of maturity

Source