Create a graphical representation of the evolution of PFG abundance through time for 5 (or more) pixels of a FATE simulation

This script is designed to produce one graphical representation for a FATE simulation : the evolution through time of the abundance of each PFG for 5 (or more) randomly selected cells of the studied area.

POST_FATE.graphic_evolutionPixels(
  name.simulation,
  file.simulParam = NULL,
  opt.cells_ID = NULL,
  opt.doPlot = TRUE
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

file.simulParam
default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

opt.cells_ID
(optional) default NULL.
The cells ID of the studied area for which PFG abundances will be extracted

opt.doPlot
(optional) default TRUE.
If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned

Value

A list containing one data.frame object with the following columns, and one ggplot2 object :

tab

TYPE

concerned information (either 'light', 'abundance' or 'soil')

GROUP

concerned entity (either 'STRATUM_[...]', PFG name or 'soil')

ID.pixel

number of the concerned pixel

HAB

habitat of the concerned pixel

YEAR

concerned simulation year

value

concerned value extracted from .csv files produced by POST_FATE.temporalEvolution

plot

ggplot2 object, representing the evolution of each PFG abundance, and light and soil resources if those modules were activated

One POST_FATE_TABLE_PIXEL_evolution_pixels_[...].csv file is created :

pixels ids

always, containing the data.frame detailed above

One POST_FATE_[...].pdf file is created :

GRAPHIC_A
pixels

to visualize for each PFG the evolution of its abundance within each selected pixel through simulation time

Details

This function allows to obtain, for a specific FATE simulation and a specific parameter file within this simulation, one preanalytical graphic :

the evolution of abundance of each Plant Functional Group through simulation time, within 5 (or more) randomly selected pixels of the studied area (FATE arbitrary unit)

if light was activated (see PRE_FATE.params_globalParameters),
evolution of light resources within the selected pixels is also represented (1: Low, 2: Medium, 3: High)

if soil was activated (see PRE_FATE.params_globalParameters),
evolution of soil resources within the selected pixels is also represented (user-defined scale)

It requires that the POST_FATE.temporalEvolution function has been run and that the file POST_FATE_TABLE_PIXEL_evolution_abundance.csv exists (as well as the POST_FATE_TABLE_PIXEL_evolution_light.csv and POST_FATE_TABLE_PIXEL_evolution_soil.csv files if those modules were activated).

See also

Author

Maya Guéguen