Create a map of the Plant Functional Group cover
for one (or several) specific year of a FATE-HD simulation

This script is designed to produce a raster map of PFG cover for one (or several) specific FATE-HD simulation year.

POST_FATE.graphic_mapPFGcover(name.simulation, file.simulParam = NULL,
  year, strata_min, opt.mat.cover.obs = NULL, opt.ras.cover.obs = NULL,
  opt.no_CPU = 1, opt.doPlot = TRUE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG abundance and binary maps

strata_min
an integer corresponding to the lowest stratum from which PFG abundances are summed up to the highest stratum

opt.mat.cover.obs
default NULL (optional). A data.frame with 3 columns : X, Y, obs

opt.ras.cover.obs
default NULL (optional). A string that corresponds to the file name of a raster containing observed values for vegetation cover

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

opt.doPlot
default TRUE (optional). If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned.

Value

One POST_FATE_[...].pdf file is created :

GRAPHIC_B
PFGcover

to visualize the PFG cover within the studied area

A raster file is created into simulation results folder.

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, one preanalytical graphic.

For each PFG and each selected simulation year, raster maps are retrieved from the results folders ABUND_perPFG_perStrata and BIN_perPFG_perStrata and unzipped. Informations extracted lead to the production of one graphic before the maps are compressed again :

the map of Plant Functional Group cover for each selected simulation year(s), representing the relative cumulated abundance of PFG present in each pixel above a height threshold

See also

Examples

if (FALSE) {
POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = 850
                              , strata_min = 3
                              , opt.no_CPU = 1)

POST_FATE.graphic_mapPFGcover(name.simulation = "FATE_simulation"
                              , file.simulParam = "Simul_parameters_V1.txt"
                              , year = c(850, 950)
                              , strata_min = 3
                              , opt.no_CPU = 1)
}