Create presence / absence (binary) maps for each Plant Functional Group
for one (or several) specific year of a FATE-HD simulation

This script is designed to produce raster maps of PFG simulated presences and absences for one (or several) specific FATE-HD simulation year.

POST_FATE.relativeAbund_presenceAbsence(name.simulation,
  file.simulParam = NULL, year, strata_min = 1, opt.no_CPU = 1)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG abundance maps

strata_min
an integer corresponding to the lowest stratum for which PFG relative abundances are calculated and then transformed into binary results

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

Value

Two result folders are created :

BIN_perPFG
_perStrata

containing presence/absence raster maps for each PFG within each stratum

BIN_perPFG
_allStrata

containing presence/absence raster maps for each PFG across all strata

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, raster maps of PFG presence/absence.

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_perPFG_perStrata and unzipped. Informations extracted lead to the production of the same number of raster before the maps are compressed again :

for each selected simulation year(s), relative abundances within each stratum are calculated : $$\frac{\text{Abund } PFG_i \text{ } Stratum_j} {\Sigma \text{ Abund } PFG_{all} \text{ } Stratum_j}$$

binary maps for each PFG within each stratum are obtained by converting relative abundances in 0 if < 5 %, in 1 otherwise.

binary maps for each PFG within each pixel are obtained from all binary maps of a PFG within all strata :
if the PFG is present within one stratum, it is considered as present (1) within the pixel ;
it is considered absent otherwise (0).

See also

Examples

# NOT RUN {
# }
# NOT RUN {
POST_FATE.relativeAbund_presenceAbsence(name.simulation = "FATE_simulation"
                                        , file.simulParam = "Simul_parameters_V1.txt"
                                        , year = 850
                                        , strata_min = 1
                                        , opt.no_CPU = 1)

POST_FATE.relativeAbund_presenceAbsence(name.simulation = "FATE_simulation"
                                        , file.simulParam = "Simul_parameters_V1.txt"
                                        , year = c(850, 950)
                                        , strata_min = 1
                                        , opt.no_CPU = 1)
# }# NOT RUN {
                                    
                                    

# }