Create tables of pixel temporal evolution of PFG abundances (and light and soil resources if activated) for a FATE simulation

This script is designed to produce from 1 to 3 tables containing pixel temporal evolution of PFG abundances, as well as light and soil resources if those modules were activated, in a FATE simulation.

POST_FATE.temporalEvolution(
  name.simulation,
  file.simulParam = NULL,
  no_years,
  opt.ras_habitat = NULL,
  opt.no_CPU = 1
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

file.simulParam
default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

no_years
an integer corresponding to the number of simulation years that will be used to extract PFG abundance / light / soil maps

opt.ras_habitat
(optional) default NULL.
A string corresponding to the file name of a raster mask, with an integer value within each pixel, corresponding to a specific habitat

opt.no_CPU
(optional) default 1.
The number of resources that can be used to parallelize the unzip/zip of raster files, as well as the extraction of values from raster files

Value

A list containing three data.frame objects with the following columns :

PFG

concerned plant functional group (for abundance)

STRATUM

concerned height stratum (for LIGHT)

ID.pixel

number of the concerned pixel

X, Y

coordinates of the concerned pixel

HAB

habitat of the concerned pixel

years

values of the corresponding object (abundance / LIGHT / SOIL) for each selected simulation year(s)

One to three POST_FATE_TABLE_PIXEL_evolution_[...].csv files are created :

abundance

always

light

if light module was activated

soil

if soil module was activated

Details

This function allows to obtain, for a specific FATE simulation and a specific parameter file within this simulation, one to three preanalytical tables that can then be used to create graphics.

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_perPFG_allStrata and unzipped. Informations extracted lead to the production of one table before the maps are compressed again :

the value of abundance for each Plant Functional Group for each selected simulation year(s) in every pixel in which the PFG is present for at least one of the selected simulation year(s)

If the light module was activated (see PRE_FATE.params_globalParameters), for each height stratum and each selected simulation year, raster maps are retrieved from the results folder LIGHT and unzipped. Informations extracted lead to the production of one table before the maps are compressed again :

the value of light resources for each height stratum for each selected simulation year(s) in every pixel

If the soil module was activated (see PRE_FATE.params_globalParameters), for each selected simulation year, raster maps are retrieved from the results folder SOIL and unzipped. Informations extracted lead to the production of one table before the maps are compressed again :

the value of soil resources for each selected simulation year(s) in every pixel

If a raster mask for habitat has been provided, the tables will also contain information about the pixel habitat.

These .csv files can then be used by other functions :

to produce graphics of temporal evolution of modelled abundances and space occupancy at the whole area level
(see POST_FATE.graphic_evolutionCoverage)

to produce graphics of temporal evolution of modelled abundances and / or resources at the pixel level
(see POST_FATE.graphic_evolutionPixels)

to produce graphics of temporal evolution of community composition at the habitat level
(see POST_FATE.graphic_evolutionStability)

See also

Author

Maya Guéguen

Examples

if (FALSE) {                      
POST_FATE.temporalEvolution(name.simulation = "FATE_simulation"
                            , file.simulParam = "Simul_parameters_V1.txt"
                            , no_years = 50
                            , opt.no_CPU = 1)
}
                                                        
                                                        
                                                        
## ----------------------------------------------------------------------------------------- ##
## Load example data