Compute validation data for habitat, PFG richness and composition for a FATE simulation.

This script is designed to compare observed and simulated PFG distribution for one specific FATE simulation year by computing a) difference in PFG richness ; b) similarity between distribution quantiles ; c) a random forest model to predict habitat classes.

POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  mat.obs,
  mat.hab,
  ras.habitat,
  doRichness = TRUE,
  doComposition = TRUE,
  doHabitat = TRUE,
  RF.seed = 123,
  RF.training = 0.7,
  doHabitat.allMap = FALSE,
  opt.ras_validation = NULL,
  opt.keep_PFG = NULL,
  opt.keep_strata = NULL
)

Arguments

name.simulation

a string corresponding to the main directory or simulation name of the FATE simulation

file.simulParam

default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

year

an integer corresponding to the simulation year that will be used to extract PFG abundance table

mat.obs

a data.frame with at least 5 columns :
site, x, y, PFG, abund
(and optionally, strata, code.habitat) (see Details)

mat.hab

a data.frame with 2 columns : code.habitat, habitat

ras.habitat

a string corresponding to the file name of a raster mask, with an integer value within each pixel, corresponding to a specific habitat

doRichness

(logical) default TRUE.
If TRUE, difference in observed and simulated PFG richness will be run.

doComposition

(logical) default TRUE.
If TRUE, difference in observed and simulated PFG distribution over habitat will be run.

doHabitat

(logical) default TRUE.
If TRUE, a random forest model to predict habitat from PFG abundances will be run.

RF.seed

default 123.
An integer to be given to set.seed function, in order to fix the produced results if needed, as dataset will be divided randomly into training and testing datasets for the random forest model

RF.training

default 0.7.
A numeric between 0 and 1 corresponding to the percentage of data that will be used as training dataset in the random forest model

doHabitat.allMap

(logical) default TRUE.
If TRUE, habitat prediction over the whole simulation map will be run.

opt.ras_validation

(optional) default NULL.
A string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the validation will take place

opt.keep_PFG

(optional) default NULL.
A vector of character corresponding to the names of the PFG to keep for the validation

opt.keep_strata

(optional) default NULL.
A list with names corresponding to the strata to keep for the validation, and each element containing correspondence with FATE strata definition

Value

A list containing two to ten elements depending on the options selected :

PFG richness

three vector objects :

rich.obs

vector containing names of PFG in observed data

rich.sim

vector containing names of PFG in simulated data

rich.diff

vector containing names of PFG in observed but not in simulated data

PFG composition

compo.distrib

data.frame object with the following columns :

PFG

concerned plant functional group

code.habitat

concerned habitat code

strata

concerned height stratum

quantile.perc

concerned quantile

quantile.obs

corresponding observed abundance

quantile.sim

corresponding simulated abundance

compo.proximity

data.frame object with the following columns :

PFG

concerned plant functional group

code.habitat

concerned habitat code

strata

concerned height stratum

proximity

composition similarity between each habitat/strata combination

Habitat

hab.RF.model

randomForest object obtained from randomForest function

hab.RF.perf

data.frame object with the following columns :

dataset

train, test or valid

habitat

concerned habitat code

sensitivity

true positive rate

specificity

true negative rate

weight

share of each habitat in the observed habitat distribution

TSS

mean True Skill Statistic

TSSw

mean weighted True Skill Statistic

hab.tab.pred

data.frame object with the following columns :

pixel

concerned pixel

x

corresponding x-coordinate

y

corresponding y-coordinate

code.habitat

concerned habitat code

habitat.obs

observed habitat

habitat.sim

simulated habitat through random forest model

habitat.final

simulated habitat through random forest model, with wrong predictions indicated as failure

fail_succ

wether or not the simulated habitat is the same as the observed

color

corresponding color code

hab.ras.pred

raster object containing habitat predictions

hab.plot

ggplot2 object, representing hab.ras.pred raster

One to four POST_FATE_TABLE_[...].csv files are created :

HAB_validation_compo_distribution_

if composition module was activated, containing compo.distrib

HAB_validation_compo_proximity_

if composition module was activated, containing compo.proximity

HAB_validation_RF_performance_

if habitat module was activated, containing hab.RF.perf

PIXEL_validation_RF_prediction_

if habitat module was activated, containing hab.tab.pred

HabitatPrediction_YEAR_[...].tif and POST_FATE_GRAPHIC_D_map_habitat_[...].pdf files are created containing random forest predicted habitat raster and plot respectively.

Details

This function allows to obtain, for a specific FATE simulation and a specific parameter file within this simulation, PFG validation habitat model and predictions and one preanalytical graphic.

PFG richness

check if all simulated PFG abundances are superior to 0 within the simulated area (only a subset of PFG can be examined with the opt.keep_PFG parameter).

PFG composition

for each PFG and each habitat / stratum combination, abundances for each quartiles (25, 50, 75 and 100 %) is calculated for both observed and simulated distributions.
(Only a subset of PFG and strata can be examined with the opt.keep_PFG and opt.keep_strata parameters.)

Then, a composition similarity between each PFG / habitat / strata combination is calculated as a pseudo-distance between observed and simulated quartiles Q : $$S_{\text{ Habitat}_j \text{, }\text{Stratum}_k} = \sum S_{\text{ PFG}_i \text{, }\text{Habitat}_j \text{, }\text{Stratum}_k}$$ with $$S_{\text{ PFG}_i \text{, }\text{Habitat}_j \text{, }\text{Stratum}_k} = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ q} \text{, }sim} - Q_{\text{ q} \text{, }obs})$$ with q varying from 1 to 4.

Habitat validation

a randomForest model is built to predict habitat (provided through ras.habitat) in function of observed PFG abundances. Only habitat code in mat.hab are effectively used and predicted. TSS (True Skill Statistic) is computed between observed and simulated habitat, and averaged over each habitat. TSS can also be weighted by the share of each habitat in the observed habitat distribution.

Author

Matthieu Combaud, Maxime Delprat, Maya Guéguen

Examples

library(raster)

## Create a simulation folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")
.loadData("Champsaur_results_V1", "7z")

## Please extract results files in the 'FATE_Champsaur/RESULTS' folder

## Define a vector to choose habitats taken into account
mat.hab = data.frame(ID = c(6, 5, 7, 8)
                     , habitat = c("coniferous.forest"
                                   , "deciduous.forest"
                                   , "natural.grassland"
                                   , "woody.heatland"))

## Habitat & validation maps
ras_simulation = Champsaur_params$stk.mask$Champsaur
ras.habitat = Champsaur_params$stk.mask$habitat
ras.habitat = projectRaster(from = ras.habitat, to = ras_simulation, method = "ngb")
# writeRaster(ras_simulation, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
mat.obs = Champsaur_params$tab.releves

## Transform observed PFG abundances into relative abundances
mat.obs$abund = PRE_FATE.abundBraunBlanquet(mat.obs$abund) / 100
mat.obs = aggregate(abund ~ site + PFG + strata + x + y
                    , data = mat.obs, FUN = "sum")
                        
## Create Global and Simulation parameters
PRE_FATE.params_globalParameters(name.simulation = "FATE_Champsaur"
                                 , opt.saving_abund_PFG_stratum = TRUE
                                 , opt.saving_abund_PFG = TRUE
                                 , opt.saving_abund_stratum = FALSE
                                 , required.no_PFG = 15
                                 , required.no_strata = 7
                                 , required.simul_duration = 2000
                                 , required.seeding_duration = 1000
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.potential_fecundity = 1
                                 , required.max_abund_low = 1000
                                 , required.max_abund_medium = 2000
                                 , required.max_abund_high = 3000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , DISPERSAL.saving = FALSE
                                 , doHabSuitability = TRUE
                                 , HABSUIT.mode = 1)
                                
PRE_FATE.params_simulParameters(name.simulation = "FATE_Champsaur"
                                , name.MASK = "MASK_Champsaur.tif")

simul.param = "Simul_parameters_V1.txt"
# simul.param = paste0("FATE_Champsaur/PARAM_SIMUL/", simul.param)

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , file.simulParam = simul.param
                     , year = 2000
                     , mat.obs = mat.obs
                     , mat.hab = mat.hab
                     , ras.habitat = ras.habitat
                     , doHabitat = TRUE
                     , doHabitat.allMap = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)