Compute habitat performance and create a prediction plot of habitat for a whole map of a FATE simulation.

This script compare habitat simulations and observations and create a map to visualize this comparison with all the FATE and observed data.

POST_FATE.validation_habitat(
  name.simulation,
  sim.version,
  obs.path,
  releves.PFG,
  releves.sites,
  hab.obs,
  validation.mask
)

Arguments

name.simulation

simulation folder name.

sim.version

name of the simulation we want to validate (it works with only one sim.version).

obs.path

the function needs observed data, please create a folder for them in your simulation folder and then indicate in this parameter the access path to this new folder.

releves.PFG

name of file which contain the observed Braund-Blanquet abundance at each site and each PFG and strata (with extension).

hab.obs

name of the file which contain the extended studied map in the simulation (with extension).

validation.mask

name of the file which contain a raster mask that specified which pixels need validation (with extension).

releves.sites

name of the file which contain coordinates and a description of the habitat associated with the dominant species of each site in the studied map (with extension).

Value

One folder is created in name.simulation folder :

VALIDATION/HABITAT/sim.version

containing the prepared CBNA data, RF model, the performance analyzes (confusion matrix and TSS) for the training and testing parts of the RF model, the habitat performance file, the habitat prediction file with observed and simulated habitat for each pixel of the whole map and the final prediction plot.

Details

The observed habitat is derived from the cesbio map, the simulated habitat is derived from FATE simulated relative abundance, based on a random forest algorithm trained on CBNA data. To compare observations and simulations, the function compute confusion matrix between observation and prediction and then compute the TSS for each habitat h (number of prediction of habitat h/number of observation of habitat h + number of non-prediction of habitat h/number of non-observation of habitat h). The final metrics this script use is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed habitat distribution.

Author

Matthieu Combaud, Maxime Delprat