Create DISPERSAL parameter files for a FATE simulation

This script is designed to create parameter files containing dispersal distances for each PFG (one file for each of them) used in the dispersal module of FATE.

PRE_FATE.params_PFGdispersal(
  name.simulation,
  mat.PFG.disp,
  opt.folder.name = NULL
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

mat.PFG.disp
a data.frame with 4 columns : PFG, d50, d99, ldd

opt.folder.name
(optional)
a string corresponding to the name of the folder that will be created into the name.simulation/DATA/PFGS/DISP/ directory to store the results

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/DISP/ directory with the following parameters :

DISPERS_DIST

dispersal distances (3 values) (in meters)

A DISP_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/PFGS/DISP/opt.folder.name/.

Details

A dispersal module is available to make the FATE model spatially explicit by dispersing seeds of each PFG with a kernel (see PRE_FATE.params_globalParameters).

Dispersal distances are needed for each PFG to quantify the amount of seeds dispersed into 3 different concentric circles :

d50

the distance at which 50% of seeds are dispersed

d99

the distance at which 99% of seeds are dispersed

ldd

the long dispersal distance at which 100% of seeds are dispersed

See also

Author

Maya Guéguen

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , d50 = c(50, 50, 10)
                                                        , d99 = c(1000, 1500, 2000)
                                                        , ldd = c(10000, 5000, 10000)))

## ----------------------------------------------------------------------------------------- ##
  
## Load example data