Create LIGHT parameter files for a FATE simulation

This script is designed to create parameter files containing light-related parameters for each PFG (one file for each of them) used in the light module of FATE.

PRE_FATE.params_PFGlight(
  name.simulation,
  mat.PFG.light,
  mat.PFG.tol = NULL,
  opt.folder.name = NULL
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

mat.PFG.light
a data.frame with 2 to 6 columns :

PFG,

type, (or active_germ_low, active_germ_medium, active_germ_high) (or strategy_ag)

type, light_need

(see Details)

mat.PFG.tol
(optional)
a data.frame with 2 to 4 columns :

PFG,

lifeStage, resources, tolerance (or strategy_tol)

(see Details)

opt.folder.name
(optional)
a string corresponding to the name of the folder that will be created into the name.simulation/DATA/PFGS/LIGHT/ directory to store the results

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/LIGHT/ directory with the following parameters :

NAME

name of the PFG

ACTIVE_GERM

germination rates depending on light conditions
(from 0 to 10, corresponding to 0 to 100%)

LIGHT

light value or strategy of the PFG

LIGHT_TOL

light tolerance table (in a single row).
This is a vector of 9 numbers corresponding to the ability of the PFG to survive or not :

at different life stages (Germinant (Ge), Immature (Im), Mature (Ma))

under different light conditions (Low (L), Medium (M) or High (H)).

These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH
(from 0 to 10, corresponding to 0 to 100%).

A LIGHT_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/PFGS/LIGHT/opt.folder.name/.

Details

The light module allows the user to add the effect of light interaction within a primary vegetation succession.

Several parameters, given within mat.PFG.light or mat.PFG.tol, are required for each PFG in order to set up this light interaction :

PFG

the concerned plant functional group

type

or life-form, based on Raunkier.
It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

(active_germ_low)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for Low light condition

(active_germ_medium)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for Medium light condition

(active_germ_high)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for High light condition

(strategy_ag)

a string to choose the germination strategy :
light_lover, indifferent, shade_lover

(light_need)

an integer between 0 and 5 corresponding to the light preference of the PFG (e.g. from Flora Indicativa)

lifeStage

the concerned life stage (Germinant, Immature, Mature)

resources

the concerned light condition (Low, Medium, High)

tolerance

an integer between 0 and 10 corresponding to the proportion of surviving individuals

(strategy_tol)

a string to choose the tolerance strategy :
full_light, pioneer, ubiquist, semi_shade, undergrowth

These values will allow to calculate or define a set of characteristics for each PFG :

ACTIVE_GERM

proportion of seeds that will germinate for each light condition (Low, Medium, High)

Three methods to define these proportions are available :

from predefined scenarios (using strategy_ag) :

| _L_ _M_ _H_ |

_______________

light_lover

| 50% 80% 90% |

indifferent

| 90% 90% 90% |

shade_lover

| 90% 80% 50% |

from predefined rules (using type) :

for H (herbaceous) : 50%, 80%, 90%

for C (chamaephyte) or P (phanerophyte): 90%, 90%, 90%

from user data :
with the values contained within the active_germ_low, active_germ_medium and active_germ_high columns, if provided

LIGHT_TOL

defined for each life stage (Germinant, Immature, Mature)
and each soil condition (Low, Medium, High)

Three methods to define these tolerances are available :

from predefined scenarios (using strategy_tol) :

. means Not tolerant, 1 means Tolerant (100%)

with g: Germinant, i: Immature, m: Mature

with L: low light, M: medium light, H: high light

| _ g _ | _ i _ | _ m _ |

| L M H | L M H | L M H |

_________________________

full_light

| 1 1 1 | . . 1 | . . 1 |

pioneer

| 1 1 1 | . 1 1 | . 1 1 |

ubiquist

| 1 1 1 | 1 1 1 | 1 1 1 |

semi_shade

| 1 1 . | 1 1 . | 1 1 1 |

undergrowth

| 1 1 . | 1 1 . | 1 1 . |

from predefined rules (using type and light_need):

(A)

PFG are tolerant to Low light if light <= 2

(A)

PFG are tolerant to Medium light if 2 <= light <= 4

(A)

PFG are tolerant to High light if light >= 3

(B)

all germinants are assumed to be tolerant to Low light

(C)

all mature trees or chamaephytes are assumed to be tolerant to Medium and High light conditions

(D)

all immature trees that grow in the penultimate stratum are assumed to be tolerant to High light !! desactivated !!

. means Not tolerant

A, B, C, D mean Tolerant according to one of the rule defined above

with g: Germinant, i: Immature, m: Mature

with L: low light, M: medium light, H: high light

| _ g _ | _ i _ | _ m _ |

| L M H | L M H | L M H |

_________________________

1

| A . . | A . D | A C C |

2

| A A . | A A D | A A C |

3

| B A . | . A D | . A C |

4

| B A A | . A A | . A A |

5

| B . A | . . A | . C A |

from user data :
with the values contained within the lifeStage, resources and tolerance columns, if provided

See also

Author

Maya Guéguen

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG light parameter files (with strategies) ----------------------------------------
mat.ag = data.frame(PFG = paste0('PFG', 1:6)
                    , strategy_ag = c('shade_lover', 'indifferent'
                                      , 'indifferent', 'shade_lover'
                                      , 'light_lover', 'light_lover'))

mat.tol = data.frame(PFG = paste0('PFG', 1:6)
                     , strategy_tol = c('undergrowth', 'ubiquist'
                                        , 'ubiquist', 'semi_shade'
                                        , 'pioneer', 'full_light'))

PRE_FATE.params_PFGlight(name.simulation = 'FATE_simulation'
                         , mat.PFG.light = mat.ag
                         , mat.PFG.tol = mat.tol)
                                                        

## Create PFG light parameter files (with all values) ----------------------------------------
mat.ag = data.frame(PFG = paste0('PFG', 1:6)
                    , active_germ_low = c(9, 8, 8, 8, 5, 5)
                    , active_germ_medium = rep(8, 6)
                    , active_germ_high = c(4, 8, 8, 5, 9, 9))

mat.tol = expand.grid(resources = c('Low', 'Medium', 'High')
                      , lifeStage = c('Germinant', 'Immature', 'Mature')
                      , PFG = paste0('PFG', 1:6))
mat.tol$tolerance = c(8, 8, 4, 8, 5, 4, 9, 4, 4
                      , rep(9, 9)
                      , rep(9, 9)
                      , 8, 8, 6, 8, 6, 6, 9, 5, 5
                      , 8, 8, 8, 5, 6, 9, 3, 4, 9
                      , 8, 8, 8, 5, 5, 9, 5, 5, 9)

PRE_FATE.params_PFGlight(name.simulation = 'FATE_simulation'
                         , mat.PFG.light = mat.ag
                         , mat.PFG.tol = mat.tol)
                                                        
                                                        
## -------------------------------------------------------------------------------------------

## Load example data
data(DATASET_Bauges_parameters)

## PFG traits for light
tab.traits = DATASET_Bauges_parameters$tab.light
colnames(tab.traits) = c('PFG', 'type', 'light_need')
str(tab.traits)

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = 'FATE_Bauges')

## Create PFG light parameter files ----------------------------------------------------------
PRE_FATE.params_PFGlight(name.simulation = 'FATE_Bauges'
                         , mat.PFG.light = tab.traits)