Create Namespace_constants parameter file for a FATE-HD simulation
Source: R/PRE_FATE.params_namespaceConstants.R

This script is designed to

PRE_FATE.params_namespaceConstants(name.simulation, global.abund.low,
  global.abund.med, global.abund.high, global.max.by.cohort,
  global.resource.thresh.med, global.resource.thresh.low)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

global.abund.low
Not used for now. To be removed ?

global.abund.med
Not used for now. To be removed ?

global.abund.high
an integer in the order of 1 000 000 to rescale abundance values in each pixel

global.max.by.cohort
an integer in the order of 1 000 000 to rescale abundance values of each cohort in each pixel

global.resource.thresh.med
an integer in the order of 1 000 000 to convert PFG abundances in each strata into light resources. It corresponds to the limit of abundances above which light resources are medium. PFG abundances lower than this threshold imply high amount of light. Is is consequently lower than global.resource.thresh.low.

global.resource.thresh.low
an integer in the order of 1 000 000 to convert PFG abundances in each strata into light resources. It corresponds to the limit of abundances above which light resources are low. PFG abundances higher than global.resource.thresh.med and lower than this threshold imply medium amount of light.

Value

A .txt file into the name.simulation/DATA/NAMESPACE_CONSTANTS directory with the following parameters :

GLOBAL_LOW_ABUND

GLOBAL_MEDIUM_ABUND

GLOBAL_HIGH_ABUND

GLOBAL_MAX_BY_COHORT

GLOBAL_MEDIUM_RESOURCES_TRESH

GLOBAL_LOW_RESOURCES_TRESH

GLOBAL_FULL_SOIL_COVERAGE

Details

To get abundances
per PFG per pixel

per strata

for all strata

$$abund = 10 000 * totalAbund / \text{GLOBAL HIGH ABUND}$$

To transform PFG
abundances into
light resources

$$abund < \text{GLOBAL MEDIUM RESOURCES THRESH} \Leftrightarrow Light = High$$ $$abund > \text{GLOBAL MEDIUM RESOURCES THRESH } \& \\ abund < \text{GLOBAL LOW RESOURCES THRESH} \Leftrightarrow Light = Medium$$ $$abund > \text{GLOBAL LOW RESOURCES THRESH} \Leftrightarrow Light = Low$$

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

#> 
#>  Your directory tree for your FATE-HD simulation ( FATE_simulation ) is ready!

## Create a Namespace_constants parameter file
PRE_FATE.params_namespaceConstants(name.simulation = "FATE_simulation"
                                   , global.abund.low = 1000000
                                   , global.abund.med = 5000000
                                   , global.abund.high = 8000000
                                   , global.max.by.cohort = 5000000
                                   , global.resource.thresh.med = 13000000
                                   , global.resource.thresh.low = 19000000)

#> 
#>  The parameter file  FATE_simulation/DATA/Namespace_constants.txt has been successfully created !
#>