Create SAVE parameter files for a FATE-HD simulation

This script is designed to create a parameter file containing simulation years at which the FATE-HD software must save rasters of PFG abundances or simulation objects.

PRE_FATE.params_saveYears(name.simulation, years.maps = NULL,
  years.objects = NULL, opt.folder.name = NULL)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

years.maps
a vector of simulation years at which PFG abundance maps will be saved

years.objects
a vector of simulation years at which FATE-HD simulation state will be saved

opt.folder.name
(optional)
a string that corresponds to the name of the folder that will be created into the name.simulation/DATA/SAVE/ directory to store the results

Value

Two .txt files into the name.simulation/DATA/SAVE/ directory with one line for each simulation year for which the raster maps / FATE-HD object are to be saved.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/SAVE/opt.folder.name/

Details

FATE-HD software allows the user to save two different types of outputs :

Raster maps

PFG maps can be saved for all specified simulation years. It includes maps per PFG per strata and summary maps per PFG for all height strata combined.
Raster format used is depending on input data format. It can be either .img or .tif.

Model objects

using BOOST library and its serialization functions, FATE-HD is able to save a simulation at a specific time. This object allows the user to restart a simulation from this precise state.

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

# }