Create Simul_parameters parameter file for a FATE simulation

This script is designed to create one (or several) parameter file containing PARAMETER FILENAMES used in FATE model.

PRE_FATE.params_simulParameters(
  name.simulation,
  name.MASK,
  name.SAVED_STATE = NULL,
  name.DIST = NULL,
  name.DROUGHT = NULL,
  name.FIRE = NULL,
  name.ELEVATION = NULL,
  name.SLOPE = NULL,
  opt.global.name = NULL,
  opt.folder.name = NULL
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

name.MASK
a string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the succession (core) module of the FATE simulation will take place (see PRE_FATE.params_globalParameters)

name.SAVED_STATE
(optional)
a string corresponding to the file name of a FATE object, obtained from a previous simulation and from which to restart this new simulation

name.DIST
(optional)
a string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the disturbance module of the FATE simulation will take place (see PRE_FATE.params_globalParameters)

name.DROUGHT
(optional)
a string corresponding to the name of a raster file, with a numeric value within each pixel corresponding to the drought intensity experienced by this pixel through the drought (or fire) disturbance module of the FATE simulation (see PRE_FATE.params_globalParameters)

name.FIRE
(optional)
a string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the fire disturbance module of the FATE simulation will take place (see PRE_FATE.params_globalParameters)

name.ELEVATION
(optional)
a string corresponding to the name of a raster file, with a numeric value within each pixel corresponding to the elevation of this pixel and used by the fire disturbance module of the FATE simulation (see PRE_FATE.params_globalParameters)

name.SLOPE
(optional)
a string corresponding to the name of a raster file, with a numeric value within each pixel corresponding to the slope of this pixel and used by the fire disturbance module of the FATE simulation (see PRE_FATE.params_globalParameters)

opt.global.name
(optional)
a string corresponding to the name of the global parameter file in the folder name.simulation/DATA/GLOBAL_PARAMETERS/ that will be used to build the simulation parameter file

opt.folder.name
(optional)
a string corresponding to the name of the folder in each name.simulation/DATA/PFGS/module/ from which PFG file names will be extracted to build the simulation parameter file

Value

A .txt file into the name.simulation/PARAM_SIMUL/ directory with the following parameters :

--GLOBAL_PARAMS--

--SAVING_DIR--

--SAVED_STATE-- (optional)

--SAVING_YEARS_MAPS-- (optional)

--SAVING_YEARS_OBJECTS-- (optional)

--MASK--

--MASK_CHANGEMASK_YEARS-- (optional)

--MASK_CHANGEMASK_FILES-- (optional)

--PFG_PARAMS_LIFE_HISTORY--

--PFG_PARAMS_LIGHT-- (optional)

--PFG_PARAMS_SOIL-- (optional)

--PFG_PARAMS_DISPERSAL-- (optional)

--PFG_MASK_HABSUIT-- (optional)

--HABSUIT_CHANGEMASK_YEARS-- (optional)

--HABSUIT_CHANGEMASK_FILES-- (optional)

--PFG_PARAMS_DISTURBANCES-- (optional)

--DIST_MASK-- (optional)

--DIST_CHANGEMASK_YEARS-- (optional)

--DIST_CHANGEMASK_FILES-- (optional)

--PFG_PARAMS_DROUGHT-- (optional)

--DROUGHT_MASK-- (optional)

--DROUGHT_CHANGEMASK_YEARS-- (optional)

--DROUGHT_CHANGEMASK_FILES-- (optional)

--PFG_MASK_ALIENS-- (optional)

--ALIENS_CHANGEMASK_YEARS-- (optional)

--ALIENS_CHANGEMASK_FILES-- (optional)

--ALIENS_CHANGEFREQ_YEARS-- (optional)

--ALIENS_CHANGEFREQ_FILES-- (optional)

--PFG_PARAMS_FIRE-- (optional)

--FIRE_MASK-- (optional)

--FIRE_CHANGEMASK_YEARS-- (optional)

--FIRE_CHANGEMASK_FILES-- (optional)

--FIRE_CHANGEFREQ_YEARS-- (optional)

--FIRE_CHANGEFREQ_FILES-- (optional)

--ELEVATION_MASK-- (optional)

--SLOPE_MASK-- (optional)

--END_OF_FILE--

Details

The FATE software takes only one input parameter : a file containing links to other files containing all the parameters and data needed by the program to run.

GLOBAL_PARAMS

file where parameters related to the simulation definition are referred (e.g. number of PFG involved, number of height strata, simulation duration, computer resources, modules loaded, etc)
(see PRE_FATE.params_globalParameters)

SAVING_DIR

directory where simulation outputs will be stored

SAVED_STATE (optional)

file containing the results of a previous FATE simulation from which to restart this new simulation

SAVING_YEARS_
ARRAYS (optional)

file containing the years for which simulation maps will be saved
(see PRE_FATE.params_savingYears)

SAVING_YEARS_
OBJECTS (optional)

file containing the years for which simulation outputs will be saved
(see PRE_FATE.params_savingYears)

MASK

raster mask that will define the study area

MASK_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the succession module
(see PRE_FATE.params_changingYears)

MASK_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the succession module
(see PRE_FATE.params_changingYears)

PFG_PARAMS_
LIFE_HISTORY

PFG life history related parameters (one by PFG)
(see PRE_FATE.params_PFGsuccession)

PFG_PARAMS_
LIGHT (optional)

PFG light preferences and tolerance related parameters (one by PFG)
(see PRE_FATE.params_PFGlight)

PFG_PARAMS_
SOIL (optional)

PFG soil contribution and tolerance related parameters (one by PFG)
(see PRE_FATE.params_PFGsoil)

PFG_PARAMS_
DISPERSAL (optional)

PFG dispersal capacity related parameters (one by PFG)
(see PRE_FATE.params_PFGdispersal)

PFG_MASK_HABSUIT
(optional)

raster masks (one by PFG) containing PFG habitat suitability for the study area

HABSUIT_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the habitat suitability module
(see PRE_FATE.params_changingYears)

HABSUIT_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the habitat suitability module
(see PRE_FATE.params_changingYears)

PFG_PARAMS_
DISTURBANCES (optional)

PFG disturbance related parameters in terms of resprouting and mortality (one by PFG)
(see PRE_FATE.params_PFGdisturbance)

DIST_MASK
(optional)

raster masks that will define the disturbance areas

DIST_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the disturbance module
(see PRE_FATE.params_changingYears)

DIST_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the disturbance module
(see PRE_FATE.params_changingYears)

PFG_PARAMS_DROUGHT
(optional)

PFG drought disturbance related parameters in terms of resprouting and mortality (one by PFG)
(see PRE_FATE.params_PFGdrought)

DROUGHT_MASK
(optional)

raster mask that will define the drought intensity of the area

DROUGHT_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the drought disturbances module
(see PRE_FATE.params_changingYears)

DROUGHT_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the drought disturbances module
(see PRE_FATE.params_changingYears)

PFG_MASK_ALIENS
(optional)

raster masks (one by alien) containing alien introduction zones for the study area

ALIENS_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the aliens introduction module
(see PRE_FATE.params_changingYears)

ALIENS_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the aliens introduction module
(see PRE_FATE.params_changingYears)

ALIENS_CHANGEFREQ_YEARS
(optional)

file containing the years to change frequencies for the aliens introduction module
(see PRE_FATE.params_changingYears)

ALIENS_CHANGEFREQ_FILES
(optional)

file containing the files to change frequencies for the aliens introduction module
(see PRE_FATE.params_changingYears)

PFG_PARAMS_FIRE
(optional)

PFG fire disturbance related parameters in terms of resprouting and mortality (one by PFG)
(see PRE_FATE.params_PFGdisturbance)

FIRE_MASK
(optional)

raster mask that will define the fire disturbance areas

FIRE_CHANGEMASK_YEARS
(optional)

file containing the years to change rasters for the fire disturbances module
(see PRE_FATE.params_changingYears)

FIRE_CHANGEMASK_FILES
(optional)

file containing the files to change rasters for the fire disturbances module
(see PRE_FATE.params_changingYears)

FIRE_CHANGEFREQ_YEARS
(optional)

file containing the years to change frequencies for the fire disturbances module
(see PRE_FATE.params_changingYears)

FIRE_CHANGEFREQ_FILES
(optional)

file containing the files to change frequencies for the fire disturbances module
(see PRE_FATE.params_changingYears)

ELEVATION_MASK
(optional)

raster mask that will define the elevation of the area

SLOPE_MASK
(optional)

raster mask that will define the slope of the area

Note

The function produces links to files that are NOT absolute paths BUT relative ones.
When relative paths are used, the user should be careful of the folder from which the simulation is launched.
A function allows to transform these relative paths into absolute ones. (see examples of .setPattern).

The order of files matters!
For instance the first link below --PFG_PARAMS_LIFE_HISTORY-- flag (e.g. PFG Albert) has to match with the first item below the --PFG_PARAMS_DISPERSAL-- flag (must be PFG Albert too).

See also

Author

Maya Guéguen

Examples

           
## ----------------------------------------------------------------------------------------- ##
                                
## Load example data