Create the skeleton folder for a FATE simulation

This script is designed to create a user-friendly directory tree to run a FATE simulation.

PRE_FATE.skeletonDirectory(name.simulation = "FATE_simulation")

Arguments

name.simulation

a string that will be used as the main directory and simulation name

Value

A directory tree with folders to contain the parameter files, the simulation files and the results.

Details

FATE requires only one input parameter (see PRE_FATE.params_simulParameters),

which is a file containing the names of parameter files,

which may themselves contain

parameters (e.g. succession, dispersal files...)

or other file names (e.g. disturbance or environmental change masks).

The user could give names of files stored everywhere on a machine, and does not have to put them all in one same place.
But as this is more practical, this function proposes a way to organize all those files or parameter files that will or could be used by a FATE simulation.

The tree structure is detailed below :

DATA

this folder will contain all the data or parameters that are needed by the model

GLOBAL_PARAMETERS

files containing global parameters for the simulation
(see PRE_FATE.params_globalParameters)

MASK

raster maps used in the model

SCENARIO

files containing information about changes in raster maps input
(see PRE_FATE.params_changingYears)

SAVE

files containing information about simulation times to save outputs
(see PRE_FATE.params_savingYears)

PFGS

SUCC

life history parameter files
(see PRE_FATE.params_PFGsuccession)

LIGHT

response to light interaction parameter files
(see PRE_FATE.params_PFGlight)

SOIL

response to soil interaction parameter files
(see PRE_FATE.params_PFGsoil)

DISP

dispersal parameter files
(see PRE_FATE.params_PFGdispersal)

HABSUIT

habitat suitability maps

DIST

response to disturbances parameter files
(see PRE_FATE.params_PFGdisturbance)

DROUGHT

response to drought disturbance parameter files
(see PRE_FATE.params_PFGdrought)

ALIENS

aliens introduction maps

PARAM_SIMUL

this folder will contain simulation files that can be given as input to the software
(see PRE_FATE.params_simulParameters)

RESULTS

this folder will collect all the results produced by the software with a folder for each simulation

NB :
All the functions of the RFate package are based on this folder structure.

See also

Author

Maya Guéguen

Examples

## Create a skeleton folder with the default name ('FATE_simulation') ------------------------
PRE_FATE.skeletonDirectory()

## Create a skeleton folder with a specific name ---------------------------------------------
PRE_FATE.skeletonDirectory(name.simulation = 'FATE_AlpineForest')