Calculate PFG traits values based on determinant species traits values

This script is designed to calculate PFG traits values based on determinant species traits values. Either the mean or the median is used depending on the trait class (i.e. numeric or categorical).

PRE_FATE.speciesClustering_step3(mat.traits)

Arguments

mat.traits
a data.frame with at least 3 columns :

species

the ID of each determinant species (see PRE_FATE.speciesClustering_step2)

PFG

the corresponding Plant Functional Group (see PRE_FATE.speciesClustering_step2)

...

one column for each functional trait (see Details)

Value

A list containing one data.frame with the following columns, and one list with as many ggplot2 objects as functional traits given in mat.traits :

tab

PFG

the concerned plant functional group

no.species

the number of species contained in this PFG

...

one column for each functional trait, computed as the mean (for numeric traits) or the median (for categorical traits) of the values of the determinant species of this PFG

plot

...

one for each functional trait, 'specific' cases excepted (see Details)

The information is written in PRE_FATE_PFG_TABLE_traits.csv and PRE_FATE_CLUSTERING_STEP_3_PFGtraitsValues.pdf files.
This .csv file can be used to build parameter files to run a FATE simulation (e.g. PRE_FATE.params_PFGsuccession).

Details

This function allows to obtain 'average' functional trait values for each Plant Functional Group, based on values at the determinant species level.

A graphic is automatically produced for each functional trait given, with boxplot representing the values of determinant species, and colored points the values calculated for each PFG.
However, some traits can have 'specific' representation, as long as their names within mat.traits match one of the configuration detailed below :

maturity, longevity

to visualize the difference between these two values, for the maturity time has an impact on the fecundity of the PFG within FATE (see CORE module)
If there is NO values for longevity within one PFG, and some maturity values are available, some values might be inferred as \(\text{maturity} * 2\). If there is NO values for maturity within one PFG, and some longevity values are available, some values might be inferred as \(\text{longevity} / 2\).

height, light

to visualize the PFG light preference, and help decide and understand the choice of the height limits of strata in FATE (see LIGHT competition module)

soil_contrib,
soil_tol_min,
soil_tol_max

to visualize the PFG soil preference, and help parameterize the global parameters of the soil competition module within FATE (see SOIL competition module)

soil_contrib,
soil_tolerance

same as the previous one, but soil_tol_min and soil_tol_max values are obtained by adding or removing soil_tolerance to soil_contrib

See also

Examples

## Load example data
data(DATASET_Bauges_PFG)

## Species traits
tab.traits = DATASET_Bauges_PFG$dom.traits
str(tab.traits)

## Determinant species
sp.DETERM = DATASET_Bauges_PFG$dom.determ
str(sp.DETERM)

## Merge traits and PFG informations
mat.traits = merge(sp.DETERM[, c("species", "PFG")]
                   , tab.traits
                   , by = "species", all.x = TRUE)

## Compute traits per PFG : no specific graphic ----------------------------------------------
sp.PFG = PRE_FATE.speciesClustering_step3(mat.traits = mat.traits)
names(sp.PFG)
str(sp.PFG$tab)
names(sp.PFG$plot)
plot(sp.PFG$plot$DISPERSAL)
plot(sp.PFG$plot$LIGHT)
plot(sp.PFG$plot$MOISTURE)

## Compute traits per PFG : with one specific graphic ----------------------------------------
colnames(mat.traits) = c("species", "PFG", "GROUP", "DISPERSAL"
                         , "light", "NITROGEN", "MOISTURE", "height")
sp.PFG = PRE_FATE.speciesClustering_step3(mat.traits = mat.traits)
names(sp.PFG)
str(sp.PFG$tab)
names(sp.PFG$plot)
plot(sp.PFG$plot$height_light)