Computation of distances between species based on traits and niche overlap

This script is designed to create a distance matrix between species, combining functional distances (based on functional trait values) and niche overlap (based on co-occurrence of species).

PRE_FATE.speciesDistance(
  mat.traits,
  mat.overlap,
  opt.maxPercent.NA = 0,
  opt.maxPercent.similarSpecies = 0.25,
  opt.min.sd = 0.3
)

Arguments

mat.traits
a data.frame with at least 3 columns :

species

the ID of each studied species

GROUP

a factor variable containing grouping information to divide the species into data subsets (see Details)

...

one column for each functional trait

mat.overlap
two options :

a data.frame with 2 columns :

species

the ID of each studied species

raster

path to raster file with species distribution

a dissimilarity structure representing the niche overlap between each pair of species.
It can be a dist object, a niolap object, or simply a matrix.

opt.maxPercent.NA
(optional) default 0.
Maximum percentage of missing values (NA) allowed for each trait (between 0 and 1)

opt.maxPercent.similarSpecies
(optional) default 0.25.
Maximum percentage of similar species (same value) allowed for each trait (between 0 and 1)

opt.min.sd
(optional) default 0.5.
Minimum standard deviation allowed for each trait (trait unit)

Value

A dist object corresponding to the distance between each pair of species, or a list of dist objects, one for each GROUP value.

The information is written in PRE_FATE_DOMINANT_speciesDistance.csv file (or if necessary, one file is created for each group).

Details

This function allows to obtain a distance matrix between species, based on two types of distance information :

Functional traits :

The GROUP column is required if species must be separated to have one final distance matrix per GROUP value.
If the column is missing, all species will be considered as part of a unique dataset.

The traits can be qualitative or quantitative, but previously identified as such
(i.e. with the use of functions such as as.numeric, as.factor and ordered).

Functional distance matrix is calculated with Gower dissimilarity, using the gowdis function.

This function allows NA values.
However, too many missing values lead to misleading results. Hence, 3 parameters allow the user to play with the place given to missing values, and therefore the selection of traits that will be used for the distance computation :

opt.maxPercent.NA

traits with too many missing values are removed

opt.maxPercent
.similarSpecies

traits with too many similar values are removed

opt.min.sd

traits with too little variability are removed

Niche overlap :

If a data.frame is given, the degree of niche overlap will be computed using the niche.overlap.

Functional distances and niche overlap informations are then combined according to the following formula :

$$\text{mat.DIST}_{sub-group} = \frac{[ \text{mat.OVERLAP}_{sub-group} + \text{mat.FUNCTIONAL}_{sub-group} * n_{traits} ]}{[ n_{traits} + 1 ]}$$

meaning that distance matrix obtained from functional information is weighted by the number of traits used.

See also

Examples

## Load example data
data(DATASET_Bauges_PFG)

## Species traits
tab.traits = DATASET_Bauges_PFG$dom.traits
str(tab.traits)

#> 'data.frame':	284 obs. of  7 variables:
#>  $ species  : Factor w/ 4457 levels "X1000","X10008",..: 2199 2212 1716 2151 2216 2385 2446 2980 3059 3072 ...
#>  $ GROUP    : chr  "Phanerophyte" "Phanerophyte" NA NA ...
#>  $ DISPERSAL: Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 4 4 3 7 2 1 NA 7 5 3 ...
#>  $ LIGHT    : Ord.factor w/ 5 levels "1"<"2"<"3"<"4"<..: 2 2 NA NA 3 3 NA 3 3 4 ...
#>  $ NITROGEN : Ord.factor w/ 9 levels "1"<"1.5"<"2"<..: 5 5 5 5 2 7 NA 6 4 3 ...
#>  $ MOISTURE : Ord.factor w/ 9 levels "1"<"1.5"<"2"<..: 5 6 NA NA 4 6 NA 5 4 3 ...
#>  $ HEIGHT   : num  7.82 7.82 NA NA 3.99 ...

## Species niche overlap (similarity distance)
tab.overlap = DATASET_Bauges_PFG$dom.dist_overlap
tab.overlap[1:5, 1:5]

#>           X10113    X10130    X10328    X10502    X10512
#> X10113        NA 0.3658476 0.5682751 0.7287500 0.6877509
#> X10130 0.6572168        NA 0.1029676 0.4860218 0.1602655
#> X10328 0.8243148 0.2691582        NA 0.5529661 0.6170380
#> X10502 0.9359201 0.7582322 0.7777290        NA 0.4915670
#> X10512 0.8841843 0.3711542 0.8725871 0.7480937        NA

## Give warnings -----------------------------------------------------------------------------
sp.DIST = PRE_FATE.speciesDistance(mat.traits = tab.traits
                                   , mat.overlap = as.matrix(tab.overlap))

#> 
#> 
#>  #------------------------------------------------------------#
#>  # PRE_FATE.speciesDistance
#>  #------------------------------------------------------------# 
#> 
#>  ---------- INFORMATION : AVAILABLE 
#> 
#> > FOR TRAITS 
#>   Number of species :  284
#>   Groups :  Chamaephyte, Herbaceous, Phanerophyte
#>   Measured traits :  DISPERSAL, LIGHT, NITROGEN, MOISTURE, HEIGHT

#> Warning: Missing data!
#>  `mat.traits` contains some species with no trait values : X323, X40623, X41178
#> These species will not be taken into account ! 
#> 

#> 
#> > FOR OVERLAP 
#>   Number of species :  275
#> 
#> > FOR BOTH 
#>   Number of species with traits and no overlap information :  17
#>   Number of species with overlap and no traits information :  11
#>   Number of species with both trait and overlap distances:  264
#> 
#>   Comparison of groups' dimensions : 
#> 
#> > Group Chamaephyte :    trait values = 31   overlap values =  28
#> > Group Herbaceous :    trait values = 216   overlap values =  204
#> > Group Phanerophyte :    trait values = 32   overlap values =  30

#> Warning: Missing data!
#>  `mat.traits` contains some traits with too many missing values or not enough variation between species. 
#> These traits will not be taken into account ! 
#> 
#> Columns below represent for each trait :
#>  - the percentage of missing values 
#>  - the percentage of similar species 
#>  - the standard deviation of pairwise distances 
#> 
#> In group Chamaephyte : 
#>  >> DISPERSAL			0.38	45.87	17.83
#>  >> LIGHT			0	33.6	23.93
#>  >> NITROGEN			0	29.63	21.15
#> In group Herbaceous : 
#>  >> DISPERSAL			2.27	17.15	28.67
#>  >> LIGHT			0.76	37.25	27.07
#>  >> MOISTURE			1.14	15.74	23.08
#>  >> NITROGEN			0	25.87	19.76
#> In group Phanerophyte : 
#>  >> DISPERSAL			0	33.56	22.79
#>  >> LIGHT			0	32.64	23.99
#>  >> MOISTURE			0.38	18.72	23.58
#>  >> NITROGEN			0	28.05	28.49

#> 
#>  ---------- INFORMATION : USED 
#> 
#>   Traits used to calculate functional distances : 
#> 
#> > Group Chamaephyte : HEIGHT MOISTURE
#> > Group Herbaceous : HEIGHT
#> > Group Phanerophyte : HEIGHT
#> 
#>   Number of species :  262
#>   Groups :  Chamaephyte, Herbaceous, Phanerophyte
#>   Number of species in each group :  28 204 30
#>   Number of NA values due to `gowdis` function :  2
#> 
#> > Done!

#> 
#>  The output files 
#>  > PRE_FATE_DOMINANT_speciesDistance_Chamaephyte.csv 
#>  > PRE_FATE_DOMINANT_speciesDistance_Herbaceous.csv 
#>  > PRE_FATE_DOMINANT_speciesDistance_Phanerophyte.csv 
#> have been successfully created !

str(sp.DIST)

#> List of 3
#>  $ Chamaephyte : 'dist' num [1:378] 0.2286 0.23 0.0955 0.1376 0.0809 ...
#>   ..- attr(*, "Labels")= chr [1:28] "X10559" "X10896" "X11038" "X11102" ...
#>   ..- attr(*, "Size")= int 28
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE
#>  $ Herbaceous  : 'dist' num [1:20706] 0.181 0.113 0.139 0.36 0.168 ...
#>   ..- attr(*, "Labels")= chr [1:204] "X10113" "X10130" "X10328" "X10502" ...
#>   ..- attr(*, "Size")= int 204
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE
#>  $ Phanerophyte: 'dist' num [1:435] 0.0224 0.3209 0.1192 0.1826 0.3868 ...
#>   ..- attr(*, "Labels")= chr [1:30] "X10512" "X10889" "X11057" "X11313" ...
#>   ..- attr(*, "Size")= int 30
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE

if (FALSE) {
require(foreach)
require(ggplot2)
require(ggdendro)
pp = foreach(x = names(sp.DIST)) %do%
{
  hc = hclust(sp.DIST[[x]])
  pp = ggdendrogram(hc, rotate = TRUE) +
    labs(title = paste0("Hierarchical clustering based on species distance "
                        , ifelse(length(names(sp.DIST)) > 1
                        , paste0("(group ", x, ")")
                        , "")))
  return(pp)
}
plot(pp[[1]])
plot(pp[[2]])
plot(pp[[3]])
}

## Change parameters to allow more NAs (and change traits used) ------------------------------
sp.DIST = PRE_FATE.speciesDistance(mat.traits = tab.traits
                                   , mat.overlap = as.matrix(tab.overlap)
                                   , opt.maxPercent.NA = 0.05
                                   , opt.maxPercent.similarSpecies = 0.3
                                   , opt.min.sd = 0.3)

#> 
#> 
#>  #------------------------------------------------------------#
#>  # PRE_FATE.speciesDistance
#>  #------------------------------------------------------------# 
#> 
#>  ---------- INFORMATION : AVAILABLE 
#> 
#> > FOR TRAITS 
#>   Number of species :  284
#>   Groups :  Chamaephyte, Herbaceous, Phanerophyte
#>   Measured traits :  DISPERSAL, LIGHT, NITROGEN, MOISTURE, HEIGHT

#> Warning: Missing data!
#>  `mat.traits` contains some species with no trait values : X323, X40623, X41178
#> These species will not be taken into account ! 
#> 

#> 
#> > FOR OVERLAP 
#>   Number of species :  275
#> 
#> > FOR BOTH 
#>   Number of species with traits and no overlap information :  17
#>   Number of species with overlap and no traits information :  11
#>   Number of species with both trait and overlap distances:  264
#> 
#>   Comparison of groups' dimensions : 
#> 
#> > Group Chamaephyte :    trait values = 31   overlap values =  28
#> > Group Herbaceous :    trait values = 216   overlap values =  204
#> > Group Phanerophyte :    trait values = 32   overlap values =  30

#> Warning: Missing data!
#>  `mat.traits` contains some traits with too many missing values or not enough variation between species. 
#> These traits will not be taken into account ! 
#> 
#> Columns below represent for each trait :
#>  - the percentage of missing values 
#>  - the percentage of similar species 
#>  - the standard deviation of pairwise distances 
#> 
#> In group Chamaephyte : 
#>  >> DISPERSAL			0.38	45.87	17.83
#>  >> LIGHT			0	33.6	23.93
#> In group Herbaceous : 
#>  >> LIGHT			0.76	37.25	27.07
#> In group Phanerophyte : 
#>  >> DISPERSAL			0	33.56	22.79
#>  >> LIGHT			0	32.64	23.99

#> 
#>  ---------- INFORMATION : USED 
#> 
#>   Traits used to calculate functional distances : 
#> 
#> > Group Chamaephyte : HEIGHT MOISTURE NITROGEN
#> > Group Herbaceous : DISPERSAL HEIGHT MOISTURE NITROGEN
#> > Group Phanerophyte : HEIGHT MOISTURE NITROGEN
#> 
#>   Number of species :  262
#>   Groups :  Chamaephyte, Herbaceous, Phanerophyte
#>   Number of species in each group :  28 204 30
#>   Number of NA values due to `gowdis` function :  2
#> 
#> > Done!

#> 
#>  The output files 
#>  > PRE_FATE_DOMINANT_speciesDistance_Chamaephyte.csv 
#>  > PRE_FATE_DOMINANT_speciesDistance_Herbaceous.csv 
#>  > PRE_FATE_DOMINANT_speciesDistance_Phanerophyte.csv 
#> have been successfully created !

str(sp.DIST)

#> List of 3
#>  $ Chamaephyte : 'dist' num [1:378] 0.1914 0.1925 0.0716 0.1032 0.0607 ...
#>   ..- attr(*, "Labels")= chr [1:28] "X10559" "X10896" "X11038" "X11102" ...
#>   ..- attr(*, "Size")= int 28
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE
#>  $ Herbaceous  : 'dist' num [1:20706] 0.3 0.184 0.059 0.526 0.286 ...
#>   ..- attr(*, "Labels")= chr [1:204] "X10113" "X10130" "X10328" "X10502" ...
#>   ..- attr(*, "Size")= int 204
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE
#>  $ Phanerophyte: 'dist' num [1:435] 0.0201 0.1604 0.1905 0.1836 0.2855 ...
#>   ..- attr(*, "Labels")= chr [1:30] "X10512" "X10889" "X11057" "X11313" ...
#>   ..- attr(*, "Size")= int 30
#>   ..- attr(*, "call")= language as.dist.default(m = mat)
#>   ..- attr(*, "Diag")= logi FALSE
#>   ..- attr(*, "Upper")= logi FALSE