Save data to reproduce building of parameter files

This script is designed to gather all data and parameters used to build a FATE simulation folder.

SAVE_FATE.step2_parameters(
  name.dataset,
  name.simulation = NA,
  strata.limits,
  mat.PFG.succ,
  mat.PFG.light = NA,
  mat.PFG.light.tol = NA,
  mat.PFG.soil = NA,
  mat.PFG.soil.tol = NA,
  mat.PFG.disp = NA,
  mat.PFG.dist = NA,
  mat.PFG.dist.tol = NA,
  mat.PFG.drought = NA,
  mat.PFG.drought.tol = NA,
  rasters = list(name.MASK = NA, name.DIST = NA, name.DROUGHT = NA, name.FIRE = NA,
    name.ELEVATION = NA, name.SLOPE = NA),
  multipleSet = list(name.simulation.1 = NA, name.simulation.2 = NA, file.simulParam.1
    = NA, file.simulParam.2 = NA, no_simulations = NA, opt.percent_maxAbund = NA,
    opt.percent_seeding = NA, opt.percent_light = NA, opt.percent_soil = NA,
    do.max_abund_low = NA, do.max_abund_medium = NA, do.max_abund_high = NA,
    do.seeding_duration = NA, do.seeding_timestep = NA, do.seeding_input = NA,
    do.no_strata = NA, do.LIGHT.thresh_medium = NA, do.LIGHT.thresh_low = NA,
    do.SOIL.init = NA, do.SOIL.retention = NA, do.DISPERSAL.mode = NA,     
    do.HABSUIT.mode = NA)
)

Arguments

name.dataset
a string corresponding to the name to give to archive folder

name.simulation
(optional) default NA.
A string corresponding to the name of the simulation folder

strata.limits
a vector of integer containing height strata limits

mat.PFG.succ
a data.frame with at least 5 columns :
PFG, type, height, maturity, longevity
(and optionally, max_abundance, potential_fecundity, immature_size, is_alien, flammability)
(see PRE_FATE.params_PFGsuccession)

mat.PFG.light
(optional) default NA.
A data.frame with 2 to 6 columns :

PFG,

type, (or active_germ_low, active_germ_medium, active_germ_high) (or strategy_ag)

type, light_need

(see PRE_FATE.params_PFGlight)

mat.PFG.light.tol
(optional) default NA.
A data.frame with 2 to 4 columns :

PFG,

lifeStage, resources, tolerance (or strategy_tol)

(see PRE_FATE.params_PFGlight)

mat.PFG.soil
(optional) default NA.
A data.frame with 3 to 7 columns :

PFG,

type, (or active_germ_low, active_germ_medium, active_germ_high) (or strategy_ag)

soil_contrib, soil_tol_min, soil_tol_max (or strategy_contrib)

(see PRE_FATE.params_PFGsoil)

mat.PFG.soil.tol
(optional) default NA.
A data.frame with 2 to 4 columns :

PFG,

lifeStage, resources, tolerance (or strategy_tol)

(see PRE_FATE.params_PFGsoil)

mat.PFG.disp
(optional) default NA.
A data.frame with 4 columns : PFG, d50, d99, ldd (see PRE_FATE.params_PFGdispersal)

mat.PFG.dist
(optional) default NA.
A data.frame with 5 columns :
PFG, type, maturity, longevity, age_above_150cm (see PRE_FATE.params_PFGdisturbance)

mat.PFG.dist.tol
(optional) default NA.
A data.frame with 3 to 7 columns :

nameDist,

PFG,

(responseStage, breakAge, resproutAge),

responseStage, killedIndiv, resproutIndiv (or strategy_tol)

(see PRE_FATE.params_PFGdisturbance)

mat.PFG.drought
(optional) default NA.
A data.frame with 4 or 6 columns :

PFG,

threshold_moderate, threshold_severe,

counter_recovery, counter_sens, counter_cum (or strategy_drou)

(see PRE_FATE.params_PFGdrought)

mat.PFG.drought.tol
(optional) default NA.
A data.frame with 3 to 7 columns :

nameDist,

PFG,

(responseStage, breakAge, resproutAge),

responseStage, killedIndiv, resproutIndiv (or strategy_tol)

(see PRE_FATE.params_PFGdrought)

rasters
a list containing all the rasters given to the PRE_FATE.params_simulParameters function, if used (name.MASK, name.DIST, name.DROUGHT, name.FIRE, name.ELEVATION, name.SLOPE)

multipleSet
a list containing all the parameter values given to the PRE_FATE.params_multipleSet function, if used (name.simulation.1, name.simulation.2, file.simulParam.1, file.simulParam.2, no_simulations, opt.percent_maxAbund, opt.percent_seeding, opt.percent_light, opt.percent_soil, do.max_abund_low, do.max_abund_medium, do.max_abund_high, do.seeding_duration, do.seeding_timestep, do.seeding_input, do.no_strata, do.LIGHT.thresh_medium, do.LIGHT.thresh_low, do.SOIL.init, do.SOIL.retention, do.DISPERSAL.mode, do.HABSUIT.mode)

Value

A list containing all the elements given to the function and checked, and two archive files :

name.dataset

name of the dataset

strata.limits

height strata limits

mat.PFG.succ

(mat.PFG.light)

(mat.PFG.light.tol)

(mat.PFG.soil)

(mat.PFG.soil.tol)

(mat.PFG.disp)

(mat.PFG.dist)

(mat.PFG.dist.tol)

(mat.PFG.drought)

(mat.PFG.drought.tol)

rasters

raster files of all simulation masks

(multipleSet)

(name.simulation)

name of the simulation folder

(DATA folder)

contained in name.simulation folder and archived

(PARAM_SIMUL folder)

contained in name.simulation folder and archived

The information is written in FATE_dataset_[name.dataset]_step2_parameters.RData file.

See also

Author

Maya Guéguen

Examples

## Load example data