Compare observed and simulated habitat of a FATE simulation at the last simulation year.

To compare observations and simulations, this function compute confusion matrix between observation and prediction and then compute the TSS for each habitat.

do.habitat.validation(
  output.path,
  RF.model,
  habitat.FATE.map,
  validation.mask,
  simulation.map,
  predict.all.map,
  sim.version,
  name.simulation,
  perStrata,
  hab.obs,
  year,
  list.strata.releves,
  list.strata.simulations
)

Arguments

output.path

access path to the for the folder where output files will be created.

RF.model

random forest model trained on CBNA data (train.RF.habitat function)

habitat.FATE.map

a raster map of the observed habitat in the studied area.

validation.mask

a raster mask that specified which pixels need validation.

simulation.map

a raster map of the whole studied area use to check the consistency between simulation map and the observed habitat map.

predict.all.map

Logical. If TRUE, the script will predict habitat for the whole map.

sim.version

name of the simulation to validate.

name.simulation

simulation folder name.

perStrata

Logical. If TRUE, the PFG abundance is defined by strata in each pixel. If FALSE, PFG abundance is defined for all strata.

hab.obs

a raster map of the observed habitat in the extended studied area.

year

simulation year selected for validation.

list.strata.releves

a character vector which contain the observed strata definition, extracted from observed PFG releves.

list.strata.simulations

a character vector which contain FATE strata definition and correspondence with observed strata definition.

Value

Habitat performance file.
If option selected, the function also returns an habitat prediction file with observed and simulated habitat for each pixel of the whole map.

Details

After several preliminary checks, the function is going to prepare the observations database by extracting the observed habitat from a raster map. Then, for the simulation sim.version, the script take the evolution abundance for each PFG and all strata (or for each PFG & each strata if option selected) file and predict the habitat for the whole map (if option selected) thanks to the RF model. Finally, the function computes habitat performance based on TSS for each habitat.

Author

Matthieu Combaud & Maxime Delprat