Compare observed and simulated habitat of a FATE simulation to a chosen year.

To compare observations and simulations, this function compute confusion matrix between observation and prediction and then compute the TSS for each habitats.

do_habitat_validation(
  output.path,
  RF.model,
  predict.all.map,
  sim,
  simu_PFG,
  habitat.whole.area.df,
  list.strata,
  perStrata
)

Arguments

output.path

Access path to the for the folder where output files will be created.

RF.model

Random forest model trained on observed abundance data (train_RF_habitat function)

predict.all.map

Logical. If TRUE, the script will predict habitat for the whole map.

sim

Name of the single simulation to validate.

simu_PFG

A data.frame provides by POST_FATE.temporalEvolution with simulated abundance for each PFG and strata (if option selected) and pixel ID (see POST_FATE.temporalEvolution).

habitat.whole.area.df

A data.frame provides by POST_FATE.validation with 3 columns :
pixel which contains the ID of each pixel in the study area.
code.habitat which contains the ID of the habitat in each pixel.
for.validation for each pixel, 0 if does not need validation, 1 if needs validation.

list.strata

If abundance file is defined by strata : a character vector which contains FATE strata definition and correspondence with observed strata definition.
If abundance file is defined for all strata : a character vector with value "all".

perStrata

Logical. Default TRUE. If TRUE, PFG abundance is defined by strata. If FALSE, PFG abundance defined for all strata.

Value

Habitat performance file.
If option selected, the function also returns an habitat prediction file with observed and simulated habitat for each pixel of the whole map.

Details

For a given simulation sim, this script takes the evolution abundance for each PFG and all strata (or for each PFG & each strata if option selected) data frame and predicts the habitat for the whole map (if option selected) thanks to the RF model. Finally, the function computes habitat performance based on TSS for each habitat.

Author

Matthieu Combaud & Maxime Delprat