Adapt all raster maps of a FATE simulation folder (change NA to 0, and save as .tif)

This function scan all the raster files within a FATE simulation folder, change all NA values to 0, potentially reproject onto a raster mask and save them with the specified extension.

.adaptMaps(
  name.simulation,
  opt.name.file = NULL,
  extension.old,
  extension.new = NULL,
  opt.name.MASK = NULL
)

Arguments

name.simulation

a string corresponding to the main directory or simulation name of the FATE simulation

opt.name.file

(optional)
a string corresponding to the complete or partial name of the file in which to search and change the pattern

extension.old

a string corresponding to the extension of raster files to be found

extension.new

(optional)
a string (either tif or img) corresponding to the new extension to save all the maps

opt.name.MASK

(optional) default NULL.
A string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the succession (core) module of the FATE simulation will take place (see PRE_FATE.params_globalParameters)

Author

Maya Guéguen