Load a dataset

This function loads one of the available data sets, containing :

PNE_PFG

elements to create the Plant Functional Groups (PFG) over the Ecrins National Park (PNE)
(Boulangeat et al. 2012 GCB)

PNE_PARAM

all necessary files to build the FATE simulation folder as well as the parameter files
(Boulangeat et al. 2014 GCB)

PNE_RESULTS

results obtained from outputs of FATE simulation
(Boulangeat et al. 2014 GCB)

.loadData(data.name)

Arguments

data.name
a string corresponding to the name of the dataset that will be loaded

Value

PNE_PFG

A list object with 6 elements to help building the Plant Functional Group :

sp.observations

a data.frame of dimension 168313 x 7
containing releves data about plant species in the PNE
to be used with the PRE_FATE.selectDominant function

sites : sites ID

X : x-axis coordinates in Lambers (lcc)

Y : y-axis coordinates in Lambers (lcc)

habitat : habitat ID from DELPHINE :

0 : glaciers and eternal snows

31 : uncolonized rocks

40 : lawns and meadows

50 : low moors

60 : open areas, brush

70 : semi-closed areas

81 : closed areas

83 : forests

species : species ID

abund_BB : Braun-Blanquet abundance
(see PRE_FATE.abundBraunBlanquet function for details)

abund : relative abundance obtained from the abund_BB column with the PRE_FATE.abundBraunBlanquet function

dom.traits

data.frame of dimension 359 x 6
containing traits for dominant species
to be used with the PRE_FATE.speciesDistance function

species : species ID

GROUP : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

height : mean plant height (cm)

dispersal : classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

palatability : classes (from 0 to 3) (CBNA)

light : Ellenberg indicator value for light (from 1 to 9)

dom.dist_overlap

matrix of dimension 358 x 358
containing niche overlap distance for dominant species
to be used with the PRE_FATE.speciesDistance function

dom.determ

data.frame of dimension 359 x 5
containing dominant species information relative to PFG
obtained from the PRE_FATE.speciesClustering_step2 function

species : species ID

name : species name (taxonomic)

GROUP : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

PFG : name of assigned Plant Functional Group

determinant : is the species kept as determinant species within the PFG ? (see PRE_FATE.speciesClustering_step2 function for details)

nb.clusters

vector of length 3
number of groups kept for each life-form class, obtained by cutting the hierarchical tree obtained from species distances

PFG.traits

data.frame of dimension 24 x 10
containing traits for plant functional groups
obtained from the PRE_FATE.speciesClustering_step3 function

PFG_name : full descriptive Plant Functional Group name

PFG : Plant Functional Group short name

type : rough generalization of Raunkier life-forms :

H : herbaceous

C : chamaephytes

P : phanerophytes

strata : maximum height stratum that can be reached by the PFG (from 1 to 5)

disp : MEDIAN classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

light : MEDIAN Ellenberg indicator value for light (from 1 to 9)

height : MEAN PFG height (cm)

palatability : MEDIAN classes (from 0 to 3) (CBNA)

longevity : MEAN age of lifespan

maturity : MEAN age of maturity

PNE_PARAM

A list object with 13 elements to help building the simulation files and folders to run a FATE simulation :

masks

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 7 mask layers with binary values (0 or 1) to be used in a FATE simulation :

maskEcrins : simulation map, where occurs succession

noDisturb : perturbation map, when there is none

mowing : perturbation map, where occurs mowing

grazing1 : perturbation map, where occurs light grazing

grazing2 : perturbation map, where occurs extensive grazing

grazing3 : perturbation map, where occurs intensive grazing

grazingAll : perturbation map, where occurs grazing, all types combined

HS_0

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 24 layers with probability values (between 0 and 1) representing Habitat Suitability for initialization phase and equilibrium or current time (0) for each PFG and to be used in a FATE simulation.
These maps are coming from Species Distribution Models and methods to obtain them are described in Supplementary Materials of Boulangeat et al. 2014 GCB.

HS_15

same as HS_0 but 15 years after equilibrium

HS_30

same as HS_0 but 30 years after equilibrium

HS_45

same as HS_0 but 45 years after equilibrium

HS_60

same as HS_0 but 60 years after equilibrium

HS_75

same as HS_0 but 75 years after equilibrium

HS_90

same as HS_0 but 90 years after equilibrium

strata_limits

a vector of length 5
containing height of stratum limits in centimeters
to be used with the PRE_FATE.params_PFGsuccession and PRE_FATE.params_PFGlight functions

succ_light

a data.frame of dimension 24 x 6
containing traits for plant functional groups
obtained from the PRE_FATE.speciesClustering_step3 function
to be used with the PRE_FATE.params_PFGsuccession and PRE_FATE.params_PFGlight functions

disp

a data.frame of dimension 24 x 4
containing dispersal values (in meters) for plant functional groups
to be used with the PRE_FATE.params_PFGdispersal function

dist

a data.frame of dimension 384 x 5
containing response of plant functional groups to disturbances
to be used with the PRE_FATE.params_PFGdisturbance function

global

a vector of length 18
containing global parameter values for a FATE simulation in the PNE
to be used with the PRE_FATE.params_globalParameters function

PNE_RESULTS

A list object with 5 elements :

evaluation

a data.frame of dimension 24 x 11
containing statistical metrics to evaluate the predictive quality of both Habitat Suitability models (HS) and FATE simulation outputs

PFG : name of Plant Functional Group

nb.obs.absences : number of observed absences

nb.obs.presences : number of observed presences

specificity.FATE : proportion of actual negatives that are correctly identified as such by the FATE model (true negative rate)

specificity.HS : proportion of actual negatives that are correctly identified as such by the Habitat Suitability model (true negative rate)

sensitivity.FATE : proportion of actual positives that are correctly identified as such by the FATE model (true positive rate)

sensitivity.HS : proportion of actual positives that are correctly identified as such by the Habitat Suitability model (true positive rate)

TSS.FATE : True Skill Statistic for FATE model

TSS.HS : True Skill Statistic for Habitat Suitability model

error.rate.FATE : percentage of bad predictions for FATE model

error.rate.HS : percentage of bad predictions for Habitat Suitability model

abund_str.equilibrium

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 120 layers representing FATE abundances for year 800 of initialization phase (= equilibrium or current time).
Maps are per PFG and per height stratum.

forest_cover.init

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 16 layers representing FATE relative forest cover through initialization phase (from year 50 to 800).
Maps are summing all PFGs through all height strata above 1.5 meters.

forest_cover
CC_BAU

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 14 layers representing FATE relative forest cover through Climate Change (CC) + Business-As-Usual (BAU) scenario (from year 850 to 1500).
Maps are summing all PFGs through all height strata above 1.5 meters.

forest_cover
CC_Abandon

a stack object of dimension 782 x 619 with a resolution of 100m and Lambers (lcc) projection, containing 14 layers representing FATE relative forest cover through Climate Change (CC) + Abandonment (BAU) scenario (from year 850 to 1500).
Maps are summing all PFGs through all height strata above 1.5 meters.

Author

Maya Guéguen