Compute distribution of relative abundance over observed relevés

This script is designed to compute distribution, per PFG/strata/habitat, of relative abundance, from observed data.

get_observed_distribution(
  releves.PFG,
  hab.obs,
  studied.habitat = NULL,
  PFG.considered_PFG.compo,
  strata.considered_PFG.compo,
  habitat.considered_PFG.compo,
  perStrata,
  output.path
)

Arguments

releves.PFG

a data frame with abundance (column named abund) at each site and for each PFG and strata.

hab.obs

a raster map of the extended studied map in the simulation, with same projection & resolution than simulation mask.

studied.habitat

default NULL. If NULL, the function will take into account of habitats define in the hab.obs map. Otherwise, please specify in a 2 columns data frame the habitats (2nd column) and the ID (1st column) for each of them which will be taken into account for the validation.

PFG.considered_PFG.compo

a character vector of the list of PFG considered in the validation.

strata.considered_PFG.compo

a character vector of the list of precise strata considered in the validation.

habitat.considered_PFG.compo

a character vector of the list of habitat(s) considered in the validation.

perStrata

Logical. All strata together (FALSE) or per strata (TRUE).

name.simulation

simulation folder name.

Value

Details

The function takes the releves.PFG and releves.sites files and aggregate coverage per PFG. Then, the code get habitat information from also the hab.obs map, keep only interesting habitat, strata and PFG, and transform the data into relative metrics. Finally, the script computes distribution per PFG, and if require per strata/habitat (else all strata/habitat will be considered together).

Author

Matthieu Combaud, Maxime Delprat