Create a random forest algorithm trained on observed vegetation data

This script is designed to produce a random forest model trained on observed PFG abundance and a map of observed habitat.

train_RF_habitat(
  releves.PFG,
  hab.obs.RF = NULL,
  external.training.mask = NULL,
  studied.habitat,
  RF.param,
  output.path,
  perStrata = FALSE,
  seed
)

Arguments

releves.PFG

a data.frame with at least 5 columns :
site, x, y, abund, PFG
(and optionally, strata, code.habitat)
(see Details)

external.training.mask

(optional) default NULL.
a raster map for keeping releves data only in a specific area.

studied.habitat

a data.frame with 2 columns :
ID ,habitat
(see Details)

RF.param

a list of 2 parameters to fit a random forest model :
share.training defines the size of the training part of the data base.
ntree is the number of trees build by the algorithm, it allows to reduce the prediction error.

output.path

access path to the folder where output files will be created.

perStrata

(Logical) default FALSE.
If TRUE, the PFG abundance must be defined by strata in each site. If FALSE, PFG abundance must be defined for all strata.

Value

2 prepared observed releves files are created before the building of the random forest model in a folder previously defined.
5 more files are created at the end of the script to save the RF model and the performance analyzes (confusion matrix and TSS) for the training and testing parts.

Details

This function transforms PFG abundance in relative abundance, gets habitat information from an habitat data frame previously defined and a habitat map, keep releves on interesting habitat(s) and then builds a random forest model. Finally, the function analyzes the model performance with computation of confusion matrix and TSS between the training and testing sample.

Author

Matthieu Combaud, Maxime Delprat