Create a graphical representation of the evolution of PFG coverage and abundance through time for a FATE simulation

This script is designed to produce two graphical representations for a FATE simulation : 1) the evolution through time of the space occupation of each PFG ; 2) the evolution through time of the abundance of each PFG. These graphics represent both the evolution over the whole area.

POST_FATE.graphic_evolutionCoverage(
  name.simulation,
  file.simulParam = NULL,
  opt.fixedScale = TRUE,
  opt.doPlot = TRUE
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

file.simulParam
default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

opt.fixedScale
(optional) default TRUE.
If FALSE, the ordinate scale will be adapted for each PFG for the graphical representation of the evolution of abundances through time

opt.doPlot
(optional) default TRUE.
If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned

Value

A list containing two data.frame objects with the following columns, and two ggplot2 objects :

tab.spaceOccupancy

PFG

concerned plant functional group (for abundance)

HAB

concerned habitat

year

concerned simulation year

spaceOccupancy

number of occupied pixels divided by the total number of pixels within the studied area

tab.totalAbundance

PFG

concerned plant functional group (for abundance)

HAB

concerned habitat

year

concerned simulation year

totalAbundance

total abundance over all the pixels within the studied area

plot.spaceOccupancy

ggplot2 object, representing the evolution of each PFG space occupancy

plot.totalAbundance

ggplot2 object, representing the evolution of each PFG total abundance

Two POST_FATE_TABLE_ZONE_evolution_[...].csv files are created :

spaceOccupancy

always, containing tab.spaceOccupancy

totalAbundance

always, containing tab.totalAbundance

One POST_FATE_GRAPHIC_A_evolution_coverage_[...].pdf file is created containing two types of graphics :

spaceOccupancy

to visualize for each PFG the evolution of its occupation of the studied area through simulation time

totalAbundance

to visualize for each PFG the evolution of its abundance within the whole studied area through simulation time

Details

This function allows to obtain, for a specific FATE simulation and a specific parameter file within this simulation, two preanalytical graphics :

the evolution of space occupancy of each plant functional group through simulation time,
with space occupancy representing the percentage of pixels within the mask of studied area where the PFG is present

the evolution of total abundance of each plant functional group through simulation time,
with total abundance being the sum over the whole studied area of the PFG abundances (FATE arbitrary unit)

If the information has been provided (see POST_FATE.temporalEvolution), the graphics will be also done per habitat.

It requires that the POST_FATE.temporalEvolution function has been run and that the file POST_FATE_TABLE_PIXEL_evolution_abundance.csv exists.

See also

Author

Maya Guéguen

Examples

if (FALSE) {                      
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")
                                    
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.fixedScale = FALSE)
}

## ----------------------------------------------------------------------------------------- ##
## Load example data