Create a graphical representation of the evolution of PFG coverage and abundance through time for a FATE-HD simulation

This script is designed to produce two graphical representations for a FATE-HD simulation : 1) the evolution through time of the space occupation of each PFG ; 2) the evolution through time of the abundance of each PFG. These graphics represent both the evolution over the whole area.

POST_FATE.graphic_evolutionCoverage(name.simulation,
  file.simulParam = NULL, opt.abund_fixedScale = TRUE,
  opt.doPlot = TRUE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

opt.abund_fixedScale
default TRUE. If FALSE, the ordinate scale will be adapted for each PFG for the graphical representation of the evolution of abundances through time

opt.doPlot
default TRUE (optional). If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned.

Value

A list containing two data.frame objects with the following columns, and two ggplot2 objects :

tab.spaceOccupancy

PFG

the concerned Plant Functional Group (for abundance)

HAB

the concerned habitat

YEAR

the concerned simulation year

Occupancy

the number of occupied pixels divided by the total number of pixels within the studied area

tab.abundance

PFG

the concerned Plant Functional Group (for abundance)

HAB

the concerned habitat

YEAR

the concerned simulation year

Abund

the total abundance over all the pixels within the studied area

graph.spaceOccupancy

ggplot2 object, representing the evolution of each PFG space occupancy

graph.abundance

ggplot2 object, representing the evolution of each PFG total abundance

Two POST_FATE_TABLE_ZONE_evolution_[...].csv files are created :

spaceOccupancy

always, containing tab.spaceOccupancy

abundance

always, containing tab.abundance

Two POST_FATE_[...].pdf files are created :

GRAPHIC_A
spaceOccupancy

to visualize for each PFG the evolution of its occupation of the studied area through simulation time

GRAPHIC_A
abundance

to visualize for each PFG the evolution of its abundance within the whole studied area through simulation time

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, two preanalytical graphics :

the evolution of space occupancy of each Plant Functional Group through simulation time, with space occupancy representing the percentage of pixels within the mask of studied area where the PFG is present

the evolution of abundance of each Plant Functional Group through simulation time, with abundance being the sum over the whole studied area of the PFG abundances (FATE-HD arbitrary unit)

If the information has been provided, the graphics will be also done per habitat.

It requires that the POST_FATE.temporalEvolution function has been run and that the file POST_FATE.evolution_abundance_PIXEL_[...].csv exists.

See also

Examples

if (FALSE) {
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)

POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 4)

POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
}