Create a graphical representation of the evolution of PFG coverage and abundance through time for a FATE-HD simulation

This script is designed to produce two graphical representations for a FATE-HD simulation : 1) the evolution through time of the space occupation of each PFG ; 2) the evolution through time of the abundance of each PFG.

POST_FATE.graphic_evolutionCoverage(name.simulation,
  file.simulParam = NULL, no.years = 10, opt.abund_fixedScale = TRUE,
  opt.no_CPU = 1)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

no.years
an integer corresponding to the number of simulation years that will be used to extract PFG abundance maps

opt.abund_fixedScale
default TRUE. If FALSE, the ordinate scale will be adapted for each PFG for the graphical representation of the evolution of abundances through time

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

Value

Two POST_FATE_[...].pdf files are created :

GRAPHIC_A
spaceOccupancy

to visualize for each PFG the evolution of its occupation of the studied area through simulation time

GRAPHIC_B
abundance

to visualize for each PFG the evolution of its abundance within the whole studied area through simulation time

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, two preanalytical graphics.

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_perPFG_allStrata and unzipped. Informations extracted lead to the production of two graphics before the maps are compressed again :

the evolution of space occupancy of each Plant Functional Group through simulation time, with space occupancy representing the percentage of pixels within the mask of studied area where the PFG is present

the evolution of abundance of each Plant Functional Group through simulation time, with abundance being the sum over the whole studied area of the PFG abundances (FATE-HD arbitrary unit)

See also

Examples

# NOT RUN {
# }
# NOT RUN {
POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.no_CPU = 1)

POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.no_CPU = 4)

POST_FATE.graphic_evolutionCoverage(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , no.years = 50
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 1)
# }# NOT RUN {
                                    
                                    

# }