Create a graphical representation of the evolution of PFG abundance through time for 5 (or more) pixels of a FATE-HD simulation

This script is designed to produce one graphical representation for a FATE-HD simulation : the evolution through time of the abundance of each PFG for 5 (or more) randomly selected cells of the studied area.

POST_FATE.graphic_evolutionPixels(name.simulation,
  file.simulParam = NULL, opt.abund_fixedScale = TRUE,
  opt.cells_ID = NULL, opt.doPlot = TRUE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

opt.abund_fixedScale
default TRUE. If FALSE, the ordinate scale will be adapted for each PFG for the graphical representation of the evolution of abundances through time

opt.cells_ID
default NULL (optional). The cells ID of the studied area for which PFG abundances will be extracted.

opt.doPlot
default TRUE (optional). If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned.

Value

A list containing one data.frame object with the following columns, and one ggplot2 object :

tab

TYPE

the concerned information (either 'light', 'abundance' or 'soil')

GROUP

the concerned entity (either 'STRATUM_[...]', PFG name or 'soil')

ID

the id of the concerned pixel

HAB

the habitat of the concerned pixel

YEAR

the concerned simulation year

Occupancy

the number of occupied pixels divided by the total number of pixels within the studied area

plot

ggplot2 object, representing the evolution of each PFG abundance, and light and soil resources if those modules are activated

One POST_FATE_TABLE_PIXEL_evolution_pixels_[...].csv file is created :

pixels ids

always, containing the data.frame detailed above

One POST_FATE_[...].pdf file is created :

GRAPHIC_A
pixels

to visualize for each PFG the evolution of its abundance within each selected pixel through simulation time

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, one preanalytical graphics :

the evolution of abundance of each Plant Functional Group through simulation time, within 5 (or more) randomly selected pixels of the studied area (FATE-HD arbitrary unit)

if LIGHT is activated (see PRE_FATE.params_globalParameters), evolution of light resources within the selected pixels is also represented (1: Low, 2: Medium, 3: High)

if SOIL is activated (see PRE_FATE.params_globalParameters), evolution of soil resources within the selected pixels is also represented (user-defined scale)

It requires that the POST_FATE.temporalEvolution function has been run and that the file POST_FATE.evolution_abundance_PIXEL_[...].csv exists (as well as the files POST_FATE.evolution_light_PIXEL_[...].csv and POST_FATE.evolution_soil_PIXEL_[...].csv if those modules are activated).

See also

Examples

if (FALSE) {
POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt")

POST_FATE.graphic_evolutionPixels(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , opt.abund_fixedScale = FALSE)
}