Create a map of the Plant Functional Group richness for one (or several) specific year of a FATE-HD simulation
Source: R/POST_FATE.graphic_validationAUC.R

This script is designed to produce a raster map of PFG richness for one (or several) specific FATE-HD simulation year.

POST_FATE.graphic_validationStatistics(name.simulation,
  file.simulParam = NULL, year, mat.PFG.obs, opt.no_CPU = 1)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG abundance maps

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

opt.strata
default ALL (optional). The stratum number from which to extract PFG abundance maps

Value

One POST_FATE_[...].pdf file is created :

GRAPHIC_C
PFGrichness

to visualize the PFG richness within the studied area

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, one preanalytical graphic.

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_perPFG_allStrata (unless the opt.strata is used, then it will be from the folder ABUND_perPFG_perStrata) and unzipped. Informations extracted lead to the production of one graphic before the maps are compressed again :

the map of Plant Functional Group richness for each selected simulation year(s), representing the number of PFG present in each pixel

Examples

# NOT RUN {
# }
# NOT RUN {
POST_FATE.graphic_mapPFGrichness(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1)

POST_FATE.graphic_mapPFGrichness(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = c(850, 950)
                                 , opt.no_CPU = 1)

POST_FATE.graphic_mapPFGrichness(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1
                                 , opt.strata = 2)
# }# NOT RUN {
                                    
                                    

# }