Create a graphical representation of several statistics for each PFG to asses the quality of the model
for one (or several) specific year of a FATE-HD simulation

This script is designed to produce a graphical representation of several statistics (sensitivity, specificity, TSS, AUC) for quality assessment for one (or several) specific FATE-HD simulation year.

POST_FATE.graphic_validationStatistics(name.simulation,
  file.simulParam = NULL, year, mat.PFG.obs, opt.ras_habitat = NULL,
  opt.no_CPU = 1, opt.doPlot = TRUE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG binary maps

mat.PFG.obs
a data.frame with 4 columns : PFG, X, Y, obs

opt.ras_habitat
default NULL (optional). A string that corresponds to the file name of a raster mask, with an integer value within each pixel, corresponding to a specific habitat

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

opt.doPlot
default TRUE (optional). If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned.

Value

A list containing one data.frame object with the following columns, and one ggplot2 object :

tab

PFG

the concerned Plant Functional Group

AUC.sd

standard deviation of the AUC values

sensitivity.sd

standard deviation of the sensitivity values

specificity.sd

standard deviation of the specificity values

variable

name of the calculated statistic among 'sensitivity', 'specificity', 'TSS' and 'AUC'

value

value of the corresponding statistic

plot

ggplot2 object, representing the values for each PFG of these four validation statistics (sensitivity, specificity, TSS, AUC)

One POST_FATE_TABLE_YEAR_[...].csv file is created :

validationStatistics

containing the data.frame detailed above

Two folders are created :

BIN_perPFG
_allStrata

containing presence / absence raster maps for each PFG across all strata

BIN_perPFG
_perStrata

containing presence / absence raster maps for each PFG for each stratum

One POST_FATE_[...].pdf file is created :

GRAPHIC_B
validationStatistics

to assess the modeling quality of each PFG based on given observations within the studied area

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, raster maps of PFG presence / absence and one preanalytical graphic.

Observation records (presences and absences) are required for each PFG within the mat.PFG.obs object :

PFG

the concerned Plant Functional Group

X and Y

the coordinates of each observations, matching with the projection of the mask of name.simulation

obs

either 0 or 1 to indicate presences or absences

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_REL_perPFG_allStrata and unzipped. Informations extracted lead to the production of presence/absence maps and one graphic before the maps are compressed again :

for each selected simulation year(s), the value of several statistics to evaluate the predictive quality of the model for each Plant Functional Group (sensitivity, specificity, TSS, AUC)

for each selected simulation year(s), presence / absence maps of each Plant Functional Group within each stratum are obtained by finding the best cutoff to transform abundanceq in 0/1 while optimizing the values of sensitivity and specificity (see .getCutoff)

If a raster mask for habitat has been provided, the graphics will be also done per habitat.

It requires that the POST_FATE.relativeAbund function has been run and that the folder ABUND_REL_perPFG_allStrata exists.

These binary raster files can then be used by other functions :

to produce graphics of PFG abundances vs PFG Habitat Suitability maps (see POST_FATE.graphic_mapPFGvsHS)

See also

Examples

if (FALSE) {
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                       , file.simulParam = "Simul_parameters_V1.txt"
                                       , year = 850
                                       , mat.PFG.obs =
                                       , opt.no_CPU = 1)

POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                       , file.simulParam = "Simul_parameters_V1.txt"
                                       , year = c(850, 950)
                                       , mat.PFG.obs =
                                       , opt.no_CPU = 1)
}

## ----------------------------------------------------------------------------------------- ##
## Load example data
PNE_PFG = .loadData("PNE_PFG")
PNE_PARAM = .loadData("PNE_PARAM")
PNE_RESULTS = .loadData("PNE_RESULTS")

## PNE_PFG$PFG.observations : data.frame
## PNE_PARAM$succ_light : data.frame
## PNE_PARAM$strata_limits : vector
## PNE_PARAM$disp : data.frame
## PNE_PARAM$dist : data.frame
## PNE_PARAM$global : vector
## PNE_PARAM$masks : rasterStack
## PNE_RESULTS$abund_str.equilibrium : rasterStack

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_PNE")

## Create PFG succession parameter files : predefined of strata limits
tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity")]
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_PNE"
                              , mat.PFG.succ = tab
                              , strata.limits = PNE_PARAM$strata_limits
                              , strata.limits_reduce = FALSE)

## Create PFG light parameter files : predefined of strata limits
tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity", "light")]
PRE_FATE.params_PFGlight(name.simulation = "FATE_PNE"
                         , mat.PFG.succ = tab
                         , strata.limits = PNE_PARAM$strata_limits
                         , strata.limits_reduce = FALSE)

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_PNE"
                             , mat.PFG.disp = PNE_PARAM$disp)

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_PNE"
                               , mat.PFG.dist = PNE_PARAM$dist)

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_PNE"
                                 , required.no_PFG = PNE_PARAM$global["NB_FG"]
                                 , required.no_STRATA = PNE_PARAM$global["NB_STRATUM"]
                                 , required.simul_duration = PNE_PARAM$global["SIMULATION_DURATION"]
                                 , required.seeding_duration = PNE_PARAM$global["SEEDING_DURATION"]
                                 , required.seeding_timestep = PNE_PARAM$global["SEEDING_TIMESTEP"]
                                 , required.seeding_input = PNE_PARAM$global["SEEDING_INPUT"]
                                 , required.max_abund_low = PNE_PARAM$global["MAX_ABUND_LOW"]
                                 , required.max_abund_medium = PNE_PARAM$global["MAX_ABUND_MEDIUM"]
                                 , required.max_abund_high = PNE_PARAM$global["MAX_ABUND_HIGH"]
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = PNE_PARAM$global["LIGHT.thresh_medium"]
                                 , LIGHT.thresh_low = PNE_PARAM$global["LIGHT.thresh_low"]
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = PNE_PARAM$global["DISPERSAL.mode"]
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = PNE_PARAM$global["HABSUIT.ref_option"]
                                 , doDisturbances = TRUE
                                 , DIST.no = PNE_PARAM$global["DIST.no"]
                                 , DIST.no_sub = PNE_PARAM$global["DIST.no_sub"]
                                 , DIST.freq = rep(PNE_PARAM$global["DIST.freq"]
                                                   , PNE_PARAM$global["DIST.no"])
)

## Create simulation masks
library(raster)
writeRaster(PNE_PARAM$masks$maskEcrins
            , file = "FATE_PNE/DATA/MASK/mask.tif"
            , overwrite = TRUE)
writeRaster(PNE_PARAM$masks$noDisturb
            , file = "FATE_PNE/DATA/MASK/noDisturb.tif"
            , overwrite = TRUE)

## Create simulation parameters file
PRE_FATE.params_simulParameters(name.simulation = "FATE_PNE"
                                , name.mask = "mask.tif"
                                , name.dist = "noDisturb.tif")

## Create results folders
name.folder = "FATE_PNE"
name.simul = "SIMUL_V1"
dir1 = paste0(name.folder, "/RESULTS/", name.simul, "/ABUND_perPFG_allStrata")
dir2 = paste0(name.folder, "/RESULTS/", name.simul, "/ABUND_perPFG_perStrata")
dir3 = paste0(name.folder, "/RESULTS/", name.simul, "/LIGHT")
dir4 = paste0(name.folder, "/RESULTS/", name.simul, "/SOIL")

dir.create(dir1, recursive = TRUE)
dir.create(dir2, recursive = TRUE)
dir.create(dir3, recursive = TRUE)
dir.create(dir4, recursive = TRUE)

## Create results files
PFG.names = PNE_PARAM$succ_light$PFG
PFG.short = sapply(PFG.names, function(x) strsplit(x, "_")[[1]][1])
for (pfg in PFG.names)
{
  ind = grep(pfg, names(PNE_RESULTS$abund_str.equilibrium))
  stk = PNE_RESULTS$abund_str.equilibrium[[ind]]
  writeRaster(stk
              , filename = paste0(dir2, "/Abund_YEAR_800_", pfg, "_STRATA_"
                                  , sub(".*str", "", names(stk)), ".tif")
              , overwrite = TRUE
              , bylayer = TRUE)
  ras = sum(stk)
  writeRaster(ras
              , filename = paste0(dir1, "/Abund_YEAR_800_", pfg, "_STRATA_all.tif")
              , overwrite = TRUE)
}

## Create relative abundance maps
POST_FATE.relativeAbund(name.simulation = "FATE_PNE"
                        , file.simulParam = "Simul_parameters_V1.txt"
                        , year = 800
                        , opt.no_CPU = 1)

## Create binary maps
library(reshape2)
tab = PNE_PFG$PFG.observations
tab = melt(tab, id.vars = c("sites", "X", "Y"))
colnames(tab) = c("sites", "X", "Y", "PFG", "obs")
tab = tab[, c("PFG", "X", "Y", "obs")]
tab = tab[which(tab$PFG != "Others"), ]
tab$PFG = sapply(tab$PFG, function(x) names(PFG.short)[which(PFG.short == x)])
tab$obs = ifelse(tab$obs > 0, 1, 0)
str(tab)

validStats = POST_FATE.graphic_validationStatistics(name.simulation = "FATE_PNE"
                                                    , file.simulParam = "Simul_parameters_V1.txt"
                                                    , year = 800
                                                    , mat.PFG.obs = tab
                                                    , opt.no_CPU = 1)

str(validStats$`FATE_PNE/PARAM_SIMUL/Simul_parameters_V1.txt`$tab$`800`)
plot(validStats$`FATE_PNE/PARAM_SIMUL/Simul_parameters_V1.txt`$plot$`800`$ALL)