Create a graphical representation of several statistics for each PFG to asses the quality of the model for one (or several) specific year of a FATE-HD simulation

This script is designed to produce a graphical representation of several statistics (sensitivity, specificity, TSS, AUC) for quality assessment for one (or several) specific FATE-HD simulation year.

POST_FATE.graphic_validationStatistics(name.simulation,
  file.simulParam = NULL, year, mat.PFG.obs, opt.no_CPU = 1)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

file.simulParam
a string that corresponds to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE-HD simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG abundance maps

mat.PFG.obs
a data.frame with 4 columns : PFG, X, Y, obs

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

Value

A data.frame with the following columns :

PFG

the concerned Plant Functional Group

AUC.sd

standard deviation of the AUC values

sensitivity.sd

standard deviation of the sensitivity values

specificity.sd

standard deviation of the specificity values

variable

name of the calculated statistic among 'sensitivity', 'specificity', 'TSS' and 'AUC'

value

value of the corresponding statistic

One POST_FATE_[...].pdf file is created :

GRAPHIC_E
validationStatistics

to assess the modeling quality of each PFG based on given observations within the studied area

Details

This function allows one to obtain, for a specific FATE-HD simulation and a specific parameter file within this simulation, one preanalytical graphic.

For each PFG and each selected simulation year, raster maps are retrieved from the results folder ABUND_perPFG_allStrata and unzipped. Informations extracted lead to the production of one graphic before the maps are compressed again :

the value of several statistics for the predictive quality of the model for each Plant Functional Group and for each selected simulation year(s)

Observation records (presences and absences) are required for each PFG within the mat.PFG.obs obejct :

PFG

the concerned Plant Functional Group

X and Y

the coordinates of each observations, matching with the projection of the mask of name.simulation

obs

either 0 or 1 to indicate presences or absences

Examples

# NOT RUN {
# }
# NOT RUN {
POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = 850
                                 , opt.no_CPU = 1)

POST_FATE.graphic_validationStatistics(name.simulation = "FATE_simulation"
                                 , file.simulParam = "Simul_parameters_V1.txt"
                                 , year = c(850, 950)
                                 , opt.no_CPU = 1)
# }# NOT RUN {
                                    
                                    

# }