Create all possible graphical representations for a FATE simulation

This script is designed to produce a set of graphical representations for a FATE simulation. Graphics can be of three types : 1) representing an evolution through time (of abundance, light, soil) ; 2) vizualising the goodness of the modelisation (presence/absence, validation statistics) : 3) or representing a spatial distribution for a specific year (richness, abundance, light, soil).

POST_FATE.graphics(
  name.simulation,
  file.simulParam,
  year,
  no.years,
  opt.strata_min = 1,
  opt.strata = "ALL",
  opt.mat.PFG.succ = NULL,
  opt.mat.PFG.obs = NULL,
  opt.mat.CWM.obs = NULL,
  opt.mat.cover.obs = NULL,
  opt.mat.light.obs = NULL,
  opt.mat.soil.obs = NULL,
  opt.ras.cover.obs = NULL,
  opt.ras.light.obs = NULL,
  opt.ras.soil.obs = NULL,
  opt.ras_habitat = NULL,
  opt.cells_ID = NULL,
  opt.abund_fixedScale = TRUE,
  opt.no_CPU = 1,
  opt.doPlot = TRUE,
  opt.doFunc.evolutionCoverage = TRUE,
  opt.doFunc.evolutionPixels = TRUE,
  opt.doFunc.validation = TRUE,
  opt.doFunc.mapPFGvsHS = TRUE,
  opt.doFunc.mapPFGrichness = TRUE,
  opt.doFunc.mapPFGcover = TRUE,
  opt.doFunc.mapPFGlight = TRUE,
  opt.doFunc.mapPFGsoil = TRUE
)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE simulation

file.simulParam
default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

year
an integer corresponding to the simulation year(s) that will be used to extract PFG abundance and/or binary maps, as well as light or soil resources

no.years
an integer corresponding to the number of simulation years that will be used to extract PFG abundance maps

opt.strata_min
default 1 (optional). An integer corresponding to the lowest stratum from which PFG abundances are summed up to the highest stratum

opt.strata
default ALL (optional). The stratum number from which to extract PFG binary maps

opt.mat.PFG.succ
default NULL (optional). A data.frame with 2 or 3 columns : PFG, light, soil_contrib

opt.mat.PFG.obs
default NULL (optional). A data.frame with 4 columns : PFG, X, Y, obs (0/1)

opt.mat.CWM.obs
default NULL (optional). A data.frame with 3 or 4 columns : X, Y, light (CWM), soil (CWM)

opt.mat.cover.obs
default NULL (optional). A data.frame with 3 columns : X, Y, obs (0/1)

opt.mat.light.obs
default NULL (optional). A data.frame with 3 columns : X, Y, obs (0/1)

opt.mat.soil.obs
default NULL (optional). A data.frame with 3 columns : X, Y, obs (0/1)

opt.ras.cover.obs
default NULL (optional). A string that corresponds to the file name of a raster containing observed values for vegetation cover

opt.ras.light.obs
default NULL (optional). A string that corresponds to the file name of a raster containing observed values for community weighted mean of light

opt.ras.soil.obs
default NULL (optional). A string that corresponds to the file name of a raster containing observed values for community weighted mean of soil

opt.ras_habitat
default NULL (optional). A string that corresponds to the file name of a raster mask, with an integer value within each pixel, corresponding to a specific habitat

opt.cells_ID
default NULL (optional). The cells ID of the studied area for which abund / light / soil resources will be extracted

opt.abund_fixedScale
default TRUE. If FALSE, the ordinate scale will be adapted for each PFG for the graphical representation of the evolution of abundances through time

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the unzip/zip of raster files

opt.doPlot
default TRUE (optional). If TRUE, plot(s) will be processed, otherwise only the calculation and reorganization of outputs will occur, be saved and returned

opt.doFunc.evolutionCoverage
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.evolutionPixels
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.validation
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.mapPFGvsHS
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.mapPFGrichness
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.mapPFGcover
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.mapPFGlight
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

opt.doFunc.mapPFGsoil
default TRUE (optional). If TRUE, corresponding POST_FATE.graphic_... function will be run.

Value

The following POST_FATE_[...].pdf files are created :

GRAPHIC_A
spaceOccupancy

to visualize for each PFG the evolution of its occupation of the studied area through simulation time

GRAPHIC_A
abundance

to visualize for each PFG the evolution of its abundance within the whole studied area through simulation time

GRAPHIC_A
pixels

to visualize for each PFG the evolution of its abundance within each selected pixel through simulation time, as well as the evolution of light and soil resources

GRAPHIC_B
validationStatistics

to assess the modeling quality of each PFG based on given observations within the studied area

GRAPHIC_B
PFGvsHS

to visualize the PFG presence within the studied area (probability and simulated occurrence)

GRAPHIC_C
PFGrichness

to visualize the PFG richness within the studied area

GRAPHIC_C
PFGcover

to visualize the PFG cover within the studied area

GRAPHIC_C
PFGlight

to visualize the light CWM within the studied area

GRAPHIC_C
PFGlsoil

to visualize the soil CWM within the studied area

Three folders are created :

ABUND_REL_perPFG
_allStrata

containing relative abundance raster maps for each PFG across all strata (see POST_FATE.relativeAbund)

BIN_perPFG
_allStrata

containing presence / absence raster maps for each PFG across all strata (see POST_FATE.graphic_validationStatistics)

BIN_perPFG
_perStrata

containing presence / absence raster maps for each PFG for each stratum (see POST_FATE.graphic_validationStatistics)

Details

This function allows one to obtain, for a specific FATE simulation and a specific parameter file within this simulation, up to ten preanalytical graphics.

For each PFG and each selected simulation year, raster maps are retrieved from the results folders (ABUND_perPFG_allStrata or ABUND_perPFG_allStrata or ABUND_REL_perPFG_allStrata or BIN_perPFG_perStrata or BIN_perPFG_allStrata or LIGHT or SOIL) and unzipped. Informations extracted lead to the production of the following graphics before the maps are compressed again :

the evolution of space occupancy of each Plant Functional Group through simulation time, with space occupancy representing the percentage of pixels within the mask of studied area where the PFG is present (see POST_FATE.graphic_evolutionCoverage)

the evolution of abundance of each Plant Functional Group through simulation time, with abundance being the sum over the whole studied area of the PFG abundances (FATE arbitrary unit) (see POST_FATE.graphic_evolutionCoverage)

the evolution of abundance of each Plant Functional Group through simulation time, within 5 (or more) randomly selected pixels of the studied area (FATE arbitrary unit), as well as light resources of each stratum (1: Low, 2: Medium, 3: High) and soil resources if these modules are selected (see POST_FATE.graphic_evolutionPixels)

the value of several statistics for the predictive quality of the model for each Plant Functional Group and for each selected simulation year(s) (see POST_FATE.graphic_validationStatistics)

the maps of Plant Functional Group habitat suitability and occurrences for each selected simulation year(s), representing the probability of presence of each PFG in each pixel compared to its simulated presence (see POST_FATE.graphic_mapPFGvsHS)

the map of Plant Functional Group richness for each selected simulation year(s), representing the number of PFG present in each pixel (see POST_FATE.graphic_mapPFGrichness)

the map of Plant Functional Group cover for each selected simulation year(s), representing the relative cumulated abundance of PFG present in each pixel above a height threshold (see POST_FATE.graphic_mapPFGcover)

the map of light Community Weighted Mean for each selected simulation year(s), representing the simulated value of light within each pixel above a height threshold (see POST_FATE.graphic_mapPFGlight)

the map of soil Community Weighted Mean for each selected simulation year(s), representing the simulated value of soil within each pixel (see POST_FATE.graphic_mapPFGsoil)

See also

Examples

if (FALSE) {
POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10)

POST_FATE.graphics(name.simulation = "FATE_simulation"
                                    , file.simulParam = "Simul_parameters_V1.txt"
                                    , year = 100
                                    , no.years = 10
                                    , opt.abund_fixedScale = FALSE
                                    , opt.no_CPU = 4)
}