Compute validation data for habitat, PFG richness and composition for a FATE simulation.

This script is designed to compare observed and simulated PFG distribution for one specific FATE simulation year by computing a) difference in PFG richness ; b) similarity between distribution quantiles ; c) a random forest model to predict habitat classes.

POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  mat.obs,
  mat.hab,
  ras.habitat,
  doRichness = TRUE,
  doComposition = TRUE,
  doHabitat = TRUE,
  RF.seed = 123,
  RF.training = 0.7,
  doHabitat.allMap = FALSE,
  opt.ras_validation = NULL,
  opt.keep_PFG = NULL,
  opt.keep_strata = NULL
)

Arguments

name.simulation

a string corresponding to the main directory or simulation name of the FATE simulation

file.simulParam

default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation

year

an integer corresponding to the simulation year that will be used to extract PFG abundance table

mat.obs

a data.frame with at least 5 columns :
site, x, y, PFG, abund
(and optionally, strata, code.habitat) (see Details)

mat.hab

a data.frame with 2 columns : code.habitat, habitat

ras.habitat

a string corresponding to the file name of a raster mask, with an integer value within each pixel, corresponding to a specific habitat

doRichness

(logical) default TRUE.
If TRUE, difference in observed and simulated PFG richness will be run.

doComposition

(logical) default TRUE.
If TRUE, difference in observed and simulated PFG distribution over habitat will be run.

doHabitat

(logical) default TRUE.
If TRUE, a random forest model to predict habitat from PFG abundances will be run.

RF.seed

default 123.
An integer to be given to set.seed function, in order to fix the produced results if needed, as dataset will be divided randomly into training and testing datasets for the random forest model

RF.training

default 0.7.
A numeric between 0 and 1 corresponding to the percentage of data that will be used as training dataset in the random forest model

doHabitat.allMap

(logical) default TRUE.
If TRUE, habitat prediction over the whole simulation map will be run.

opt.ras_validation

(optional) default NULL.
A string corresponding to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the validation will take place

opt.keep_PFG

(optional) default NULL.
A vector of character corresponding to the names of the PFG to keep for the validation

opt.keep_strata

(optional) default NULL.
A list with names corresponding to the strata to keep for the validation, and each element containing correspondence with FATE strata definition

Value

Habitat

into the name.simulation/VALIDATION/HABITAT/ directory :
1 .csv file containing the prepared observed data (relative abundances & habitat information).
1 .rds file containing the RF model.
5 .csv files containing the performance analysis (confusion matrix and TSS) for the training & testing parts of the RF model and the final habitat performance.
If option selected, 1 .csv file containing the habitat prediction within each pixel, 1 .csv file with sucess or failure of the prediction and 1 .png prediction map.

PFG Composition

into the name.simulation/VALIDATION/PFG COMPOSITION/ directory :
1 .csv file containing the proximity between observed and simulated data computed for each PFG/strata/habitat.
1 .csv file containing the observed releves transformed into relative metrics.
1 .csv file containing the final output with the distribution per PFG, strata and habitat.

PFG richness

into the name.simulation/VALIDATION/PFG RICHNESS/ directory :
1 .csv file of PFG richness in a FATE simulation.
1 .csv fie of the PFG extinction frequency in a FATE simulation.
1 .rds file which is the abundance per PFG file.

Details

This function allows to obtain, for a specific FATE simulation and a specific parameter file within this simulation, PFG validation habitat model and predictions and one preanalytical graphic.

Habitat validation

Observed habitat

is provided by the ras.habitat map. The final set of habitats taken into account in the validation is provided by mat.hab table which contains all the habitats, with their corresponding ID.

Simulated habitat

is determined from FATE simulated relative abundances (see POST_FATE.temporalEvolution), thanks to a Random Forest algorithm. The model is trained on mat.obs data. Information about PFG abundances at each sites with xy coordinates are necessary. PFG abundances can be given in absolute abundance (only in sites where at least one PFG is present) or in presence-absence data.
Eventually, habitat ID information can be provided, as well as strata names. If not, habitat information will be taken from ras.habitat map, and PFG abundance will be considered for all strata.

To compare observations and simulations, the function computes confusion matrix between observations and predictions and then computes the TSS for each habitat h. $$TSS_{\text{ h}} = (\frac{\text{True Positives}_{\text{ h}}}{\text{True Positives}_{\text{ h}}{\text{ + False Negatives}_{\text{ h}}}} + \frac{\text{True Negatives}_{\text{ h}}}{\text{True Negatives}_{\text{ h}}{\text{ + False Positives}_{\text{ h}}}}) - \text{1}$$ The final metric used is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed habitat distribution. The habitat validation also provides a visual comparison of observed and simulated habitat on the whole studied area, if option selected.

PFG composition validation

Observed composition

is computed for a set of PFG, habitat & strata (for all strata if strata definition is not activated) by computing 4 quartiles of the composition, based on mat.obs data provided. Habitats are chosen in mat.hab, PFG list is defined in the file file.simulParam and, if defined, mat.obs strata are taken into account. Optionally, PFGs contained in opt.keep_PFG can be excluded from the analysis.

Simulated composition

is computed in the same way than the observed composition, with FATE abundances
(see POST_FATE.temporalEvolution).

Then, a composition similarity between each habitat/strata combinations is provided by computing a pseudo-distance between observed and simulated quartiles for each PFG y. $$S_{\text{ habitat, strata}} = \sum S_{\text{ y}{\text{, }habitat}{\text{, }strata}}$$ with $$S_{\text{ y}{\text{, }habitat}{\text{, }strata}} = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ i}{\text{, }sim}} - Q_{\text{ i}{\text{, }obs}})$$ with i varying from 1 to 4.

PFG richness validation

the observed PFG richness corresponds to the PFG list contained in the file file.simulParam, the simulated PFG richness is the number of PFG for which abundance over the simulation area is strictly superior to zero for the simulation year under scrutiny.
Then, observed and simulated richness are compared for all the PFGs in the simulation in order to quantify the PFG mortality. Optionally, PFGs contained in opt.keep_PFG can be excluded from the analysis.

Author

Matthieu Combaud, Maxime Delprat, Maya Guéguen

Examples

library(raster)

## Create a simulation folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")
.loadData("Champsaur_results_V1", "7z")

## Please extract results files in the 'FATE_Champsaur/RESULTS' folder

## Define a vector to choose habitats taken into account
mat.hab = data.frame(ID = c(6, 5, 7, 8)
                             , habitat = c("coniferous.forest"
                                           , "deciduous.forest"
                                           , "natural.grassland"
                                           , "woody.heatland"))

## Habitat & validation maps
ras.habitaterved = Champsaur_params$stk.mask$habitat
ras_simulation = Champsaur_params$stk.mask$Champsaur
ras.habitat = projectRaster(from = ras.habitaterved, to = ras_simulation, method = "ngb")
writeRaster(ras_simulation, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
mat.obs = Champsaur_params$tab.releves

## Transform observed PFG abundances into relative abundances
mat.obs$abund = PRE_FATE.abundBraunBlanquet(mat.obs$abund)/100
mat.obs = aggregate(abund ~ site + PFG + strata + x + y
                        , data = mat.obs, FUN = "sum")
                        
## Create Global and Simulation parameters
PRE_FATE.params_globalParameters(name.simulation = "FATE_Champsaur"
                                , opt.saving_abund_PFG_stratum = TRUE
                                , opt.saving_abund_PFG = TRUE
                                , opt.saving_abund_stratum = FALSE
                                , required.no_PFG = 15
                                , required.no_strata = 7
                                , required.simul_duration = 2000
                                , required.seeding_duration = 1000
                                , required.seeding_timestep = 1
                                , required.seeding_input = 100
                                , required.potential_fecundity = 1
                                , required.max_abund_low = 1000
                                , required.max_abund_medium = 2000
                                , required.max_abund_high = 3000
                                , doDispersal = TRUE
                                , DISPERSAL.mode = 1
                                , DISPERSAL.saving = FALSE
                                , doHabSuitability = TRUE
                                , HABSUIT.mode = 1)
                                
PRE_FATE.params_simulParameters(name.simulation = "FATE_Champsaur"
                                , name.MASK = "MASK_Champsaur.tif")

param = "Simul_parameters_V1.txt"
simul.param = paste0("FATE_Champsaur/PARAM_SIMUL/", param)

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , file.simulParam = simul.param
                     , year = 2000
                     , doHabitat = TRUE
                     , mat.obs = mat.obs
                     , ras.habitat = ras.habitat
                     , mat.hab = mat.hab
                     , opt.keep_strata = opt.keep_strata
                     , doHabitat.allMap = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)