Compute validation data for habitat, PFG richness and composition for a FATE simulation.

This script is designed to compute validation data for :

Habitat

Compares simulated and observed habitat and create a map to visualize this comparison with all the FATE and observed data
(if option selected).

PFG Composition

Computes the similarity between observed and simulated PFG composition for a chosen set of PFG/habitat/strata combinations.

PFG Richness

Computes the PFG richness over the whole simulation area for a FATE simulation and computes the difference between observed and simulated PFG richness.

POST_FATE.validation(
  name.simulation,
  file.simulParam,
  year,
  perStrata = FALSE,
  opt.no_CPU = 1,
  doHabitat = TRUE,
  releves.PFG,
  hab.obs,
  studied.habitat,
  predict.all.map = FALSE,
  RF.seed = 123,
  RF.training = 0.7,
  validation.mask = NULL,
  list.strata.simulations = NULL,
  doComposition = TRUE,
  doRichness = TRUE,
  exclude.PFG = NULL
)

Arguments

name.simulation

a string corresponding to the simulation folder name.

file.simulParam

default NULL.
A string corresponding to the name of a parameter file that will be contained into the PARAM_SIMUL folder of the FATE simulation.

year

an integer corresponding to the year of simulation for validation.

perStrata

(logical) default FALSE.
If TRUE, habitat & PFG composition are computed with PFG abundance defined by strata.
If FALSE, habitat & PFG composition are computed with PFG abundance defined for all strata.

opt.no_CPU

default 1.
An integer corresponding to the number of resources that can be used to parallelize the computation of prediction performance for habitat & richness validation.

doHabitat

(logical) default TRUE.
If TRUE, habitat validation module is activated, if FALSE, habitat validation module is disabled.

releves.PFG

a data.frame with at least 5 columns :
site, x, y, abund, PFG
(and optionally, strata, code.habitat)
(see Details).

hab.obs

a raster map of the habitats within the studied area, with same projection & resolution than simulation mask.

studied.habitat

a data.frame with 2 columns :
ID ,habitat
(see Details).

predict.all.map

(logical) default FALSE.
If TRUE, the function will compute habitat prediction & performance over the whole map and will provide a prediction map.

RF.seed

default 123.
An integer corresponding to the number of seeds to set in order to generate a Random Forest model.

RF.training

default 0.7.
A double between 0 & 1 corresponding to the part of the data used for training a Random Forest model on releves.PFG data.

validation.mask

(optional) default NULL.
A raster map (with 0 or 1 in each pixel) that specified on which pixels the performance of the habitat prediction will be compute, with same projection & resolution than simulation mask
If NULL, the function will take the simulation mask (which means that the performance will be compute over the whole map).

list.strata.simulations

(optional) default NULL.
If perStrata = TRUE, a list which contains, for each stratum defined in releves.PFG, the correspondence with FATE strata definition.
If perStrata = FALSE, please specify NULL value.

doComposition

(logical) default TRUE.
If TRUE, PFG composition validation module is activated.

doRichness

(logical) default TRUE.
If TRUE, PFG richness validation module is activated.

exclude.PFG

(optional) default NULL.
A character vector containing the name(s) of the PFG(s) that will be excluded from the composition and richness analysis.

Value

Habitat

into the name.simulation/VALIDATION/HABITAT/ directory :
1 .csv file containing the prepared observed data (relative abundances & habitat information).
1 .rds file containing the RF model.
5 .csv files containing the performance analysis (confusion matrix and TSS) for the training & testing parts of the RF model and the final habitat performance.
If option selected, 1 .csv file containing the habitat prediction within each pixel, 1 .csv file with sucess or failure of the prediction and 1 .png prediction map.

PFG Composition

into the name.simulation/VALIDATION/PFG COMPOSITION/ directory :
1 .csv file containing the proximity between observed and simulated data computed for each PFG/strata/habitat.
1 .csv file containing the observed releves transformed into relative metrics.
1 .csv file containing the final output with the distribution per PFG, strata and habitat.

PFG richness

into the name.simulation/VALIDATION/PFG RICHNESS/ directory :
1 .csv file of PFG richness in a FATE simulation.
1 .csv fie of the PFG extinction frequency in a FATE simulation.
1 .rds file which is the abundance per PFG file.

Details

Habitat validation

Observed habitat

is provided by the hab.obs map. The final set of habitats taken into account in the validation is provided by studied.habitat table which contains all the habitats, with their corresponding ID.

Simulated habitat

is determined from FATE simulated relative abundances (see POST_FATE.temporalEvolution), thanks to a Random Forest algorithm. The model is trained on releves.PFG data. Information about PFG abundances at each sites with xy coordinates are necessary.
Eventually, habitat ID information can be provided, as well as strata name. If not, habitat information will be taken from hab.obs map, and PFG abundance will be considered for all strata.

To compare observations and simulations, the function computes confusion matrix between observations and predictions and then computes the TSS for each habitat h. $$TSS_{\text{ h}} = (\frac{\text{number of prediction}_{\text{ h}}}{\text{number of observation}_{\text{ h}}} + \frac{\text{number of non-prediction}_{\text{ h}}}{\text{number of non-observation}_{\text{ h}}}) - \text{1}$$ The final metric used is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed habitat distribution. The habitat validation also provides a visual comparison of observed and simulated habitat on the whole studied area, if option selected.

PFG composition validation

Observed composition

is computed for a set of PFG, habitat & strata (for all strata if strata definition is not activated) by computing 4 quartiles of the composition, based on releves.PFG data provided. Habitats are chosen in studied.habitat, PFG list is defined in the file file.simulParam and, if defined, releves.PFG strata are taken into account. Optionally, PFGs contained in exclude.PFG can be excluded from the analysis.

Simulated composition

is computed in the same way than the observed composition, with FATE abundances
(see POST_FATE.temporalEvolution).

Then, a composition similarity between each habitat/strata combinations is provided by computing a pseudo-distance between observed and simulated quartiles for each PFG y. $$S_{\text{ habitat, strata}} = \sum S_{\text{ y}{\text{, }habitat}{\text{, }strata}}$$ with $$S_{\text{ y}{\text{, }habitat}{\text{, }strata}} = 1 - \frac{\text{1}}{4} * \sum abs(Q_{\text{ i}{\text{, }sim}} - Q_{\text{ i}{\text{, }obs}})$$ with i varying from 1 to 4.

PFG richness validation

the observed PFG richness corresponds to the PFG list contained in the file file.simulParam, the simulated PFG richness is the number of PFG for which abundance over the simulation area is strictly superior to zero for the simulation year under scrutiny.
Then, observed and simulated richness are compared for all the PFGs in the simulation in order to quantify the PFG mortality. Optionally, PFGs contained in exclude.PFG can be excluded from the analysis.

Author

Matthieu Combaud, Maxime Delprat, Maya Guéguen

Examples

library(raster)

## Load example data
Champsaur_params = .loadData("Champsaur_params", "RData")

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_Champsaur")

## Load results from a simulation
.loadData("Champsaur_results_V1", "7z")

## Please extract files from the 7z folder in 'FATE_CHAMPSAUR/RESULTS'

## Define a vector to choose habitats taken into account
studied.habitat = data.frame(ID = c(6, 5, 7, 8), habitat = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland"))

## Habitat & validation maps
hab.observed = Champsaur_params$stk.mask$habitat
simulation.map = Champsaur_params$stk.mask$Champsaur
hab.obs = projectRaster(from = hab.observed, res = res(simulation.map)[1], crs = crs(projection(simulation.map)), method = "ngb")
writeRaster(simulation.map, filename = "FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")

## Observed data
releves.PFG = Champsaur_params$tab.releves

## List of PFG taken into account in a FATE simulation
list.PFG = c("C1", "C2", "C3", "C4", "H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5")

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , sim.version = "SIMUL_V1.1"
                     , year = 2000
                     , doHabitat = TRUE
                     , releves.PFG = releves.PFG
                     , hab.obs = hab.obs
                     , studied.habitat = studied.habitat
                     , predict.all.map = TRUE
                     , doComposition = TRUE
                     , doRichness = TRUE)