Computes validation data for habitat, PFG richness and composition for a FATE simulation.

This script is designed to compute validation data for : Habitat : compares habitat simulations and observations and create a map to visualize this comparison with all the FATE and observed data. PFG Composition : produced a computation of observed distribution of relative abundance in the simulation area and a computation of distance between observed and simulated distribution. PFG Richness : computes the PFG richness over the whole simulation area for a FATE simulation and computes the difference between observed and simulated PFG richness.

POST_FATE.validation(
  name.simulation,
  sim.version,
  year,
  doHabitat = TRUE,
  obs.path,
  releves.PFG,
  releves.sites,
  hab.obs,
  validation.mask,
  studied.habitat = NULL,
  doComposition = TRUE,
  PFG.considered_PFG.compo,
  habitat.considered_PFG.compo,
  doRichness = TRUE,
  list.PFG,
  exclude.PFG = NULL
)

Arguments

name.simulation

simulation folder name.

sim.version

name of the simulation to validate (it works with only one sim.version).

year

year of simulation for validation.

doHabitat

logical. Default TRUE. If TRUE, habitat validation module is activated, if FALSE, habitat validation module is disabled.

obs.path

the function needs observed data, please create a folder for them in your simulation folder and then indicate in this parameter the access path to this new folder (habitat & PFG composition validation).

releves.PFG

name of file which contain the observed Braund-Blanquet abundance at each site and each PFG and strata (habitat & PFG composition validation).

hab.obs

name of the file which contain the extended studied map in the simulation (habitat & PFG composition validation).

validation.mask

name of the file which contain a raster mask that specified which pixels need validation (habitat & PFG composition validation).

studied.habitat

default NULL. If NULL, the function will take into account of all habitats in the hab.obs map. Otherwise, please specify in a vector habitats that will be take into account for the validation (habitat validation).

doComposition

logical. Default TRUE. If TRUE, PFG composition validation module is activated, if FALSE, PFG composition validation module is disabled.

PFG.considered_PFG.compo

a character vector of the list of PFG considered in the validation (PFG composition validation).

habitat.considered_PFG.compo

a character vector of the list of habitat(s) considered in the validation (PFG composition validation).

doRichness

logical. Default TRUE. If TRUE, PFG richness validation module is activated, if FALSE, PFG richness validation module is disabled.

list.PFG

a character vector which contain all the PFGs taken account in the simulation and observed in the simulation area (PFG richness validation).

exclude.PFG

default NULL. A character vector containing the names of the PFG you want to exclude from the analysis (PFG richness validation).

releves.site

name of the file which contain coordinates and a description of the habitat associated with the dominant species of each site in the studied map (habitat & PFG composition validation).

Value

Details

Habitat validation

The observed habitat is derived from a map of the area, the simulated habitat is derived from FATE simulated relative abundance, based on a random forest algorithm trained on observed data. To compare observations and simulations, the function computes confusion matrix between observation and prediction and then computes the TSS for each habitat h (number of prediction of habitat h/number of observation of habitat h + number of non-prediction of habitat h/number of non-observation of habitat h). The final metrics this script use is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed habitat distribution.

PFG composition validation

This code firstly run the get.observed.distribution function in order to have a obs.distri file which contain the observed distribution per PFG, strata and habitat. This file is also an argument for the do.PFG.composition.validation function run next. This second sub function provide the computation of distance between observed and simulated distribution.
NB : The argument strata.considered_PFG.compo is by default "A" in the 2 sub functions because it's easier for a FATE simulation to provide PFG abundances for all strata.
The argument perStrata.compo is by default NULL for the same reasons.

PFG richness validation

Firstly, the function updates the list.PFG with exclude.PFG vector. Then, the script takes the abundance per PFG file from the results of the FATE simulation and computes the difference between the list.PFG and all the PFG which are presents in the abundance file, in order to obtain the PFG richness for a simulation. The function also determine if an observed PFG is missing in the results of the simulation at a specific year.

Author

Matthieu Combaud, Maxime Delprat

Examples

## Habitat validation ---------------------------------------------------------------------------------
POST_FATE.validation(name.simulation = "FATE_Champsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = TRUE
                      , obs.path = "FATE_Champsaur/DATA_OBS/"
                      , releves.PFG = "releves.PFG.abundance.csv"
                      , releves.sites = "releves.sites.shp"
                      , hab.obs = "simplified.cesbio.map.grd"
                      , validation.mask = "certain.habitat.100m.restricted.grd"
                      , studied.habitat = NULL
                      , doComposition = FALSE
                      , doRichness = FALSE)
                      
## PFG composition validation --------------------------------------------------------------------------
list.PFG<-as.factor(c("C1","C2","C3","C4","H1","H2","H3","H4","H5","H6","P1","P2","P3","P4","P5"))
habitat.considered = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland")
POST_FATE.validation(name.simulation = "FATE_Champsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = FALSE
                      , obs.path = "FATE_Champsaur/DATA_OBS/"
                      , releves.PFG = "releves.PFG.abundance.csv"
                      , releves.sites = "releves.sites.shp"
                      , hab.obs = "simplified.cesbio.map.grd"
                      , validation.mask = "certain.habitat.100m.restricted.grd"
                      , studied.habitat = NULL
                      , doComposition = TRUE
                      , PFG.considered_PFG.compo = list.PFG
                      , habitat.considered_PFG.compo = habitat.considered
                      , doRichness = FALSE)
                      
## PFG richness validation -----------------------------------------------------------------------------
list.PFG<-as.factor(c("C1","C2","C3","C4","H1","H2","H3","H4","H5","H6","P1","P2","P3","P4","P5"))
POST_FATE.validation(name.simulation = "FATE_CHampsaur"
                      , sim.version = "SIMUL_V4.1"
                      , year = 2000
                      , doHabitat = FALSE
                      , doComposition = FALSE
                      , doRichness = TRUE
                      , list.PFG = list.PFG
                      , exclude.PFG = NULL)