Computes validation data for habitat, PFG richness and composition for a FATE simulation.

This script is designed to compute validation data for :
Habitat : Compares habitat simulations and observations and create a map to visualize this comparison with all the FATE and observed data (if option selected).
PFG Composition : Produced a computation of observed distribution of relative abundance in the simulation area and a computation of distance between observed and simulated distribution.
PFG Richness : Computes the PFG richness over the whole simulation area for a FATE simulation and computes the difference between observed and simulated PFG richness.

POST_FATE.validation(
  name.simulation,
  sim.version,
  year,
  perStrata = FALSE,
  opt.no_CPU = 1,
  doHabitat = TRUE,
  releves.PFG,
  hab.obs,
  studied.habitat,
  predict.all.map = FALSE,
  seed,
  validation.mask = NULL,
  list.strata.simulations = NULL,
  doComposition = TRUE,
  PFG.considered_PFG.compo,
  habitat.considered_PFG.compo,
  strata.considered_PFG.compo = "A",
  doRichness = TRUE,
  list.PFG,
  exclude.PFG = NULL
)

Arguments

name.simulation

Simulation folder name.

sim.version

A character vector with the name(s) of the simulation(s) to validate.

year

Year of simulation for validation.

perStrata

Logical.
Default FALSE. If TRUE, PFG abundance is defined by strata. If FALSE, PFG abundance defined for all strata (habitat & PFG composition & PFG richness validation).

opt.no_CPU

Default 1.
The number of resources that can be used to parallelize the computation of prediction performance for habitat & richness validation.

doHabitat

Logical. Default TRUE.
If TRUE, habitat validation module is activated, if FALSE, habitat validation module is disabled.

releves.PFG

A data frame with abundance (column named abund) at each site and for each PFG and strata, if necessary, & sites coordinates (habitat & PFG composition validation).

hab.obs

A raster map of the extended studied map in the simulation, with same projection & resolution than simulation mask (habitat & PFG composition validation).

studied.habitat

A 2 columns data frame which contains the habitats (2nd column) and the ID (1st column) for each of them which will be taken into account for the validation (habitat & PFG composition validation).

predict.all.map

Default FALSE.
If TRUE, the function will compute habitat prediction over the whole map and will provide a prediction map.

seed

Numerical. Number of seeds to set in order to generate a Random Forest model.

validation.mask

(Optional). Default NULL.
A raster mask (with 0 or 1 in each pixel) that specified on which pixels the performance of the prediction will be compute, with same projection & resolution than simulation mask (habitat & PFG composition validation).
If NULL, the function will take the simulation mask (which means that the performance will be compute over the whole map)

list.strata.simulations

(Optional). Default NULL.
If perStrata = TRUE, a character vector which contain FATE strata definition and correspondence with observed strata definition. If perStrata = FALSE, please specify NULL value.

doComposition

Logical. Default TRUE.
If TRUE, PFG composition validation module is activated, if FALSE, PFG composition validation module is disabled.

PFG.considered_PFG.compo

A character vector of the list of PFG considered in the validation (PFG composition validation).

habitat.considered_PFG.compo

A character vector of the list of habitat(s) considered in the validation (PFG composition validation).

strata.considered_PFG.compo

Default "A".
If perStrata = FALSE, a character vector with value "A" (selection of one or several specific strata disabled). If perStrata = TRUE, a character vector with at least one of the observed strata (PFG composition validation).

doRichness

Logical. Default TRUE.
If TRUE, PFG richness validation module is activated, if FALSE, PFG richness validation module is disabled.

list.PFG

A character vector which contain all the PFGs taken account in the simulation and observed in the simulation area (PFG richness validation).

exclude.PFG

(Optional). Default NULL.
A character vector containing the names of the PFG you want to exclude from the analysis (PFG richness validation).

Value

Details

Habitat validation

The observed habitat is derived from a map of the area or, if defined, from studied.habitat, the simulated habitat is derived from FATE simulated relative abundance, based on a random forest algorithm trained on observed releves data (see train_RF_habitat).
To compare observations and simulations, the function computes confusion matrix between observations and predictions and then compute the TSS for each habitat h (number of prediction of habitat h/number of observation of habitat h + number of non-prediction of habitat h/number of non-observation of habitat h).
The final metrics this script use is the mean of TSS per habitat over all habitats, weighted by the share of each habitat in the observed habitat distribution. The habitat validation also provides a visual comparison of observed and simulated habitat on the whole studied area, if option selected (see do_habitat_validation & plot_predicted_habitat).

PFG composition validation

This code firstly run the get_observed_distribution function in order to have a obs.distri file which contain the observed distribution per PFG, strata and habitat. This file is also an argument for the do_PFG_composition_validation function run next. This second sub function provides the computation of distance between observed and simulated distribution.

PFG richness validation

Firstly, the function updates the list.PFG with exclude.PFG vector. Then, the script takes the abundance per PFG (and per strata if option selected) file from the results of the FATE simulation and computes the difference between the list.PFG and all the PFG which are presents in the abundance file, in order to obtain the PFG richness for a simulation. The function also determine if an observed PFG is missing in the results of the simulation at a specific year.

Author

Matthieu Combaud, Maxime Delprat

Examples

library(raster)
library(sf)

## Define a vector to choose habitats taken into account
studied.habitat = data.frame(ID = c(6, 5, 7, 8), habitat = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland"))
## Habitat & validation maps
hab.observed = raster("FATE_Champsaur/DATA_OBS/simplified.cesbio.map.grd")
simulation.map = raster("FATE_Champsaur/DATA/MASK/MASK_Champsaur.tif")
hab.obs = projectRaster(from = hab.observed, res = res(simulation.map)[1], crs = crs(projection(simulation.map)), method = "ngb")
## Observed data
releves.sites = as.data.frame(st_read("FATE_Champsaur/DATA_OBS/releves.sites.shp"))
releves.PFG = as.data.frame(read.csv("FATE_Champsaur/DATA_OBS/releves.PFG.abundance.csv"))
releves.PFG = merge(releves.PFG, releves.sites[,c("site","geometry")], by = "site")
coor = SpatialPoints(st_coordinates(releves.PFG$geometry), proj4string = crs(hab.observed))
coor = spTransform(coor, CRSobj = crs(hab.obs))
releves.PFG$geometry = NULL
releves.PFG[,c("x","y")] = coor@coords
colnames(releves.PFG) = c("site", "abund", "PFG", "strata", "x", "y")
## List of FATE strata & correspondence with observed strata (if perStrata = TRUE, if not = NULL)
list.strata.simulations = list(S = c(1,2,3), M = c(4), B = c(5,6,7))
## List of PFG taken into account in a FATE simulation
list.PFG = as.factor(c("C1", "C2", "C3", "C4", "H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5"))
## Habitat, strata and PFG considered in PFG compo validation
habitat.considered = c("coniferous.forest", "deciduous.forest", "natural.grassland", "woody.heatland")
strata.considered_PFG.compo = c("S", "M", "B")
PFG.considered_PFG.compo = as.factor(c("H1", "H2", "H3", "H4", "H5", "H6", "P1", "P2", "P3", "P4", "P5"))

POST_FATE.validation(name.simulation = "FATE_Champsaur"
                     , sim.version = c("SIMUL_V4.1", "SIMUL_V8.1")
                     , year = 2000
                     , perStrata = TRUE
                     , opt.no_CPU = 1
                     , doHabitat = TRUE
                     , releves.PFG = releves.PFG
                     , hab.obs = hab.obs
                     , studied.habitat = studied.habitat
                     , predict.all.map = TRUE
                     , seed = 123
                     , list.strata.simulations = list.strata.simulations
                     , doComposition = TRUE
                     , PFG.considered_PFG.compo = PFG.considered_PFG.compo
                     , habitat.considered_PFG.compo = habitat.considered
                     , strata.considered_PFG.compo = strata.considered_PFG.compo
                     , doRichness = TRUE
                     , list.PFG = list.PFG
                     , exclude.PFG = NULL)