Transform Braun-Blanquet values into relative abundances

This script is designed to transform Braun-Blanquet abundance information into relative abundances or average recovery values (between 0 and 100).

PRE_FATE.abundBraunBlanquet(abund)

Arguments

abund
a vector with abundance values from Braun-Blanquet (+, r, 1, 2, 3, 4, 5), and with NA when no information

Value

A vector with numerical values between 0 and 100 corresponding to the median of each recovery class.

Details

Braun-Blanquet values allow to estimate the abundance-dominance of a species based on an estimation of the number of individuals and the covering surface. A correspondence has been defined between this index and average recovery values :

Braun-Blanquet		.....

Recovery class (%)		.....		Average recovery (%)

+		<1

0.5			1

1-5		3			2

5-25		15

3		25-50		37.5

4		50-75		62.5

5		75-100

87.5					Braun-Blanquet

Braun-Blanquet J., Roussine N. & Nègre R., 1952. Les groupements végétaux de la France méditerranéenne. Dir. Carte Group. Vég. Afr. Nord , CNRS, 292 p.

Baudière A. & Serve L., 1975. Les groupements végétaux du Plade Gorra-Blanc (massif du Puigmal, Pyrénées Orientales). Essai d'interprétation phytosociologique et phytogéographique. Nat. Monsp., sér. Bot., 25, 5-21.

Foucault B. (de), 1980. Les prairies du bocage virois (Basse-Normandie, France). Typologie phytosociologique et essai de reconstitution des séries évolutives herbagères. Doc. Phytosoc., N.S., 5, 1-109.

Examples

# NOT RUN {
## Load example data
data(MontBlanc)
str(MontBlanc)

## MontBlanc$mat.releves : data.frame
summary(MontBlanc$mat.releves$abund)

## Transformation of Braun-Blanquet abundances
MontBlanc$mat.releves$abund = PRE_FATE.abundBraunBlanquet(abund = MontBlanc$mat.releves$abund)

summary(MontBlanc$mat.releves$abund)

# }