Create DISPERSAL parameter files for a FATE-HD simulation

This script is designed to create parameter files containing dispersal distances for each PFG (one file for each of them) used in the dispersal module of FATE-HD.

PRE_FATE.params_PFGdispersal(name.simulation, mat.PFG.disp)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.disp
a data.frame with 5 columns : PFG, MODE, d50, d99, ldd

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/DISP/ directory with the following parameters :

MODE_DISPERS

DISPERS_DIST

Details

For each PFG, a dispersal module is available to disperse seeds with a kernel.
Dispersal distances are needed to quantify the amount of seeds dispersed into 3 different concentric circles :

d50

the distance at which 50% of seeds are dispersed

d99

the distance at which 49% of seeds are dispersed

ldd

the long dispersal distance at which 1% of seeds are dispersed

Three modes of dispersal are available :

uniform kernel [1]

homogeneous dispersal within the d50, d99 and ldd circles

exponential kernel [2]

seeds are dispersed within each concentric circle according to a decreasing exponential density law (lambda = 1)

exponential kernel
with probability [3]

seeds are dispersed within each concentric circle according to a decreasing exponential density law (lambda = 1) and a continuous decreasing probability with distance

homogeneous dispersal
EVERYWHERE

(!not available YET!)

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , MODE = 1
                                                        , d50 = c(500, 500, 100)
                                                        , d99 = c(10000, 15000, 20000)
                                                        , ldd = c(100000, 50000, 100000)))

# }