Create DISTURBANCE parameter files for a FATE-HD simulation

This script is designed to create parameter files containing response to disturbance parameters for each PFG (one file for each of them) used in the disturbance module of FATE-HD.

PRE_FATE.params_PFGdisturbance(name.simulation, mat.PFG.dist,
  mat.PFG.succ = paste0(name.simulation,
  "/DATA/PFGS/SUCC_COMPLETE_TABLE.csv"), opt.folder.name = NULL)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.dist
a data.frame with 5 columns :
name, responseStage, PFG, KilledIndiv, ResproutIndiv (see Details)

mat.PFG.succ
a data.frame with at least 6 columns :
PFG, type, MATURITY, LONGEVITY, STRATA and CHANG_STR_AGES_to_str_3.
Such an object can be obtained with the PRE_FATE.params_PFGsuccession function.

opt.folder.name
(optional)
a string that corresponds to the name of the folder that will be created into the name.simulation/DATA/PFGS/DIST/ directory to store the results

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/DIST/ directory with the following parameters :

BREAK_AGE : the age when each PFG changes of response stage (in years)

RESPR_AGE : the PFG resprouting age table (in a single row)
This is a vector of no.DIST * no.responseStages numbers (in years) corresponding to the age at which each living PFG can be rejuvenated (younger than the actual one) :

at different response stages (!currently 4, not yet editable!)

for each disturbance.

FATES : the PFG disturbance response table (in a single row)
This is a vector of no.DIST * no.responseStages * 2 numbers
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)
corresponding to the proportion of individuals :

that will be killed or resprout

at different response stages (!currently 4, not yet editable!)

for each disturbance.

PROP_KILLED : the proportion of propagules killed by each disturbance
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

ACTIVATED_SEED : the proportion of seeds activated by each disturbance
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

A DIST_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/PFGS/DIST/opt.folder.name/

Details

The disturbance module of FATE-HD allows the user to simulate spatial perturbation(s) that will impact each PFG in terms of resprouting and mortality on the different response stages (!currently 4, not yet editable!).

Several parameters, given within mat.PFG.succ, are required for each PFG in order to set up these responses :

type

or life-form, based on Raunkier. It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

MATURITY

the age from which the PFG can reproduce

LONGEVITY

the maximum or average lifespan of the PFG

STRATA

the maximum height stratum that the PFG can reach

CHANG_STR_AGES

at what age each PFG goes into the upper stratum

These values will allow to calculate or define a set of characteristics for each PFG :

BREAK_AGE

= each PFG can respond to a disturbance in several different ways
(!currently 4, not yet editable!) that depend on the PFG age
= ages at which each PFG changes of response stage

These response classes are the same for all disturbances for each PFG.
These ages are defined according to the maturity and longevity of the PFG, its life form and its ability to grow above 1.5 meter :

age from class 1 to 2 :

maturity - 2 for herbaceous

1 for chamaephytes and phanerophytes

age from class 2 to 3 :

maturity for herbaceous

min(maturity - 2 , age_above_150cm) for chamaephytes and phanerophytes

age from class 3 to 4 :

longevity - 2 for herbaceous

min(longevity - 2 , age_above_150cm) for chamaephytes and phanerophytes

Some corrections are made for short-living plants (annuals, biennials...) : as they die after 1 or 2 years, they are not affected differently according to life stages. Break ages from class 1 to 3 are set to 1, and break age from 3 to 4 is set to their longevity (1 or 2).

RESPR_AGE

= when subject to a perturbation, each PFG can either stay undisturbed, be killed, or resprout (in years)
= ages at which each PFG will be rejuvenated by a disturbance

It does not impact dead inviduals, only living ones.
It is defined according to the maturity and longevity of the PFG, and its ability to grow above 1.5 meter :

individuals within 1st and 2nd response stage are too young to resprout

individuals within 3rd response stage :
min(maturity - 2, age_above_150cm)

individuals within 4th response stage : longevity - 2

short-living plants (annuals, biennials...) always start back at 0

A second file, mat.PFG.dist, is required to define the importance of the response of each PFG to each disturbance :

name

the name of each perturbation (several can be defined at the same time)

responseStage

the concerned response class (!currently 4, not yet editable!)

PFG

the concerned plant functional group (!should match with the ones given within mat.PFG.succ!)

KilledIndiv

the proportion of killed individuals

ResproutIndiv

the proportion of resprouting individuals

These values will allow to define a third parameter for each PFG :

FATES

= proportion of killed or resprouting individuals
= for each disturbance and for each response stage
Two methods are available to give the proportion of killed and resprouting individuals :

for each life form : H (herbaceous), C (chamaephyte) or P (phanerophyte)

for each PFG

Both methods can be combined (but are applied in the order given by the PFG column).

Two parameters are also defined, but currently set to 0 :

PROP_KILLED

= the proportion of propagules killed by each disturbance

It is currently set to 0 for all PFG and disturbances.

ACTIVATED_SEED

= the proportion of seeds activated by each disturbance

It is currently set to 0 for all PFG and disturbances.

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)))

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4 * 3)
                                                           , responseStage = rep(1:4, 2 * 3)
                                                           , PFG = rep(c("C", "H", "P"), each = 2 * 4)
                                                           , KilledIndiv = c(c(0,10,10,10,1,1,0,0)
                                                                             , c(0,0,0,0,1,1,0,0)
                                                                             , c(10,10,10,10,10,0,0,0))
                                                           , ResproutIndiv = c(c(0,0,0,0,0,0,5,1)
                                                                               , c(0,0,9,10,0,0,5,1)
                                                                               , c(0,0,0,0,0,0,0,0))))

## ----------------------------------------------------------------------------------------- ##

## Load example data
PNE_PARAM = .loadData("PNE_PARAM")

## PNE_PARAM$succ_light : data.frame
## PNE_PARAM$strata_limits : vector
## PNE_PARAM$dist : data.frame

tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity")]

## Create PFG succession parameter files : predefined of strata limits
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = tab
                            , strata.limits = PNE_PARAM$strata_limits
                            , strata.limits_reduce = FALSE)

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = PNE_PARAM$dist)