Create SOIL parameter files for a FATE simulation

This script is designed to create parameter files containing soil contribution and tolerance for each PFG (one file for each of them) used in the soil module of FATE.

PRE_FATE.params_PFGsoil(
  name.simulation,
  mat.PFG.soil,
  mat.PFG.tol = NULL,
  opt.folder.name = NULL
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

mat.PFG.soil
a data.frame with 3 to 7 columns :

PFG,

type, (or active_germ_low, active_germ_medium, active_germ_high) (or strategy_ag)

soil_contrib, soil_tol_min, soil_tol_max (or strategy_contrib)

(see Details)

mat.PFG.tol
(optional)
a data.frame with 2 to 4 columns :

PFG,

lifeStage, resources, tolerance (or strategy_tol)

(see Details)

opt.folder.name
(optional)
a string corresponding to the name of the folder that will be created into the name.simulation/DATA/PFGS/SOIL/ directory to store the results

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SOIL/ directory with the following parameters :

NAME

name of the PFG

ACTIVE_GERM

germination rates depending on soil conditions
(from 0 to 10, corresponding to 0 to 100%)

SOIL_CONTRIB

contribution to the soil value of the pixel

SOIL_LOW

lower value of soil supported by the PFG,
defining the limit between Low and Medium soil resources for this PFG

SOIL_HIGH

upper value of soil supported by the PFG,
defining the limit between Medium and High soil resources for this PFG

SOIL_TOL

soil tolerance table (in a single row).
This is a vector of 9 numbers corresponding to the ability of the PFG to survive or not :

at different life stages (Germinant (Ge), Immature (Im), Mature (Ma))

under different soil conditions (Low (L), Medium (M) or High (H)).

These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH
(from 0 to 10, corresponding to 0 to 100%).

A SOIL_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/PFGS/SOIL/opt.folder.name/.

Details

The soil module allows the user to add the effect of soil interaction within a primary vegetation succession.

Several parameters, given within mat.PFG.soil or mat.PFG.tol, are required for each PFG in order to set up the soil interaction :

PFG

the concerned plant functional group

type

or life-form, based on Raunkier.
It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

(active_germ_low)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for Low soil condition

(active_germ_medium)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for Medium soil condition

(active_germ_high)

an integer between 0 and 10 corresponding to the proportion of seeds that will germinate for High soil condition

(strategy_ag)

a string to choose the germination strategy :
poor_lover, indifferent, rich_lover

soil_contrib

a value corresponding to the PFG preference for soil fertility
(e.g. nitrogen value from Ellenberg or Flora Indicativa)

soil_tol_min

the minimum soil value tolerated by the PFG (on the same scale than soil_contrib)

soil_tol_max

the maximum soil value tolerated by the PFG (on the same scale than soil_contrib)

(strategy_contrib)

a string to choose the contribution strategy :
ubiquist, tobedefined

lifeStage

the concerned life stage (Germinant, Immature, Mature)

resources

the concerned soil condition (Low, Medium, High)

tolerance

an integer between 0 and 10 corresponding to the proportion of surviving individuals

(strategy_tol)

a string to choose the tolerance strategy :
ubiquist, tobedefined

These values will allow to calculate or define a set of characteristics for each PFG :

ACTIVE_GERM

proportion of seeds that will germinate for each soil condition (Low, Medium, High)

Three methods to define these proportions are available :

from predefined scenarios (using strategy_ag) :

| _L_ _M_ _H_ |

_______________

poor_lover

| 80% 90% 50% |

indifferent

| 90% 90% 90% |

rich_lover

| 50% 90% 80% |

from predefined rules (using type) :

for H (herbaceous) : 80%, 100%, 50%

for C (chamaephyte) or P (phanerophyte) : 90%, 100%, 90%

from user data :
with the values contained within the active_germ_low, active_germ_medium and active_germ_high columns, if provided

SOIL_CONTRIB
SOIL_LOW
SOIL_HIGH

Two methods to define these values are available :

from predefined scenarios (using strategy_contrib) :

the values give the soil_tol_min, soil_contrib and soil_tol_max

with L: low soil, M: medium soil, H: high soil

| ___ L ___ | ___ M ___ | ___ H ___ |

_____________________________________

ubiquist

__________ 1.5 _ 2.5 _ 4.5 __________

to be filled

from user data :
with the values contained within the soil_contrib, soil_tol_min and soil_tol_max columns, if provided

SOIL_TOL

defined for each life stage (Germinant, Immature, Mature)
and each soil condition (Low, Medium, High)

Three methods to define these tolerances are available :

from predefined scenarios (using strategy_tol) :

the values give the percentage of surviving individuals to the concerned conditions

with g: Germinant, i: Immature, m: Mature

with L: low soil, M: medium soil, H: high soil

| ___ g ___ | ___ i ___ | ___ m ___ |

| _L _M_ H_ | _L _M_ H_ | _L _M_ H_ |

_____________________________________

ubiquist

| 90 100 90 | 90 100 90 | 90 100 90 |

to be filled

from predefined rules :

(A)

germinants are severely impacted by wrong soil conditions

(B)

immatures are half impacted by wrong soil conditions

(C)

matures are little impacted by wrong soil conditions

the values give the percentage of surviving individuals to the concerned conditions

with g: Germinant, i: Immature, m: Mature

with L: low soil, M: medium soil, H: high soil

| ___ g ___ | ___ i ___ | ___ m ___ |

| _L _M_ H_ | _L _M_ H_ | _L _M_ H_ |

_____________________________________

| 10 100 30 | 40 100 60 | 70 100 90 |

from user data :
with the values contained within the lifeStage, resources and tolerance columns, if provided

See also

Author

Maya Guéguen

Examples

                                                    
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG soil parameter files (with strategies) -----------------------------------------
PRE_FATE.params_PFGsoil(name.simulation = 'FATE_simulation'
                        , mat.PFG.soil = data.frame(PFG = paste0('PFG', 1:6)
                                                    , strategy_ag = c('rich_lover', 'indifferent'
                                                                      , 'indifferent', 'rich_lover'
                                                                      , 'indifferent', 'poor_lover')
                                                    , strategy_contrib = rep('ubiquist', 6))
                        , mat.PFG.tol = data.frame(PFG = paste0('PFG', 1:6)
                                                   , strategy_tol = rep('ubiquist', 6)))
                                                        

## Create PFG soil parameter files (with all values) -----------------------------------------
mat.tol = expand.grid(resources = c('Low', 'Medium', 'High')
                      , lifeStage = c('Germinant', 'Immature', 'Mature')
                      , PFG = paste0('PFG', 1:6))
mat.tol$tolerance = c(8, 8, 4, 8, 5, 4, 9, 4, 4
                      , rep(9, 9)
                      , rep(9, 9)
                      , 8, 8, 6, 8, 6, 6, 9, 5, 5
                      , 8, 8, 8, 5, 6, 9, 3, 4, 9
                      , 8, 8, 8, 5, 5, 9, 5, 5, 9)

PRE_FATE.params_PFGsoil(name.simulation = 'FATE_simulation'
                        , mat.PFG.soil = data.frame(PFG = paste0('PFG', 1:6)
                                                    , active_germ_low = c(5, 8, 8, 6, 8, 8)
                                                    , active_germ_medium = rep(9, 6)
                                                    , active_germ_high = c(9, 8, 8, 9, 8, 4)
                                                    , strategy_contrib = rep('ubiquist', 6))
                        , mat.PFG.tol = mat.tol)
                                                        
                                                        
## -------------------------------------------------------------------------------------------

## Load example data
data(DATASET_Bauges_parameters)

## PFG traits for soil
tab.traits = DATASET_Bauges_parameters$tab.soil
str(tab.traits)

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = 'FATE_Bauges')

## Create PFG soil parameter files -----------------------------------------------------------
PRE_FATE.params_PFGsoil(name.simulation = 'FATE_Bauges'
                        , mat.PFG.soil = tab.traits)