Create SOIL parameter files for a FATE-HD simulation

This script is designed to create parameter files containing soil contribution and tolerance for each PFG (one file for each of them) used in the soil module of FATE-HD.

PRE_FATE.params_PFGsoil(name.simulation, mat.PFG.soil)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.soil
a data.frame with 5 columns : PFG, type, soil_contrib, soil_tol_min, soil_tol_max

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SOIL/ directory with the following parameters :

NAME : name of the PFG

SOIL_CONTRIB : the contribution (influence) of the PFG on the nitrogen soil value of the pixel

SOIL_LOW : the lower value of nitrogen supported by the PFG

SOIL_HIGH : the upper value of nitrogen supported by the PFG

ACTIVE_GERM : the germination rates depending on soil conditions
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

SOIL_TOL : the PFG soil tolerance table (in a single row).
This is a vector of 9 numbers corresponding to the ability of the PFG to survive or not :

at different life stages (Germinant (Ge), Immature (Im), Mature (Ma))

under different soil conditions (Low (L), Medium (M) or High (H)).

These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH.
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

A SOIL_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.

Details

The soil module allows the user to simulate a primary vegetation succession based on soil competition.
Several parameters are required for each PFG in order to set up the soil competition :

type

or life-form, based on Raunkier. It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

soil_contrib

a value between 0 and X corresponding to the nitrogen value of the PFG (e.g. from Ellenberg)

soil_tol_min

the minimum nitrogen value tolerated by the PFG

soil_tol_max

the maximum nitrogen value tolerated by the PFG

These values will allow to calculate or define a set of characteristics for each PFG :

SOIL_CONTRIB

nitrogen contribution to the soil value of the PFG

SOIL_LOW

minimum nitrogen value tolerated by the PFG

SOIL_HIGH

maximum nitrogen value tolerated by the PFG

ACTIVE_GERM

proportion of seeds that will germinate for each soil condition (Low, Medium, High)

SOIL_TOL

defined for each life stage (Germinant, Immature, Mature) and each soil condition (Low, Medium, High) :

(A) germinants are severely impacted by wrong soil conditions :

only 10% of the cohorts survive in Low soil condition

all cohorts die in High soil condition

(B) immatures are half impacted by wrong soil conditions :

50% of the cohorts survive in Low soil condition

40% of the cohorts survive in High soil condition

(C) matures are little affected by wrong soil conditions :

90% of the cohorts survive in Low soil condition

80% of the cohorts survive in High soil condition

See also

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                    , type = c("H", "H", "C")
                                                    , soil_contrib = c(2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6)))

# }