Create SUCCESSION parameter files for a FATE-HD simulation
Source: R/PRE_FATE.params_PFGsuccession.R

This script is designed to create parameter files containing succession parameters for each PFG (one file for each of them) used in the core module of FATE-HD.

PRE_FATE.params_PFGsuccession(name.simulation, mat.PFG.disp)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.disp
a data.frame with 4 columns : PFG, d50, d99, ldd

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SUCC/ directory with the following parameters :

MATURITY : the maturity age of the PFG (in years)

LONGEVITY : the PFG life span (in years)

MAX_ABUNDANCE : the maximal (qualitative) shade that the PFG is able to produce (1: Low 2: Medium 3: High)

IMM_SIZE : the relative size of the immature PFG (0: 0

CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum (in years, put a value higher than the PFG life span if it is not supposed to rise a stratum)

WIDE_DISPERS : is the PFG able to disperse everywhere (i.e. no dispersal limits) (0: No 1: Yes)

ACTIVE_GERM : the germination rates depending on light conditions (0: 0

SHADE_TOL : the PFG shade tolerance table (in a single row). This is a vector of 9 numbers (0: Die 1: Survive) corresponding to the ability of the PFG to survive or not at different life stages (Germinate (Ge), Immature (Im), Mature (Ma)) under different light conditions (Low (L), Medium (M) or High (H) light resources). These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH.

SEED_POOL_LIFE : the maximal number of years seeds are able to survive

SEED_DORMANCY

Details

For each PFG,

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.disp = data.frame(PFG = c("PFG1", "PFG2", "PFG3")
                                                        , d50 = c(500, 500, 100)
                                                        , d99 = c(10000, 15000, 20000)
                                                        , ldd = c(100000, 50000, 100000)))

# }