Create SUCCESSION parameter files for a FATE-HD simulation

This script is designed to create parameter files containing succession parameters for each PFG (one file for each of them) used in the core module of FATE-HD.

PRE_FATE.params_PFGsuccession(name.simulation, mat.PFG.succ,
  strata.limits = c(0, 20, 50, 150, 400, 1000, 2000, 5000, 10000),
  strata.limits_reduce = TRUE, opt.folder.name = NULL)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.succ
a data.frame with 5 columns : PFG, type, height, maturity, longevity

strata.limits
a vector of integer containing values among which height strata limits will be chosen

strata.limits_reduce
default TRUE. If TRUE, stratum height limits are checked to try and bring several PFGs together in a same stratum

opt.folder.name
(optional)
a string that corresponds to the name of the folder that will be created into the name.simulation/DATA/PFGS/SUCC/ directory to store the results

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SUCC/ directory with the following parameters :

NAME

name of the PFG

MATURITY

the maturity age of the PFG (in years)

LONGEVITY

the PFG life span (in years)

MAX_ABUNDANCE

the maximal (qualitative) abundance / space that the PFG is able to produce / occupy
(qualitative) (1: Low 2: Medium 3: High)

IMM_SIZE

the relative size of the immature PFG
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

CHANG_STR_AGES

the ages at which the PFG goes in the upper stratum
(in years, put a value higher than the PFG life span if it is not supposed to rise a stratum)

SEED_POOL_LIFE

the maximal number of years seeds are able to survive (for active and dormant pool)

SEED_DORMANCY

are the seeds dormant or not (0: No 1: Yes)

POTENTIAL_
FECUNDITY

the maximum number of seeds produced by the PFG
(currently set by default to 100)

A SUCC_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory.
This file can be used to parameterize the disturbance files.

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/PFGS/SUCC/opt.folder.name/

Details

The core module of FATE-HD allows the user to simulate a primary vegetation succession based on life history traits.
Several parameters are required for each PFG in order to set up this life cycle :

type

or life-form, based on Raunkier. It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

height

the maximum or average height that reach the PFG

maturity

the age from which the PFG can reproduce

longevity

the maximum or average lifespan of the PFG

These values will allow to calculate or define a set of characteristics for each PFG :

STRATA_LIMITS

= the height values that define each stratum.

The steps are detailed below and try to homogenize the number of PFG within each stratum :

the average number of PFG per stratum should be close to the square root of the total number of PFG (no.PFG.perStrata)

strata limits should go exponentially and will be selected among
0, 20, 50, 150, 400, 1000, 2000, 5000, 10000

PFG are divided according to their height and these strata limits and then grouped in order to have per stratum a number of PFG >= (no.PFG.perStrata - 2)

STRATA

= the maximum stratum that each PFG can reach

MAX_ABUNDANCE

= maximum abundance of mature PFG in favorable conditions

It can be seen as a proxy of maximum carrying capacity for mature individuals (and therefore as a proxy a quantity of shade produced as well, if the light module is activated), and it is defined according to the number of strata potentially occupied by a PFG and its life form :

herbaceous take little place (and make little shade) (1),
chamaephytes take medium place (and make medium shade) (2)
and phanerophytes take a lot of place (and make lot of shade) (3)

all plants in first stratum take little place (and make little shade) (1)

plants other than herbaceous in stratum 2 take medium place (and make medium shade) (2)

herbaceous in stratum > 2 take medium place (and make medium shade) (2)

chamaephytes in stratum > 3 take lot of place (and make lot of shade) (3)

IMM_SIZE

= relative size of immature versus mature plants

immature herbaceous take as much space as mature herbaceous
(and contribute to shade in the same way) (100 %)

immature chamaephytes take two times less space than mature chamaephytes
(and contribute to shade half less) (50 %)

immature phanerophytes take only 10 % of their full space (and shade) capacity

intermediate percentages for herbaceous in stratum 2 (80 %) and in stratum > 2 (50 %)

immature chamaephytes in 1st stratum take as much space as mature chamaephytes
(and contribute to shade in the same way) (100 %)

immature phanerophytes with height < 10m take two times less space than mature phanerophytes
(and contribute to shade half less) (50 %)

CHANG_STR_AGES

= at what age each PFG goes into the upper stratum.

It is defined using a logistic growth curve with 2 points to parameterize it :

at age = maturity/2, height = IMM_SIZE * height

at age = longevity, height = height

POTENTIAL_FECUNDITY

= maximum number of seeds produced by the PFG.

By default, defined to the same value for all PFG (100), but can be changed by hand directly into the produced files.

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG", 1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)))

## ----------------------------------------------------------------------------------------- ##

## Load example data
PNE_PARAM = .loadData("PNE_PARAM")

## PNE_PARAM$succ_light : data.frame
## PNE_PARAM$strata_limits : vector

tab = PNE_PARAM$succ_light[, c("PFG", "type", "height", "maturity", "longevity")]

## Create PFG succession parameter files : automatic definition of strata limits
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = tab)

## Create PFG succession parameter files : predefined of strata limits
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = tab
                            , strata.limits = PNE_PARAM$strata_limits
                            , strata.limits_reduce = FALSE)