Create SUCCESSION parameter files for a FATE-HD simulation
Source: R/PRE_FATE.params_PFGsuccession.R

This script is designed to create parameter files containing succession parameters for each PFG (one file for each of them) used in the core module of FATE-HD.

PRE_FATE.params_PFGsuccession(name.simulation, mat.PFG.succ)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

mat.PFG.succ
a data.frame with 7 columns : PFG, type, height, maturity, longevity, dispersal, light

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SUCC/ directory with the following parameters :

NAME : name of the PFG

MATURITY : the maturity age of the PFG (in years)

LONGEVITY : the PFG life span (in years)

MAX_ABUNDANCE : the maximal (qualitative) shade that the PFG is able to produce
(1: Low 2: Medium 3: High)

IMM_SIZE : the relative size of the immature PFG
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum
(in years, put a value higher than the PFG life span if it is not supposed to rise a stratum)

ACTIVE_GERM : the germination rates depending on light conditions
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

SHADE_TOL : the PFG shade tolerance table (in a single row).
This is a vector of 9 numbers (0: Die 1: Survive) corresponding to the ability of the PFG to survive or not :

at different life stages (Germinant (Ge), Immature (Im), Mature (Ma))

under different light conditions (Low (L), Medium (M) or High (H)).

These parameters should be given in this order : GeL, GeM, GeH, ImL, ImM, ImH, MaL, MaM, MaH.

SEED_POOL_LIFE : the maximal number of years seeds are able to survive (for active and dormant pool)

SEED_DORMANCY : are the seeds dormant or not (0: No 1: Yes)

A SUCC_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory. This file can be used to parameterize the disturbance files.

Details

The core module of FATE-HD allows the user to simulate a primary vegetation succession based on light competition. Several parameters are required for each PFG in order to set up this life cycle :

type

or life-form, based on Raunkier. It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

height

the maximum or average height that reach the PFG

maturity

the age from which the PFG can reproduce

longevity

the maximum or average lifespan of the PFG

dispersal

the method of dispersal of the PFG

0 = no dispersal

1 = homogeneous dispersal within the d50, d99 and ldd circles

2 = negative exponential kernel within the d50, d99 and ldd circles

3 = negative exponential kernel + probability decreasing with distance within the d50, d99 and ldd circles

4 = homogeneous dispersal EVERYWHERE (!not available YET!)

light

a value between 0 and 10 corresponding to the Ellenberg value of the PFG

These values will allow to calculate or define a set of characteristics for each PFG :

STRATA_LIMITS

= the height values that define each stratum.

The steps are detailed below and try to homogenize the number of PFG within each stratum :

the average number of PFG per stratum should be close to the square root of the total number of PFG (no.PFG.perStrata)

strata limits should go exponentially and will be selected among
0, 20, 50, 150, 400, 1000, 2000, 5000, 10000

PFG are divided according to their height and these strata limits and then grouped in order to have per stratum a number of PFG >= (no.PFG.perStrata - 2)

STRATA

the maximum stratum that each PFG can reach

MAX_ABUNDANCE

= maximum abundance of mature PFG in favorable conditions
= maximum shade a PFG can make in a pixel corresponding to a number of individuals

It is defined according to the number of strata potentially occupied by a PFG :

herbaceous make little shade (1), chamaephytes make medium shade (2) and phanerophytes make lot of shade (3)

all plants in first stratum make little shade (1)

plants other than herbaceous in stratum 2 make medium shade (2)

herbaceous in stratum > 2 make medium shade (2)

chamaephytes in stratum > 3 make lot of shade (3)

IMM_SIZE

= relative shade of immature plants

immature herbaceous contribute to shade in the same way than mature herbaceous (100 %)

immature chamaephytes contribute to shade half less than mature herbaceous (50 %)

immature phanerophytes contribute to shade only by 10 % of their full capacity

intermediate percentages for herbaceous in stratum 2 (80 %) and in stratum > 2 (50 %)

immature chamaephytes in 1st stratum contribute to shade in the same way than mature chamaephytes (100 %)

immature phanerophytes with height < 10m contribute to shade half less than mature phanerophytes (50 %)

CHANG_STR_AGES

= at what age each PFG goes into the upper stratum.

It is defined using a logistic growth curve with 2 points to parameterize it :

at age = maturity/2, height = IMM_SIZE * height

at age = longevity, height = height

SHADE_TOL

defined for each life stage (Germinant, Immature, Mature) and each light condition (Low, Medium, High) :

PFG are tolerant to low light if light <= 3

PFG are tolerant to medium light if light <= 4

PFG are tolerant to high light if light >= 3

all germinants are assumed to be tolerant to Low and Medium light

all mature trees are assumed to be tolerant to High light in the upper strata

all immature trees that grow in the penultimate stratum are assumed to be tolerant to High light

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                            , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                        , type = c("C", "C", "H", "H", "P", "P")
                                                        , height = c(10, 250, 36, 68, 1250, 550)
                                                        , maturity = c(5, 5, 3, 3, 8, 9)
                                                        , longevity = c(12, 200, 25, 4, 110, 70)
                                                        , dispersal = 1
                                                        , light = c(4, 6, 3, 6, 5, 5)))

# }