Create SCENARIO parameter files for a FATE-HD simulation

This script is designed to create a parameter file containing simulation years at which the FATE-HD software must save rasters of PFG abundances or simulation objects.

PRE_FATE.params_changingYears(name.simulation, type.changing, mat.changing,
  opt.folder.name = NULL)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

type.changing
a string to choose the concerned module :
succession (MASK), habitat suitability (HS) or disturbances (DIST)

mat.changing
a data.frame with 3 columns : year, order, file.name

opt.folder.name
(optional)
a string that corresponds to the name of the folder that will be created into the name.simulation/DATA/SCENARIO/ directory to store the results

Value

Several .txt files into the name.simulation/DATA/SCENARIO/ :

type.changing_changing_times.txt : one line for each simulation year

type.changing_changing_masks_t....txt : one line for each new raster file

If the opt.folder.name has been used, the files will be into the folder name.simulation/DATA/SCENARIO/opt.folder.name/

Details

Several modules of the FATE-HD software allow the user to simulate changes over time :

succession

the CORE module is based on a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells in which the PFG can try to colonize. The available pixels can change through time, to simulate habitat loss (e.g. urbanization) or gain (e.g. glacial retreat).

habitat suitability

if this MODULE is activated within the Global_parameters file, PFG colonization depends on maps given for each PFG within the Simul_parameters file with the PFG_HAB_MASK flag.
These maps must contain values between 0 and 1 corresponding to the probability of presence of the group in each pixel. These probabilities can change through time, as they often come from Species Distribution Models (SDM) that can be based for example on climatic variables.

disturbances

if this MODULE is activated within the Global_parameters file, each disturbance relies on a raster given within the Simul_parameters file with the DIST_MASK flag.
As for succession, this mask is filled with either 0 or 1 to define where the perturbation occurs. The impacted pixels can also change through time (e.g. change in forestry practices, expansion of grazing areas, etc).

These changes require two types of information, that must be presented within two different files :

Years of change
column year

all simulation years at which the raster files of a specific module (succession, habitat suitability, disturbance) will be changed. To each given year must correspond a file with the names of the new maps.

New maps / mask
column file.name

for a specific year of change, the names of the new raster files, always in the same order.
It can be either .img or .tif.

Order of the filenames
column order

integer to be sure to always give the raster maps in the same order.

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Changing_times parameter file
PRE_FATE.params_changingYears(name.simulation = "FATE_simulation"
                              , type.changing = "DIST"
                              , mat.changing = data.frame(year = c(50,50,80,80)
                                                          , order = c(1,2,1,2)
                                                          , file.name = c("MASK_DIST1_50.tif"
                                                                          , "MASK_DIST2_50.tif"
                                                                          , "MASK_DIST1_80.tif"
                                                                          , "MASK_DIST2_80.tif")))