Create Global_parameters parameter file for a FATE-HD simulation
Source: R/PRE_FATE.params_globalParameters.R

This script is designed to create one parameter file containing GLOBAL PARAMETERS used in FATE-HD model.

PRE_FATE.params_globalParameters(name.simulation, opt.no_CPU = 1,
  required.no_PFG, required.no_STRATA, required.simul_duration,
  required.seeding_duration, required.seeding_timestep, doDispersal = FALSE,
  doHabSuitability = FALSE, HABSUIT.ref_option, doDisturbances = FALSE,
  DIST.no, DIST.no_sub, DIST.freq, doSoil = FALSE, doFire = FALSE,
  doDrought = FALSE, doHabStability = FALSE, doAliens = FALSE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the FATE-HD simulation

required.no_PFG
an integer corresponding to the number of PFG

required.no_STRATA
an integer corresponding to the number of height strata

required.simul_duration
an integer corresponding to the duration of simulation (in years)

required.seeding_duration
an integer corresponding to the duration of seeding (in years)

required.seeding_timestep
an integer corresponding to the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

doDispersal
default FALSE. If TRUE, seed dispersal is activated in the FATE-HD simulation, and associated parameters are required

doHabSuitability
default FALSE. If TRUE, habitat suitability is activated in the FATE-HD simulation, and associated parameters are required

HABSUIT.ref_option
an integer corresponding to the way of simulating the habitat suitability variation between years for each PFG, either random (1) or PFG specific (2)

doDisturbances
default FALSE. If TRUE, disturbances are applied in the FATE-HD simulation, and associated parameters are required

DIST.no
the number of disturbances

DIST.no_sub
the number of way a PFG could react to a disturbance

DIST.freq
the frequency of each disturbance (in years)

doSoil
to be filled

doFire
to be filled

doDrought
to be filled

doHabStability
to be filled

doAliens
to be filled

Value

A .txt file into the name.simulation/DATA/GLOBAL_PARAMETERS directory with the following parameters :

NB_CPUS

NB_FG

NB_STRATUM

SIMULATION_DURATION

SEEDING_DURATION

SEEDING_TIMESTEP

If the simulation includes dispersal :

DO_DISPERSAL

If the simulation includes habitat suitability :

DO_HAB_SUITABILITY

HABSUIT_OPTION

If the simulation includes disturbances :

DO_DISTURBANCES

NB_DISTURBANCES

NB_SUBDISTURBANCES

FREQ_DISTURBANCES

If the simulation includes soil competition :

DO_SOIL_COMPETITION

SOIL_DEFAULT_VALUE

NB_SOIL_CATEGORIES

SOIL_CATEGORIES_TRESHOLDS

If the simulation includes drought disturbances :

DO_DROUGHT_DISTURBANCES

NB_SUBDROUGHT

CHRONO_POST_DROUGHT

CHRONO_CURR_DROUGHT

If the simulation includes habitat stability check :

DO_HAB_STABILITY

NB_HABITATS

If the simulation includes aliens introduction :

DO_ALIENS_DISTURBANCE

FREQ_ALIENS

Details

The core module of FATE-HD requires several parameters to define general characteristics of the simulation :

Studied system

required.no_PFG : the number of Plant Functional Groups that will be included into the simulation. This number should match with the number of files that will be given with the characteristics of each group (SUCC, DISP, ...).

required.no_STRATA : the number of height strata that will be used into the succession module. This number should match with the maximum number of strata possible defined into the PFG SUCC files.

Simulation timing

required.simul_duration : the duration of simulation (in years)

required.seeding_duration : the duration of seeding (in years)

required.seeding_timestep : the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

The other modules of FATE-HD can be activated within this file, and if so, some additional parameters will be required :

DISPERSAL

HABITAT SUITABILITY

= to influence plants fecundity and seed recruitment according to PFG preferences for habitat conditions
= filter based on maps given for each PFG within the Simul_parameters file with the PFG_HAB_MASK flag.

These maps must contain values between 0 and 1 corresponding to the probability of presence of the group in each pixel. Each year (timestep), this value will be compared to a reference value, and if superior, the PFG will be able to grow and survive.

HABSUIT.ref_option : the habitat suitability reference value can be set in two possible ways :

random : for each pixel, the reference value is drawn from a uniform distribution, and the same value is used for each PFG within this pixel.

PFG specific : for each PFG, a mean value and a standard deviation value are drawn from a uniform distribution. For each pixel and for each PFG, the reference value is drawn from a normal distribution of parameters the mean and standard deviation of the PFG.

DISTURBANCES

= defined for events such as mowing, grazing, but also urbanization, crops, etc.

The impact zone is defined with a mask (map with 0 or 1) and the user will have to define how each PFG will be impacted depending on age and life stage.

DIST.no : the number of different disturbances

DIST.no_sub : the number of way a PFG could react to a perturbation

DIST.freq : the frequency of each disturbance (in years)

DROUGHT

= to experience extreme events with a direct and a delayed response on PFG.

The PFG's past drought exposure is compared to current moisture values as a proxy of drought index. Developed canopy closure helps to reduce these values. Each PFG is impacted if submitted one year to drought, and also (and more consequently) if submitted several years in a row.

DROUGHT.no_sub : the number of way a PFG could react to a perturbation

DROUGHT.chrono_post : if post-drought effects will occur previously to succession (0) or after (1)

DROUGHT.chrono_curr : if current-drought effects will occur previously to succession (0) or after (1)

HABITAT STABILITY

= to evaluate through the evolution of the PFG composition in different habitats.

The habitat distribution is given with a categorical map with each value corresponding to a specific habitat. Every year, the abundance of each PFG within each habitat is recorded, as well as the evenness of the habitat : $$ \text{evenness} = - SUM(\text{PFG.proportion} * log(\text{PFG.proportion})) / log(\text{no.PFG} - 1) $$ $$\text{with PFG.proportion} = \text{abund.PFG.in.Habitat} / \text{abund.allPFG.in.Habitat}$$ Every 5 years, a stability check is done to see if the abundance or the evenness have values out of the usual distribution.

HABSTAB.no_hab : the number of different habitats

INVASIVE
INTRODUCTION

= to add new PFG during the simulation.

The introduction areas are defined with a mask (map with 0 or 1). If the habitat suitability filter is on, suitability maps will also be needed for these new groups.

ALIEN.freq : the frequency of each introduction (in years)

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , doDispersal = 1
                                 , doHabSuitability = 1
                                 , HABSUIT.ref_option = 1
                                 )

## Create SEVERAL Global_parameters files
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , doDispersal = 1
                                 , doHabSuitability = 1
                                 , HABSUIT.ref_option = c(1,2)
                                 )

# }