Create Global_parameters parameter file for a FATE simulation

This script is designed to create parameter file(s) containing GLOBAL PARAMETERS used in FATE model.

PRE_FATE.params_globalParameters(
  name.simulation,
  opt.no_CPU = 1,
  opt.replacePrevious = FALSE,
  required.no_PFG,
  required.no_strata,
  required.simul_duration,
  required.seeding_duration,
  required.seeding_timestep,
  required.seeding_input,
  required.max_abund_low,
  required.max_abund_medium,
  required.max_abund_high,
  doLight = FALSE,
  LIGHT.thresh_medium,
  LIGHT.thresh_low,
  doSoil = FALSE,
  SOIL.init,
  SOIL.retention,
  doDispersal = FALSE,
  DISPERSAL.mode,
  doHabSuitability = FALSE,
  HABSUIT.ref_option,
  doDisturbances = FALSE,
  DIST.no,
  DIST.no_sub,
  DIST.freq,
  doDrought = FALSE,
  DROUGHT.no_sub,
  doAliens = FALSE,
  ALIEN.no,
  ALIEN.freq,
  doFire = FALSE
)

Arguments

name.simulation
a string corresponding to the main directory or simulation name of the FATE simulation

opt.no_CPU
(optional) default 1.
an integer corresponding to the number of resources that can be used to parallelize the FATE simulation

opt.replacePrevious
(optional) default FALSE.
If TRUE, pre-existing files inside name.simulation/DATA/GLOBAL_PARAMETERS folder will be replaced

required.no_PFG
an integer corresponding to the number of PFG

required.no_strata
an integer corresponding to the number of height strata

required.simul_duration
an integer corresponding to the duration of simulation (in years)

required.seeding_duration
an integer corresponding to the duration of seeding (in years)

required.seeding_timestep
an integer corresponding to the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

required.seeding_input
an integer corresponding to the number of seeds attributed to each PFG at each time step, and until the seeding duration is not over

required.max_abund_low
an integer in the order of 1 000 to rescale abundance values of small PFG

required.max_abund_medium
an integer in the order of 1 000 to rescale abundance values of intermediate PFG

required.max_abund_high
an integer in the order of 1 000 to rescale abundance values of tall PFG

doLight
default FALSE.
If TRUE, light competition is activated in the FATE simulation, and associated parameters are required

LIGHT.thresh_medium
(optional)
an integer in the order of 1 000 to convert PFG abundances in each stratum into light resources. It corresponds to the limit of abundances above which light resources are medium. PFG abundances lower than this threshold imply high amount of light. It is consequently lower than LIGHT.thresh_low.

LIGHT.thresh_low
(optional)
an integer in the order of 1 000 to convert PFG abundances in each strata into light resources. It corresponds to the limit of abundances above which light resources are low. PFG abundances higher than LIGHT.thresh_medium and lower than this threshold imply medium amount of light.

doSoil
default FALSE.
If TRUE, soil competition is activated in the FATE simulation, and associated parameters are required

SOIL.init
(optional)
a double corresponding to the soil value to initialize all pixels when starting the FATE simulation

SOIL.retention
(optional)
a double corresponding to the percentage of soil value of the previous simulation year that will be kept in the calculation of the soil value of the current simulation year

doDispersal
default FALSE.
If TRUE, seed dispersal is activated in the FATE simulation, and associated parameters are required

DISPERSAL.mode
(optional)
an integer corresponding to the way of simulating the seed dispersal for each PFG, either packets kernel (1), exponential kernel (2) or exponential kernel with probability (3)

doHabSuitability
default FALSE.
If TRUE, habitat suitability is activated in the FATE simulation, and associated parameters are required

HABSUIT.ref_option
(optional)
an integer corresponding to the way of simulating the habitat suitability variation between years for each PFG, either random (1) or PFG specific (2)

doDisturbances
default FALSE.
If TRUE, disturbances are applied in the FATE simulation, and associated parameters are required

DIST.no
(optional)
an integer corresponding to the number of disturbances

DIST.no_sub
(optional)
an integer corresponding to the number of way a PFG could react to a disturbance

DIST.freq
(optional)
a vector of integer corresponding to the frequency of each disturbance (in years)

doDrought
default FALSE.
If TRUE, drought disturbances are applied in the FATE simulation, and associated parameters are required

DROUGHT.no_sub
(optional)
an integer corresponding to the number of way a PFG could react to a drought disturbance

doAliens
default FALSE.
If TRUE, invasive plant introduction is activated in the FATE simulation, and associated parameters are required

ALIEN.no
(optional)
an integer corresponding to the number of introductions

ALIEN.freq
(optional)
a vector of integer corresponding to the frequency of each introduction (in years)

doFire
to be filled

Value

A .txt file into the name.simulation/DATA/GLOBAL_PARAMETERS directory with the following parameters :

NO_CPU

NO_PFG

NO_STRATA

SIMULATION_DURATION

SEEDING_DURATION

SEEDING_TIMESTEP

SEEDING_INPUT

MAX_ABUND_LOW

MAX_ABUND_MEDIUM

MAX_ABUND_HIGH

If the simulation includes light competition :

DO_LIGHT_COMPETITION

LIGHT_THRESH_MEDIUM

LIGHT_THRESH_LOW

If the simulation includes soil competition :

DO_SOIL_COMPETITION

SOIL_INIT

SOIL_RETENTION

If the simulation includes dispersal :

DO_DISPERSAL

DISPERSAL_MODE

If the simulation includes habitat suitability :

DO_HAB_SUITABILITY

HABSUIT_OPTION

If the simulation includes disturbances :

DO_DISTURBANCES

DIST_NO

DIST_NOSUB

DIST_FREQ

If the simulation includes drought disturbance :

DO_DROUGHT_DISTURBANCE

DROUGHT_NOSUB

If the simulation includes aliens introduction :

DO_ALIENS_INTRODUCTION

ALIENS_NO

ALIENS_FREQ

Details

The core module of FATE requires several parameters to define general characteristics of the simulation :

Studied system

no_PFG

the number of plant functional groups that will be included into the simulation.
This number should match with the number of files that will be given to parameterize the different activated modules with the characteristics of each group (SUCC, DISP, ...).

no_STRATA

the number of height strata that will be used into the succession module.
This number should match with the maximum number of strata possible defined into the PFG SUCC files.

Simulation timing

simul_duration

the duration of simulation (in years)

seeding_duration

the duration of seeding (in years)

seeding_timestep

the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

seeding_input

the number of seeds dispersed for each PFG at each time step, and until the seeding duration is not over

The other modules of FATE can be activated within this file, and if so, some additional parameters will be required :

LIGHT

= to influence seed recruitment and plant mortality according to PFG preferences for light conditions
(see PRE_FATE.params_PFGlight)
= light resources are calculated as a proxy of PFG abundances within each height stratum

To transform PFG abundances into light resources : $$abund_{\text{ PFG}_{all}\text{, }\text{Stratum}_k} < \text{LIGHT.thresh_medium} \;\; \Leftrightarrow \;\; light_{\text{ Stratum}_k} = \text{High}$$ $$\text{LIGHT.thresh_medium } < abund_{\text{ PFG}_{all}\text{, }\text{Stratum}_k} < \text{LIGHT.thresh_low} \\ \Leftrightarrow \;\; light_{\text{ Stratum}_k} = \text{Medium}$$ $$abund_{\text{ PFG}_{all}\text{, }\text{Stratum}_k} > \text{LIGHT.thresh_low} \;\; \Leftrightarrow \;\; light_{\text{ Stratum}_k} = \text{Low}$$ As light resources are directly obtained from PFG abundances, LIGHT.thresh_medium and LIGHT.thresh_low parameters should be on the same scale than required.max_abund_low, required.max_abund_medium and required.max_abund_high parameters from the core module.

SOIL

= to influence seed recruitment and plant mortality according to PFG preferences for soil conditions
(see PRE_FATE.params_PFGsoil)
= soil composition is calculated as the weighted mean of each PFG's contribution with a possible retention of the soil value of the previous simulation year $$Soil_y + \text{SOIL.retention} * (Soil_{y-1} - Soil_y)$$ with $$Soil_y = \sum abund_{\text{ PFG}_i\text{, }y} * \text{contrib}_{\text{ PFG}_i}$$

DISPERSAL

= to allow plants to disperse seeds according to 3 user-defined distances
(see PRE_FATE.params_PFGdispersal)

Three modes of dispersal (DISPERSAL.mode) are available :

packets kernel :

homogeneous dispersal of 50% of the seeds within the d50 circle

dispersal of 49% of the seeds within the d99 - d50 ring with the same concentration as in the first circle but by pairs of pixel (see Boulangeat et al, 2014)

dispersal of 1% of the seeds within the ldd - d99 ring into one random pixel

exponential kernel : seeds are dispersed within each concentric circle (d50, d99 and ldd) according to a decreasing exponential density law (lambda = 1)

exponential kernel with probability : seeds are dispersed within each concentric circle (d50, d99 and ldd) according to a decreasing exponential density law (lambda = 1) and a continuous decreasing probability with distance

HABITAT SUITABILITY

= to influence plants fecundity and seed recruitment according to PFG preferences for habitat conditions
= filter based on maps given for each PFG within the Simul_parameters file with the PFG_HAB_MASK flag
(see PRE_FATE.params_simulParameters)

These maps must contain values between 0 and 1 corresponding to the probability of presence of the PFG in each pixel. Each year (timestep), this value will be compared to a reference value, and if superior, the PFG will be able to grow and survive.
Two methods to define this habitat suitability reference value are available (HABSUIT.ref_option) :

random : for each pixel, the reference value is drawn from a uniform distribution, and the same value is used for each PFG within this pixel.

PFG specific : for each PFG, a mean value and a standard deviation value are drawn from a uniform distribution. For each pixel and for each PFG, the reference value is drawn from a normal distribution of parameters the mean and standard deviation of the PFG.

DISTURBANCES

= to influence plant mortality and / or resprouting according to PFG tolerances to these events
(see PRE_FATE.params_PFGdisturbance)
= defined for events such as mowing, grazing, but also urbanization, crops, etc
= filter based on maps given for each disturbance within the Simul_parameters file with the DIST_MASK flag
(see PRE_FATE.params_simulParameters)

These maps, containing either 0 or 1, define the impact zone of each perturbation, and the user will have to define how each PFG will be impacted depending on age and life stage.
DIST.no

the number of different disturbances

DIST.no_sub

the number of way a PFG could react to a perturbation

DIST.freq

the frequency of each disturbance (in years)

DROUGHT

= to experience extreme events with a direct and a delayed response on PFG
= based on a map given within the Simul_parameters file with the DROUGHT_MASK flag
(see PRE_FATE.params_simulParameters)

This map must contain values representing proxies for drought intensity, like moisture values, in the sense that the lower the values, the higher the chance of experiencing a drought event. Developed canopy closure helps to reduce these values. The intensity of the drought event (moderate or severe) is determined based on thresholds defined for each PFG according to, for example, their moisture preference, as well as the number of cumulated consecutive years during which the PFG experienced a drought (see PRE_FATE.params_PFGdrought).
no drought

if \(di_y > \text{threshold.MOD}_{\text{ PFG}_i}\), the counter of cumulated consecutive years of drought experienced by the PFG will decrease : $$\text{counter}_{\text{ PFG}_i} = \text{counter}_{\text{ PFG}_i} - \text{counter.RECOVERY}_{\text{ PFG}_i}$$

moderate drought

if \(\text{threshold.SEV}_{\text{ PFG}_i} < di_y < \text{threshold.MOD}_{\text{ PFG}_i}\)

if \(di_y < \text{threshold.SEV}_{\text{ PFG}_i} \;\; \text{ & } \;\; \text{counter}_{\text{ PFG}_i} = 0\)

then fecundity and recruitment are set to 0 for this year, and counter is incremented : \(\text{counter}_{\text{ PFG}_i} ++\)

severe drought

if \(di_y < \text{threshold.SEV}_{\text{ PFG}_i} \;\; \text{ & } \;\; \text{counter.SENS}_{\text{ PFG}_i} \leq \text{counter}_{\text{ PFG}_i} < \text{counter.CUM}_{\text{ PFG}_i}\)

if \(\text{counter}_{\text{ PFG}_i} \geq \text{counter.CUM}_{\text{ PFG}_i}\)

then PFG experiences immediate drought-related mortality ; and the year after, fecundity and recruitment will be set to 0 and PFG will experience delayed drought-related mortality.

As for the disturbances module, the user will have to define how each PFG will be impacted depending on age and life stage.
(DROUGHT.no)

!not required!
= 2, the immediate and delayed responses

DROUGHT.no_sub

the number of way a PFG could react to each of these two perturbations

(DROUGHT.freq)

!not required!
the map of drought intensity proxy defined within the Simul_parameters file with the DROUGHT_MASK flag, as well as the DROUGHT_CHANGEMASK_YEARS and DROUGHT_CHANGEMASK_FILES flags, make it possible to manage the frequency and the variation of drought values (see PRE_FATE.params_simulParameters)

INVASIVE
INTRODUCTION

= to add new PFG during the simulation
= defined for events such as invasive introduction, colonization, but also new crops development, reintroduction, etc
= filter based on maps given for each PFG within the Simul_parameters file with the PFG_MASK_ALIENS flag
(see PRE_FATE.params_simulParameters)

These maps, containing either 0 or 1, define the introduction areas.
If the habitat suitability filter is on, suitability maps will also be needed for these new groups.
ALIEN.no

the number of different introductions

ALIEN.freq

the frequency of each introduction (in years)

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_strata = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_abund_low = 30000
                                 , required.max_abund_medium = 50000
                                 , required.max_abund_high = 90000
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = 130000
                                 , LIGHT.thresh_low = 190000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = 1
                                 )

## Create SEVERAL Global_parameters files
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_strata = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_abund_low = 30000
                                 , required.max_abund_medium = 50000
                                 , required.max_abund_high = 90000
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = 130000
                                 , LIGHT.thresh_low = 190000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = c(1,2)
                                 )

## ----------------------------------------------------------------------------------------- ##

## Load example data