Create Global_parameters parameter file for a FATE-HD simulation

This script is designed to create one parameter file containing GLOBAL PARAMETERS used in FATE-HD model.

PRE_FATE.params_globalParameters(name.simulation, opt.no_CPU = 1,
  opt.replacePrevious = FALSE, required.no_PFG, required.no_STRATA,
  required.simul_duration, required.seeding_duration,
  required.seeding_timestep, required.seeding_input,
  required.max_by_cohort, required.max_abund_low,
  required.max_abund_medium, required.max_abund_high, doLight = FALSE,
  LIGHT.thresh_medium, LIGHT.thresh_low, doDispersal = FALSE,
  DISPERSAL.mode, doHabSuitability = FALSE, HABSUIT.ref_option,
  doDisturbances = FALSE, DIST.no, DIST.no_sub, DIST.freq,
  doSoil = FALSE, doFire = FALSE, doDrought = FALSE,
  doHabStability = FALSE, doAliens = FALSE)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

opt.no_CPU
default 1 (optional). The number of resources that can be used to parallelize the FATE-HD simulation

opt.replacePrevious
default FALSE (optional). If TRUE, pre-existing files inside name.simulation/DATA/GLOBAL_PARAMETERS folder will be replaced

required.no_PFG
an integer corresponding to the number of PFG

required.no_STRATA
an integer corresponding to the number of height strata

required.simul_duration
an integer corresponding to the duration of simulation (in years)

required.seeding_duration
an integer corresponding to the duration of seeding (in years)

required.seeding_timestep
an integer corresponding to the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

required.seeding_input
an integer corresponding to the number of seeds attributed to each PFG at each time step, and until the seeding duration is not over

required.max_by_cohort
an integer in the order of 1 000 000 to rescale abundance values of each cohort in each pixel (carrying capacity equivalent)

required.max_abund_low
an integer in the order of 1 000 000 to rescale abundance values of small PFG

required.max_abund_medium
an integer in the order of 1 000 000 to rescale abundance values of intermediate PFG

required.max_abund_high
an integer in the order of 1 000 000 to rescale abundance values of tall PFG

doLight
default FALSE. If TRUE, light competition is activated in the FATE-HD simulation, and associated parameters are required

LIGHT.thresh_medium
an integer in the order of 1 000 000 to convert PFG abundances in each strata into light resources. It corresponds to the limit of abundances above which light resources are medium. PFG abundances lower than this threshold imply high amount of light. It is consequently lower than LIGHT.thresh_low.

LIGHT.thresh_low
an integer in the order of 1 000 000 to convert PFG abundances in each strata into light resources. It corresponds to the limit of abundances above which light resources are low. PFG abundances higher than LIGHT.thresh_medium and lower than this threshold imply medium amount of light.

doDispersal
default FALSE. If TRUE, seed dispersal is activated in the FATE-HD simulation, and associated parameters are required

DISPERSAL.mode
an integer corresponding to the way of simulating the seed dispersal for each PFG, either uniform kernel (1), exponential kernel (2) or exponential kernel with probability (3)

doHabSuitability
default FALSE. If TRUE, habitat suitability is activated in the FATE-HD simulation, and associated parameters are required

HABSUIT.ref_option
an integer corresponding to the way of simulating the habitat suitability variation between years for each PFG, either random (1) or PFG specific (2)

doDisturbances
default FALSE. If TRUE, disturbances are applied in the FATE-HD simulation, and associated parameters are required

DIST.no
the number of disturbances

DIST.no_sub
the number of way a PFG could react to a disturbance

DIST.freq
the frequency of each disturbance (in years)

doSoil
default FALSE. If TRUE, soil competition is activated in the FATE-HD simulation, and associated parameters are required

doFire
to be filled

doDrought
to be filled

doHabStability
to be filled

doAliens
to be filled

Value

A .txt file into the name.simulation/DATA/GLOBAL_PARAMETERS directory with the following parameters :

NB_CPUS

NB_FG

NB_STRATUM

SIMULATION_DURATION

SEEDING_DURATION

SEEDING_TIMESTEP

SEEDING_INPUT

MAX_BY_COHORT

MAX_ABUND_LOW

MAX_ABUND_MEDIUM

MAX_ABUND_HIGH

If the simulation includes light competition :

DO_LIGHT_COMPETITION

LIGHT_THRESH_MEDIUM

LIGHT_THRESH_LOW

If the simulation includes dispersal :

DO_DISPERSAL

DISPERSAL_MODE

If the simulation includes habitat suitability :

DO_HAB_SUITABILITY

HABSUIT_OPTION

If the simulation includes disturbances :

DO_DISTURBANCES

NB_DISTURBANCES

NB_SUBDISTURBANCES

FREQ_DISTURBANCES

If the simulation includes soil competition :

DO_SOIL_COMPETITION

If the simulation includes drought disturbances :

DO_DROUGHT_DISTURBANCES

NB_SUBDROUGHT

CHRONO_POST_DROUGHT

CHRONO_CURR_DROUGHT

If the simulation includes habitat stability check :

DO_HAB_STABILITY

NB_HABITATS

If the simulation includes aliens introduction :

DO_ALIENS_DISTURBANCE

FREQ_ALIENS

Details

The core module of FATE-HD requires several parameters to define general characteristics of the simulation :

Studied system

required.no_PFG : the number of Plant Functional Groups that will be included into the simulation. This number should match with the number of files that will be given with the characteristics of each group (SUCC, DISP, ...).

required.no_STRATA : the number of height strata that will be used into the succession module. This number should match with the maximum number of strata possible defined into the PFG SUCC files.

Simulation timing

required.simul_duration : the duration of simulation (in years)

required.seeding_duration : the duration of seeding (in years)

required.seeding_timestep : the time interval at which occurs the seeding, and until the seeding duration is not over (in years)

required.seeding_input : the number of seeds dispersed for each PFG at each time step, and until the seeding duration is not over

To get abundances
per PFG per pixel

per strata

for all strata

$$abund = 10 000 * \frac{totalAbund}{\text{required.max_abund_[low/medium/high]}}$$

The other modules of FATE-HD can be activated within this file, and if so, some additional parameters will be required :

LIGHT

= to influence plants fecundity and seed recruitment according to PFG preferences for light conditions
= light resources are calculated as a proxy of PFG abundances within each height stratum

To transform PFG
abundances into
light resources $$abund < \text{LIGHT.thresh_medium} \Leftrightarrow Light = High$$ $$abund > \text{LIGHT.thresh_medium } \& \\ abund < \text{LIGHT.thresh_low} \Leftrightarrow Light = Medium$$ $$abund > \text{LIGHT.thresh_low} \Leftrightarrow Light = Low$$

DISPERSAL

= to allow plants to disperse seeds according to 3 user-defined distances

DISPERSAL.mode : three modes of dispersal are available :

uniform kernel [1] : homogeneous dispersal within the d50, d99 and ldd circles

exponential kernel [2] : seeds are dispersed within each concentric circle (d50, d99 and ldd) according to a decreasing exponential density law (lambda = 1)

exponential kernel with probability [3] : seeds are dispersed within each concentric circle (d50, d99 and ldd) according to a decreasing exponential density law (lambda = 1) and a continuous decreasing probability with distance

homogeneous dispersal
EVERYWHERE : (!not available YET!)

HABITAT SUITABILITY

= to influence plants fecundity and seed recruitment according to PFG preferences for habitat conditions
= filter based on maps given for each PFG within the Simul_parameters file with the PFG_HAB_MASK flag.

These maps must contain values between 0 and 1 corresponding to the probability of presence of the group in each pixel. Each year (timestep), this value will be compared to a reference value, and if superior, the PFG will be able to grow and survive.

HABSUIT.ref_option : the habitat suitability reference value can be set in two possible ways :

random : for each pixel, the reference value is drawn from a uniform distribution, and the same value is used for each PFG within this pixel.

PFG specific : for each PFG, a mean value and a standard deviation value are drawn from a uniform distribution. For each pixel and for each PFG, the reference value is drawn from a normal distribution of parameters the mean and standard deviation of the PFG.

DISTURBANCES

= defined for events such as mowing, grazing, but also urbanization, crops, etc.

The impact zone is defined with a mask (map with 0 or 1) and the user will have to define how each PFG will be impacted depending on age and life stage.

DIST.no : the number of different disturbances

DIST.no_sub : the number of way a PFG could react to a perturbation

DIST.freq : the frequency of each disturbance (in years)

SOIL

= to influence plants mortality according to PFG preferences for soil conditions (fertility)
= soil composition is calculated as the weighted mean of each PFG's contribution

DROUGHT

= to experience extreme events with a direct and a delayed response on PFG.

The PFG's past drought exposure is compared to current moisture values as a proxy of drought index. Developed canopy closure helps to reduce these values. Each PFG is impacted if submitted one year to drought, and also (and more consequently) if submitted several years in a row.

DROUGHT.no_sub : the number of way a PFG could react to a perturbation

DROUGHT.chrono_post : if post-drought effects will occur previously to succession (0) or after (1)

DROUGHT.chrono_curr : if current-drought effects will occur previously to succession (0) or after (1)

HABITAT STABILITY

= to evaluate through the evolution of the PFG composition in different habitats.

The habitat distribution is given with a categorical map with each value corresponding to a specific habitat. Every year, the abundance of each PFG within each habitat is recorded, as well as the evenness of the habitat : $$ \text{evenness} = - \frac{\Sigma(\text{PFG.proportion} * log(\text{PFG.proportion}))}{log(\text{no.PFG} - 1)} $$ $$\text{with PFG.proportion} = \frac{\text{abund.PFG.in.Habitat}}{\text{abund.allPFG.in.Habitat}}$$ Every 5 years, a stability check is done to see if the abundance or the evenness have values out of the usual distribution.

HABSTAB.no_hab : the number of different habitats

INVASIVE
INTRODUCTION

= to add new PFG during the simulation.

The introduction areas are defined with a mask (map with 0 or 1). If the habitat suitability filter is on, suitability maps will also be needed for these new groups.

ALIEN.freq : the frequency of each introduction (in years)

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_by_cohort = 5000000
                                 , required.max_abund_low = 3000000
                                 , required.max_abund_medium = 5000000
                                 , required.max_abund_high = 9000000
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = 13000000
                                 , LIGHT.thresh_low = 19000000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = 1
                                 )

## Create SEVERAL Global_parameters files
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 3
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = 10
                                 , required.seeding_timestep = 1
                                 , required.seeding_input = 100
                                 , required.max_by_cohort = 5000000
                                 , required.max_abund_low = 3000000
                                 , required.max_abund_medium = 5000000
                                 , required.max_abund_high = 9000000
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = 13000000
                                 , LIGHT.thresh_low = 19000000
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = 1
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = c(1,2)
                                 )

## ----------------------------------------------------------------------------------------- ##

## Load example data
.loadData("PNE_PARAM")

## PNE_PARAM$global : vector

## Create a skeleton folder
PRE_FATE.skeletonDirectory(name.simulation = "FATE_PNE")

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_PNE"
                                 , required.no_PFG = PNE_PARAM$global["NB_FG"]
                                 , required.no_STRATA = PNE_PARAM$global["NB_STRATUM"]
                                 , required.simul_duration = PNE_PARAM$global["SIMULATION_DURATION"]
                                 , required.seeding_duration = PNE_PARAM$global["SEEDING_DURATION"]
                                 , required.seeding_timestep = PNE_PARAM$global["SEEDING_TIMESTEP"]
                                 , required.seeding_input = PNE_PARAM$global["SEEDING_INPUT"]
                                 , required.max_by_cohort = PNE_PARAM$global["MAX_BY_COHORT"]
                                 , required.max_abund_low = PNE_PARAM$global["MAX_ABUND_LOW"]
                                 , required.max_abund_medium = PNE_PARAM$global["MAX_ABUND_MEDIUM"]
                                 , required.max_abund_high = PNE_PARAM$global["MAX_ABUND_HIGH"]
                                 , doLight = TRUE
                                 , LIGHT.thresh_medium = PNE_PARAM$global["LIGHT.thresh_medium"]
                                 , LIGHT.thresh_low = PNE_PARAM$global["LIGHT.thresh_low"]
                                 , doDispersal = TRUE
                                 , DISPERSAL.mode = PNE_PARAM$global["DISPERSAL.mode"]
                                 , doHabSuitability = TRUE
                                 , HABSUIT.ref_option = PNE_PARAM$global["HABSUIT.ref_option"]
                                 , doDisturbances = TRUE
                                 , DIST.no = PNE_PARAM$global["DIST.no"]
                                 , DIST.no_sub = PNE_PARAM$global["DIST.no_sub"]
                                 , DIST.freq = rep(PNE_PARAM$global["DIST.freq"]
                                                   , PNE_PARAM$global["DIST.no"])
                                 )