Create multiple set(s) of parameter files for a FATE-HD simulation

This script is designed to create multiple sets of parameters using Latin Hypercube Sampling to help find best combination of parameters (see Details)

PRE_FATE.params_multipleSet(name.simulation.1, name.simulation.2 = NULL,
  file.simulParam.1, file.simulParam.2 = NULL, no_simulations,
  opt.percent_max = 0.5, opt.percent_seeding = 0.5,
  opt.percent_light = 0.5, do.max_by_cohort = TRUE,
  do.max_abund_low = TRUE, do.max_abund_medium = TRUE,
  do.max_abund_high = TRUE, do.seeding_duration = TRUE,
  do.seeding_timestep = TRUE, do.seeding_input = TRUE,
  do.nb_stratum = TRUE, do.LIGHT.thresh_medium = TRUE,
  do.LIGHT.thresh_low = TRUE, do.DISPERSAL.mode = TRUE,
  do.HABSUIT.ref_option = TRUE)

Arguments

name.simulation.1
a string that corresponds to the main directory or simulation name of the FATE-HD simulation from which to retrieve the first parameter simulation file (file.simulParam.1), and the second if given (file.simulParam.2) and no other directory provided (name.simulation.2 = NULL)

name.simulation.2
default NULL (optional).
A string that corresponds to the main directory or simulation name of the FATE-HD simulation from which to retrieve the second parameter simulation file (file.simulParam.2)

file.simulParam.1
a string that corresponds to the name of the simulation parameter file from which to retrieve parameter values that will be used to build the multiple set of new parameters

file.simulParam.2
default NULL (optional).
A string that corresponds to the name of the second simulation parameter file from which to retrieve parameter values that will be used to build parameter ranges in comparison with values from file.simulParam.1

no_simulations
an integer corresponding to the number of set of parameters that will be produced according to Latin Hypercube Sampling (LHS)

opt.percent_max
default 0.5. Amount of variation (between 0 and 1) around the original value of MAX_BY_COHORT, MAX_ABUND_LOW, MAX_ABUND_MEDIUM, MAX_ABUND_HIGH if selected

opt.percent_seeding
default 0.5. Amount of variation (between 0 and 1) around the original value of SEEDING_DURATION, SEEDING_TIMESTEP, SEEDING_INPUT if selected

opt.percent_light
default 0.5. Amount of variation (between 0 and 1) around the original value of LIGHT_THRESH_MEDIUM, LIGHT_THRESH_LOW if selected

do.max_by_cohort
default TRUE. If TRUE, MAX_BY_COHORT parameter within Global_parameters file will be declined into a range of values

do.max_abund_low
default TRUE. If TRUE, MAX_ABUND_LOW parameter within Global_parameters file will be declined into a range of values

do.max_abund_medium
default TRUE. If TRUE, MAX_ABUND_MEDIUM parameter within Global_parameters file will be declined into a range of values

do.max_abund_high
default TRUE. If TRUE, MAX_ABUND_HIGH parameter within Global_parameters file will be declined into a range of values

do.seeding_duration
default TRUE. If TRUE, SEEDING_DURATION parameter within Global_parameters file will be declined into a range of values

do.seeding_timestep
default TRUE. If TRUE, SEEDING_TIMESTEP parameter within Global_parameters file will be declined into a range of values

do.seeding_input
default TRUE. If TRUE, SEEDING_INPUT parameter within Global_parameters file will be declined into a range of values

do.nb_stratum
default TRUE. If TRUE, NB_STRATUM parameter within Global_parameters file will be declined into a range of values, with potential impact on some parameters within PFG succession (and light) files (parameters STRATA, MAX_ABUNDANCE, IMM_SIZE, CHANG_STR_AGES (and SHADE_TOL))

do.LIGHT.thresh_medium
default TRUE. If TRUE, LIGHT_THRESH_MEDIUM parameter within Global_parameters file will be declined into a range of values

do.LIGHT.thresh_low
default TRUE. If TRUE, LIGHT_THRESH_LOW parameter within Global_parameters file will be declined into a range of values

do.DISPERSAL.mode
default TRUE. If TRUE, DISPERSAL_MODE parameter within Global_parameters file will be declined into its 3 possible values (either uniform kernel (1), exponential kernel (2) or exponential kernel with probability (3))

do.HABSUIT.ref_option
default TRUE. If TRUE, HABSUIT_OPTION parameter within Global_parameters file will be declined into its 2 possible values (either random (1) or PFG specific (2))

Value

A new folder containing the different sets of parameters asked.

Depending on what elements have been asked to be varied, three types of files can have been modified :

the global parameter file

the PFG succession files

the PFG light succession files

Below are listed the parameters that can change (if selected) within each file :

Into the name.simulation/DATA/GLOBAL_PARAMETERS folder :

NB_STRATUM

SEEDING_DURATION

SEEDING_TIMESTEP

SEEDING_INPUT

MAX_BY_COHORT

MAX_ABUND_LOW

MAX_ABUND_MEDIUM

MAX_ABUND_HIGH

If the simulation includes light competition :

LIGHT_THRESH_MEDIUM

LIGHT_THRESH_LOW

If the simulation includes dispersal :

DISPERSAL_MODE

If the simulation includes habitat suitability :

HABSUIT_OPTION

Into the name.simulation/DATA/PFGS/SUCC folder :

STRATA

MAX_ABUNDANCE

IMM_SIZE

CHANG_STR_AGES

Into the name.simulation/DATA/PFGS/LIGHT folder :

SHADE_TOL

Details

A FATE-HD simulation requires several parameters to define general characteristics of the simulation : they are saved within a Global_parameters file. To fit the model to a particular area and set of Plant Functional Groups (PFG), these are the parameters that should be optimized, since they are not data-dependant, unlike, for example, parameters related to PFG (height, maturity, dispersal distances, soil tolerance, etc).

(Note : this is true, except when varying the number of strata, which will have an impact on some parameters within SUCC and LIGHT PFG parameter files.)

The main idea is to start from a complete simulation folder, to select the parameters that should vary, and to create new parameter files with new parameter values based on pre-existing values.

Three possible scenarios are available :

1 folder -
1 simulation file

requested parameter values are extracted from the given simulation file

ranges are assigned to each parameter according to the specified value
e.g. : if opt.percent_seeding = 0.5, and do.seeding_duration is asked, values will be generated for this parameter between : $$SEEDING\_DURATION \pm SEEDING\_DURATION * \frac{50}{100}$$

according to the required number of parameter sets to be produced (no_simulations), Latin Hypercube Sampling is applied to select each new parameter values

parameter files are created for these new parameter values

1 folder -
2 simulation files

same as 1st scenario

ranges assigned to each parameter correspond to the extracted values (e.g. : if do.seeding_duration is asked, values will be generated for this parameter between : SEEDING_DURATION (file_simulation.1) and SEEDING_DURATION (file_simulation.2)

2 folders -
2 simulation files

same as 2nd scenario, except that the two given simulation files come from two different simulation folders

Latin Hypercube Sampling is a statistical method to generate a sampling of parameter values from a multidimensional space, while ensuring a good representation of the real variability. The range of each parameter is known, and depending on the number of set of parameters asked to be obtained at the end, each range is more or less finely cut and values are drawn in order to explore the whole space of combinations.

See also

Examples