Create Simul_parameters parameter file for a FATE-HD simulation

This script is designed to create one (or several) parameter file containing PARAMETER FILENAMES used in FATE-HD model.

PRE_FATE.params_simulParameters(name.simulation, name.mask,
  name.dist = NULL)

Arguments

name.simulation
a string that corresponds to the main directory or simulation name of the FATE-HD simulation

name.mask
a string that corresponds to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the succession module of the FATE-HD simulation will take place

name.dist
(optional)
a string that corresponds to the file name of a raster mask, with either 0 or 1 within each pixel, 1 corresponding to the cells of the studied area in which the succession module of the FATE-HD simulation will take place

Value

A .txt file into the name.simulation/PARAM_SIMUL/ directory with the following parameters :

--NAMESPACE_CONSTANTS--

--GLOBAL_PARAMS--

--SAVE_DIR--

--ARRAYS_SAVING_YEARS-- (optional)

--OBJECTS_SAVING_YEARS-- (optional)

--MASK--

--MASK_CHANGE_TIME-- (optional)

--MASK_CHANGE_MASK-- (optional)

--PFG_LIFE_HISTORY_PARAMS--

--PFG_DISPERSAL_PARAMS-- (optional)

--PFG_HAB_MASK-- (optional)

--HAB_CHANGE_TIME-- (optional)

--HAB_CHANGE_MASK-- (optional)

--PFG_DISTURBANCES_PARAMS-- (optional)

--DIST_MASK-- (optional)

--DIST_CHANGE_TIME-- (optional)

--DIST_CHANGE_MASK-- (optional)

--PFG_SOIL_PARAMS-- (optional)

--END_OF_FILE--

Details

The FATE-HD software takes only one input parameter : a file containing links to other files containing all the parameters and data needed by the program to run.

NAMESPACE_
CONSTANTS

file where constant parameters needed by the program to manage abundance values are given
(see PRE_FATE.params_namespaceConstants)

GLOBAL_PARAMS

file where parameters related to the simulation definition are referred (e.g. number of PFG involved, number of height strata, simulation duration, computer resources, modules loaded, ...)
(see PRE_FATE.params_globalParameters)

SAVE_DIR

directory where simulation outputs will be stored

ARRAYS_SAVING_
YEARS (optional)

file containing the years for which simulation maps will be saved
(see PRE_FATE.params_saveYears)

OBJECTS_SAVING_
YEARS (optional)

file containing the years for which simulation outputs will be saved
(see PRE_FATE.params_saveYears)

MASK

raster mask that will define the study area

MASK_CHANGE_TIME
(optional)

file containing the years to change rasters for the succession module
(see PRE_FATE.params_changingYears)

MASK_CHANGE_MASK
(optional)

file containing the files to change rasters for the succession module
(see PRE_FATE.params_changingYears)

PFG_LIFE_HISTORY_
PARAMS

PFG life history related parameters (one by PFG)
(see PRE_FATE.params_PFGsuccession)

PFG_DISPERSAL_
PARAMS (optional)

PFG dispersal capacity related parameters (one by PFG)
(see PRE_FATE.params_PFGdispersal)

PFG_HAB_MASK
(optional)

raster masks (one by PFG) containing PFG habitat suitability for the study area

HAB_CHANGE_TIME
(optional)

file containing the years to change rasters for the habitat suitability module
(see PRE_FATE.params_changingYears)

HAB_CHANGE_MASK
(optional)

file containing the files to change rasters for the habitat suitability module
(see PRE_FATE.params_changingYears)

PFG_DISTURBANCES_
PARAMS (optional)

PFG disturbance related parameters in terms of resprouting and mortality (one by PFG)
(see PRE_FATE.params_PFGdisturbance)

DIST_MASK
(optional)

raster masks that will define the disturbance areas

DIST_CHANGE_TIME
(optional)

file containing the years to change rasters for the disturbance module
(see PRE_FATE.params_changingYears)

DIST_CHANGE_MASK
(optional)

file containing the files to change rasters for the disturbance module
(see PRE_FATE.params_changingYears)

PFG_SOIL_
PARAMS (optional)

PFG soil contribution and tolerance related parameters (one by PFG)
(see PRE_FATE.params_PFGsoil)

Note

The links to files are NOT absolute paths.
When relative paths are used, the user should be careful of the folder from which the simulation is launched.
A function allows to transform these relative paths into absolute ones. (!not available yet!)

The order of files matters!
For instance the first link below --PFG_LIFE_HISTORY_PARAMS-- flag has to match (the same PFG) with the first item below the --PFG_DISPERSAL_PARAMS-- flag.

Examples

# NOT RUN {
## Create a skeleton folder with the default name ('FATE_simulation')
if (dir.exists("FATE_simulation")) system("rm -r FATE_simulation/")
PRE_FATE.skeletonDirectory()

## Create a Namespace_constants parameter file
PRE_FATE.params_namespaceConstants(name.simulation = "FATE_simulation"
                                   , global.abund.low = 1000000
                                   , global.abund.med = 5000000
                                   , global.abund.high = 8000000
                                   , global.max.by.cohort = 5000000
                                   , global.resource.thresh.med = 13000000
                                   , global.resource.thresh.low = 19000000)

## Create a Global_parameters file
PRE_FATE.params_globalParameters(name.simulation = "FATE_simulation"
                                 , required.no_PFG = 6
                                 , required.no_STRATA = 5
                                 , required.simul_duration = 100
                                 , required.seeding_duration = c(10,50)
                                 , required.seeding_timestep = 1)

## Create a SAVE_year_maps or/and SAVE_year_objects parameter file
PRE_FATE.params_saveYears(name.simulation = "FATE_simulation"
                          , years.maps = c(100, 150, 200)
                          , years.objects = 200)

## Create a Changing_times parameter file
PRE_FATE.params_changingYears(name.simulation = "FATE_simulation"
                              , type.changing = "DIST"
                              , mat.changing = data.frame(year = c(50,50,80,80)
                                                          , order = c(1,2,1,2)
                                                          , file.name = c("MASK_DIST1_50.tif"
                                                                          , "MASK_DIST2_50.tif"
                                                                          , "MASK_DIST1_80.tif"
                                                                          , "MASK_DIST2_80.tif")))

## Create PFG succession parameter files
PRE_FATE.params_PFGsuccession(name.simulation = "FATE_simulation"
                              , mat.PFG.succ = data.frame(PFG = paste0("PFG",1:6)
                                                          , type = c("C", "C", "H", "H", "P", "P")
                                                          , height = c(10, 250, 36, 68, 1250, 550)
                                                          , maturity = c(5, 5, 3, 3, 8, 9)
                                                          , longevity = c(12, 200, 25, 4, 110, 70)))

## Create PFG dispersal parameter files
PRE_FATE.params_PFGdispersal(name.simulation = "FATE_simulation"
                             , mat.PFG.disp = data.frame(PFG = paste0("PFG", 1:6)
                                                         , MODE = 1
                                                         , d50 = rep(c(500, 500, 100),2)
                                                         , d99 = rep(c(10000, 15000, 20000),2)
                                                         , ldd = rep(c(100000, 50000, 100000),2)))

## Create PFG disturbance parameter files
PRE_FATE.params_PFGdisturbance(name.simulation = "FATE_simulation"
                               , mat.PFG.dist = data.frame(name = rep(c("DIST1","DIST2"), each = 4)
                                                           , responseStage = rep(1:4, 2)
                                                           , KilledIndiv_H = c(0,0,0,0,1,1,0,0)
                                                           , KilledIndiv_C = c(0,10,10,10,1,1,0,0)
                                                           , KilledIndiv_P = c(10,10,10,10,10,0,0,0)
                                                           , ResproutIndiv_H = c(0,0,9,10,0,0,5,1)
                                                           , ResproutIndiv_C = c(0,0,0,0,0,0,5,1)
                                                           , ResproutIndiv_P = c(0,0,0,0,0,0,0,0)))

## Create PFG soil parameter files
PRE_FATE.params_PFGsoil(name.simulation = "FATE_simulation"
                        , mat.PFG.soil = data.frame(PFG = paste0("PFG", 1:6)
                                                    , soil_contrib = c(2.5, 3, 4.8, 2.5, 3, 4.8)
                                                    , soil_tol_min = c(2, 3, 3, 2, 3, 3)
                                                    , soil_tol_max = c(3, 3, 6, 3, 3, 6)))

## Create a Simulation parameter file
file.create("FATE_simulation/DATA/MASK/mask.tif")
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation"
                                , name.mask = "mask.tif")

for (pfg in 1:6){
  file.create(paste0("FATE_simulation/DATA/PFGS/ENVSUIT/mask_PFG", pfg, ".tif"))
}
PRE_FATE.params_simulParameters(name.simulation = "FATE_simulation"
                                , name.mask = "mask.tif")

# }