Selection of dominant species from abundance releves

This script is designed to select dominant species from abundance records, and habitat if the information is available.

PRE_FATE.selectDominant(mat.site.species.abund = NULL, sites = NULL,
  species = NULL, abund = NULL, habitat = NULL,
  selectionRule.quanti = 0, selectionRule.min_mean_abund = 0,
  selectionRule.min_no_abund_over25 = 0, doHabitatSelection = FALSE,
  selectionRule.min_percent_habitat = 0.05,
  selectionRule.min_no_habitat = 5)

Arguments

mat.site.species.abund
a data.frame with at least 3 columns : sites ID, species ID, abundance values as frequence values(e.g. Braun Blanquet) with NA when no information, habitat ID (optional)

sites
a vector with sites ID

species
a vector with species ID

abund
a vector with abundance values (e.g. Braun Blanquet) with NA when no information

habitat
a vector with habitat ID

selectionRule.quanti
minimum quantile of total number of sites that must occupy the species (between 0 and 1)

selectionRule.min_mean_abund
minimal average abundance of the species

selectionRule.min_no_abund_over25
minimum number of sites where species is dominant (>= 25 % of coverage)

doHabitatSelection
default FALSE. If TRUE, selection is also done including constraints on habitat class

selectionRule.min_percent_habitat
minimum percentage of habitat class occupied by the species (between 0 and 1)

selectionRule.min_no_habitat
minimum number of sites of specific habitat occupied by the species

Value

A data.frame with all the species selected and the values of parameters used to make the selection :

stat.no_sites_recorded

number of sites with information (presence-absence or abundance)

stat.no_sites_abund

number of sites with abundance information

stat.abund_median

median abundance (coverage percentage)

stat.abund_mean

mean abundance (coverage percentage)

stat.abund_max

maximal abundance (coverage percentage)

stat.no_sites
abund_max

number of sites with maximal abundance

stat.no_sites
abund_over25

number of sites with abundance >= 25

SELECTION

dataset in which the species has been selected as dominant
(global = all data, or habitat, or global + habitat...)

The information is written in PRE_FATE_DOMINANT_species_selected_[...].csv files :

COMPLETE_TABLE

the complete table of selected species with all the statistics described above

SPECIES_ONLY

only the names / ID of the species selected

Two PRE_FATE_DOMINANT_[...].pdf files are also created :

STEP_1
selectionCriteria

to visualize the values of species metrics, and the criteria used to make the selection

STEP_2
selectedSpecies

to visualize the values of species metrics, for the selected dominant species

Details

This function provides a way to select dominant species based on abundance sampling information.
This information can be given either directly in the form of a data.frame, or indirectly with a vector for each required information (sites, species, abundance, habitat class - optional).

2 rules are applied to make the species selection :

1. Presence releves

on number of sites

the species should be found in a certain number of sites, which corresponds to the quantile selectionRule.quanti of all the number of records per species
This should ensure that the selected species are covering all the studied area (or at least a determining part of it, assuming that the releves are well distributed over all the zone).

2. Abundance releves :

both conditions must be fullfilled

on average abundance

the species should have a mean abundance superior or equal to selectionRule.min_mean_abund

on dominancy

the species should be dominant (i.e. represent at least 25 % of the coverage of the site) in at least selectionRule.min_no_abund_over25 sites

This should ensure that the selected species, besides being abundant, are frequently abundant, meaning not ultra specific to a particular area.

If habitat information is available (e.g. type of environment : urban, desert, grassland... ; type of vegetation : shrubs, forest, alpine grasslands... ; etc), 1 rule with 2 conditions can be added to the selection steps, in order to keep species that are not dominant at the large scale but could be representative of a specific habitat :

3. Releves per habitat :

both conditions must be fullfilled

on occupancy

the species should occupy at least selectionRule.min_percent_habitat % of a landclass

on representation

the species should be found in at least selectionRule.min_no_habitat sites of a landclass

In the same way than for rule 2., this should allow to select species that are not globally dominant but that can be representative of a specific habitat, both in terms of abundance and dominancy.

See also

Examples

# NOT RUN {
## Load example data
data(MontBlanc)
str(MontBlanc)

## MontBlanc$mat.releves : data.frame

## Transformation of Braun-Blanquet abundances
MontBlanc$mat.releves$abund = PRE_FATE.abundBraunBlanquet(abund = MontBlanc$mat.releves$abund)

## Selection of dominant species

#########################################################################################
## EXAMPLE 1 : With default parametrization
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves)

#########################################################################################
## EXAMPLE 2 : Redefinition of global selection rules
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10)

#########################################################################################
## EXAMPLE 3 : Add habitat selection with default parametrization
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10
                                 , doHabitatSelection = TRUE)

#########################################################################################       
## EXAMPLE 4 : Redefinition of habitat selection rules
#########################################################################################

sp.DOM = PRE_FATE.selectDominant(mat.site.species.abund = MontBlanc$mat.releves
                                 , selectionRule.quanti = 0.9
                                 , selectionRule.min_mean_abund = 25
                                 , selectionRule.min_no_abund_over25 = 10
                                 , doHabitatSelection = TRUE
                                 , selectionRule.min_percent_habitat = 0.05
                                 , selectionRule.min_no_habitat = 10)

#########################################################################################

## 2 .csv and 2 .pdf files have been produced
str(sp.DOM)

# }