Create the skeleton folder for a FATE-HD simulation

This script is designed to create a user-friendly directory tree to run a FATE-HD simulation.

PRE_FATE.skeletonDirectory(name.simulation = "FATE_simulation")

Arguments

name.simulation
a string that will be used as the main directory and simulation name

Value

A directory tree with folder to contain the parameter files, the simulation files and the results.

Details

FATE-HD requires only one input parameter (see PRE_FATE.params_simulParameters), which is a file containing the names of parameter files, which may themselves contain parameters (e.g. succession, dispersal files...) or other file names (e.g. disturbance or environmental change masks.
The point is : the user could give names of files stored everywhere on a machine, and does not have to put them all in one same place.

But as this is more practical, this function proposes a way to organize all those files or parameter files that will or could be used by a FATE-HD simulation.

The tree structure is detailed below :

DATA

this folder will contain all the data or parameters that are needed by the model

GLOBAL_PARAMETERS : files containing global parameters for the simulation

MASK : all maps used in the model

SCENARIO : files containing information about changes in input data ( e.g. habitat suitability maps, disturbances maps, etc)

SAVE : files containing information about times to save outputs

PFGS :

SUCC : all the PFG life history parameter files

LIGHT : all the PFG light parameter files

DISP : all the PFG dispersal parameter files

DIST : all the PFG-response to disturbances parameter files

SOIL : all the PFG soil parameter files

HABSUIT : all the PFG habitat suitability maps

PARAM_SIMUL

this folder will contain simulation files that can be given as input to the software

RESULTS

this folder will collect all the results produced by the software with a folder for each simulation

See also

Examples

## Create a skeleton folder with the default name ('FATE_simulation')
PRE_FATE.skeletonDirectory()

## Create a skeleton folder with a specific name
PRE_FATE.skeletonDirectory(name.simulation = "FATE_AlpineForest")