Create clusters based on dissimilarity matrix

This script is designed to create clusters of species based on a distance matrix between those species. Several metrics are computed to evaluate these clusters and a graphic is produced to help the user to choose the best number of clusters (by subset of data if distance matrices are given for each of them).

PRE_FATE.speciesClustering_step1(mat.species.DIST)

Arguments

mat.species.DIST
a dist object corresponding to the distance between each pair of species, or a list of dist objects, one for each GROUP value. Such an object can be obtained with the PRE_FATE.speciesDistance function.

Value

A list object with 2 elements :

clust.dendograms

a list object with as many objects of class hclust as subset of data

clust.evaluation

a data.frame with 4 columns :

group : name of sub-dataset

nb.cluster : number of clusters used for the clustering

variable : evaluation metrics' name

value : value of evaluation metric

Two PRE_FATE_CLUSTERING_[...].pdf files are created :

STEP_1A
clusteringMethod

to account for the chosen clustering method

STEP_1B
numberOfClusters

for decision support, to help the user to choose the adequate number of clusters to be used into the hclust method

Details

This function allows one to obtain dendograms based on a distance matrix between species.

As for the PRE_FATE.speciesDistance method, clustering can be made for sub-datasets, conditioning that mat.species.DIST parameter is given as a list of dist object (instead of a dist object alone).

The process is as follows :

1. Choice of the
optimal
clustering method

hierarchical clustering on the dissimilarity matrix is realized with the hclust function.

Several methods are available for the agglomeration : complete, ward.D, ward.D2, single, average (UPGMA), mcquitty (WPGMA), median (WPGMC) and centroid (UPGMC).

Mouchet et el. (2008) proposed of measure of similarity between the input distance and the one obtained with the clustering which must be minimized to help finding the best clustering method : $$ 1 - cor( input.DIST, clustering.DIST ) ^ 2$$

For each agglomeration method, this measure is calculated. The method that minimizes it is kept and used for further analyses. A graphic (STEP_1A) is made to account for the comparison of these methods.

2. Evaluation of the
clustering

once the hierarchical clustering is done, the number of clusters to keep should be chosen. To do that, several metrics are computed :

Dunn index (mdunn) : ratio of the smallest distance between observations not in the same cluster to the largest intra-cluster distance. Value between zero and infinity, and should be maximized.

Meila's Variation of Information index (mVI) : measures the amount of information lost and gained in changing between 2 clusterings. Should be minimized.

Coefficient of determination (R2) : value between zero and one. Should be maximized.

Calinski and Harabasz index (ch) : the higher the value, the "better" is the solution.

Corrected rand index (Rand) : measures the similarity between two data clusterings. Value between 0 and 1, with 0 indicating that the two data clusters do not agree on any pair of points and 1 indicating that the data clusters are exactly the same.

Average silhouette width (av.sil) : Observations with a large s(i) (almost 1) are very well clustered, a small s(i) (around 0) means that the observation lies between two clusters, and observations with a negative s(i) are probably placed in the wrong cluster. Should be maximized.

A graphic (STEP_1B) is produced, giving the values of these metrics in function of the number of clusters used. Number of clusters with evaluation metrics' values among the 3 best are highlighted to help the user to make his/her optimal choice.

Note

The function does not return ONE dendogram (or as many as given dissimilarity structures) but a list with all tested numbers of clusters. One final dendogram can then be obtained using this result as a parameter in the PRE_FATE.speciesClustering_step2 function.

See also

Examples

# NOT RUN {
## Load example data
data(MontBlanc)
str(MontBlanc)

## MontBlanc$mat.traits : data.frame
## MontBlanc$mat.nicheOverlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits
                                   , mat.species.overlap = MontBlanc$mat.nicheOverlap
                                   , min.info.thresh = 0.9)

str(sp.DIST)

## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

# }