Choose clusters and select determinant species

This script is designed to obtain functional groups by : 1) select the number of clusters to be kept from an object obtained with the PRE_FATE.speciesClustering_step1 function ; 2) refine these groups by identifying determinant species in each of them.

PRE_FATE.speciesClustering_step2(clust.dendograms, no.clusters,
  mat.species.DIST)

Arguments

clust.dendograms
a dendogram or a list of dendograms with one for each GROUP value, as can be obtained with the PRE_FATE.speciesClustering_step1 function.

no.clusters
a vector with the number of clusters to be kept in each subset of data (if the data is split).

mat.species.DIST
a dist object corresponding to the distance between each pair of species, or a list of dist objects, one for each GROUP value. Such an object can be obtained with the PRE_FATE.speciesDistance function.

Value

A list object with 2 elements :

determ.sp

a vector with the names of all determinant species

determ.all

a data.frame containing all species (determinant and non-determinant) with 10 columns :

pfg : the ID of the PFG (group + no.cluster)

group : name of sub-dataset

no.cluster : cluster number

sp : name of species

ID : species number in each PFG

sp.mean.dist : species mean distance to other species of the same PFG

allSp.mean : mean(sp.mean.dist) within the PFG

allSp.min : mean(sp.mean.dist) - 1.64 * sd(sp.mean.dist) within the PFG

allSp.max : mean(sp.mean.dist) + 1.64 * sd(sp.mean.dist) within the PFG

toSuppr : 0 if determinant species, 1 otherwise

Two PRE_FATE_CLUSTERING_[...].pdf files are created :

STEP_2C
distantSpecies

to visualize in each PFG the distribution of mean distance of each species to other species, and non-determinant species which are outside the distribution

STEP_2C
PCO

to visualize in each PFG the distribution of species, with and without non-determinant species

Details

This function allows one to obtain a classification of dominant species into Plant Functional Groups (PFG), and the determinant species based on these PFGs.

What is the difference between dominant and determinant species ?

Dominant species are species representative of an environment or a studied area, in terms of number of releves or abundance values. They can be found with the PRE_FATE.selectDominant function of this package. These dominant species are used to build PFG with the PRE_FATE.speciesClustering_step1 function.

Once PFG are built, determinant species are defined as refined subsets of dominant species within each PFG. The process is detailed below :

each dominant species is assigned to a PFG

within each PFG :

for each species, compute its mean distance to the other species within the PFG (sp.mean.dist)

calculate the mean value of all these mean distances (allSp.mean)

calculate the deviation values around this mean value (allSp.min and allSp.max)

determinant species are the ones that are included between these deviation values

See also

Examples

# NOT RUN {
## Load example data
data(MontBlanc)

## MontBlanc$mat.traits : data.frame
## MontBlanc$mat.nicheOverlap : niolap object

## Calculate distance between species
sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits
                                   , mat.species.overlap = MontBlanc$mat.nicheOverlap
                                   , min.info.thresh = 0.9)

str(sp.DIST)

## Run hierarchical clustering and number of clusters' selection                      
sp.CLUST = PRE_FATE.speciesClustering_step1(mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.CLUST)

## Select number of clusters and find determinant species
sp.DETERM = PRE_FATE.speciesClustering_step2(clust.dendograms = sp.CLUST$clust.dendograms
                                             , no.clusters = c(11, 7, 8)
                                             , mat.species.DIST = sp.DIST)

## 2 .pdf files have been produced
str(sp.DETERM)

# }