Calculate PFG traits values based on determinant species traits values

This script is designed to create

PRE_FATE.speciesClustering_step3(mat.species.traits)

Arguments

mat.species.traits
A data.frame with at least 3 columns :

species

the ID of each determinant species (see PRE_FATE.speciesClustering_step2)

PFG

a factor variable containing grouping information to divide the species into Plant Functional Groups (see PRE_FATE.speciesClustering_step2)

...

one column for each functional trait (see Details)

Value

A .txt file per PFG into the name.simulation/DATA/PFGS/SUCC/ directory with the following parameters :

NAME : name of the PFG

MATURITY : the maturity age of the PFG (in years)

LONGEVITY : the PFG life span (in years)

MAX_ABUNDANCE : the maximal (qualitative) shade that the PFG is able to produce
(1: Low 2: Medium 3: High)

IMM_SIZE : the relative size of the immature PFG
(0: 0% 1: 10% 2: 20% 3: 30% 4: 40% 5: 50% 6: 60% 7: 70% 8: 80% 9: 90% 10: 100%)

CHANG_STR_AGES : the ages at which the PFG goes in the upper stratum
(in years, put a value higher than the PFG life span if it is not supposed to rise a stratum)

SEED_POOL_LIFE : the maximal number of years seeds are able to survive (for active and dormant pool)

SEED_DORMANCY : are the seeds dormant or not (0: No 1: Yes)

A SUCC_COMPLETE_TABLE.csv file summarizing information for all groups into the name.simulation/DATA/PFGS/ directory. This file can be used to parameterize the disturbance files.

Details

This function allows one to obtain 'average' functional trait values for each Plant Functional Group, based on values at the determinant species level.

Those functional traits can be :

type

or life-form, based on Raunkier. It should be either H (herbaceous), C (chamaephyte) or P (phanerophyte) for now

height

the maximum or average height that reach the species

maturity

the age from which the species can reproduce

longevity

the maximum or average lifespan of the species

dispersal

the age from which the species can reproduce

light

a value corresponding to the light preference of the species (from preference for shade to full light)

soil_contrib

a value between 0 and 10 corresponding to the Ellenberg nitrogen value of the PFG

soil_tolerance

the range of soil tolerance : small(1) or large(2)

palatability

the appetence of each species from null to high

See also

Examples

# NOT RUN {

# }