Computation of distances between species based on traits and niche overlap
Source: R/PRE_FATE.speciesDistance.R

This script is designed to create a distance matrix between species, combining functional distances (based on functional trait values) and niche overlap (based on co-occurrence of species).

PRE_FATE.speciesDistance(mat.species.traits, mat.species.overlap,
  min.info.thresh = 1)

Arguments

mat.species.traits
A data.frame with at least 3 columns :

species

the ID of each studied species

GROUP

a factor variable containing grouping information to divide the species into sub data sets (see Details)

...

one column for each functional trait

mat.species.overlap
Two options :

a data.frame with 2 columns :

species

the ID of each studied species

raster

path to raster file with species distribution

a dissimilarity structure representing the niche overlap between each pair of species. It can be a dist object, a niolap object, or simply a matrix.

min.info.thresh
minimum percentage of values for each trait (between 0 and 1)

Value

A dist object corresponding to the distance between each pair of species, or a list of dist objects, one for each GROUP value.

Details

This function allows one to obtain a distance matrix between species, based on two types of distance information :

Functional traits :

The GROUP column is required if species must be separated to have one final distance matrix per GROUP value. If the column is missing, all species will be considered as part of a unique dataset.

The traits can be qualitative or quantitative, but previously identified as such (i.e. with the use of functions such as as.numeric, as.factor and ordered).

Functional distance matrix is calculated with Gower dissimilarity, using the gowdis function from FD package.

This function allows NA values. In order to let the user decide what importance to give to missing values, the min.info.thresh defines which minimum percentage of species should have values for each trait.

Niche overlap :

If a data.frame is given, the degree of niche overlap will be computed using the niche.overlap function from phyloclim package.

Functional distances and niche overlap informations are then combined according to the following formula :

$$\text{mat.DIST}_{sub-group} = [ \text{mat.OVERLAP}_{sub-group} + \text{mat.FUNCTIONAL}_{sub-group} * n_{traits} ] / [ n_{traits} + 1 ]$$

meaning that distance matrix obtained from functional information is weighted by the number of traits used.

Examples

## Load example data
data(MontBlanc)
str(MontBlanc)

#> List of 3
#>  $ mat.releves     :'data.frame':	12302 obs. of  4 variables:
#>   ..$ sites  : int [1:12302] 660398 662913 663085 663104 663104 664839 751362 751377 751484 922672 ...
#>   ..$ species: chr [1:12302] "Abies alba Miller" "Abies alba Miller" "Abies alba Miller" "Abies alba Miller" ...
#>   ..$ abund  : Factor w/ 7 levels "","+","1","2",..: 2 2 2 3 6 4 2 3 2 2 ...
#>   ..$ habitat: int [1:12302] 4 2 6 4 4 4 2 2 3 8 ...
#>  $ mat.traits      :'data.frame':	337 obs. of  6 variables:
#>   ..$ species     : chr [1:337] "X23" "X139" "X190" "X197" ...
#>   ..$ Disp        : num [1:337] 4 5 7 6 7 1 4 7 7 7 ...
#>   ..$ Light       : num [1:337] 2 2 3 3 5 3 5 5 5 3 ...
#>   ..$ Height      : num [1:337] 1500 50 70 11 30 ...
#>   ..$ Palatability: Factor w/ 5 levels "","0","1","2",..: 4 5 5 5 5 5 5 5 5 5 ...
#>   ..$ GROUP       : Factor w/ 3 levels "C","H","P": 3 2 2 2 2 2 2 2 2 2 ...
#>  $ mat.nicheOverlap: niolap [1:270, 1:270] NA 0.524 0.789 0.626 0.675 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:270] "X10053" "X10106" "X10113" "X10182" ...
#>   .. ..$ : chr [1:270] "X10053" "X10106" "X10113" "X10182" ...

## MontBlanc$mat.traits : data.frame
## MontBlanc$mat.nicheOverlap : niolap object
# sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits
#                                    , mat.species.overlap = MontBlanc$mat.nicheOverlap
#                                    , min.info.thresh = 1)

sp.DIST = PRE_FATE.speciesDistance(mat.species.traits = MontBlanc$mat.traits
                                   , mat.species.overlap = MontBlanc$mat.nicheOverlap
                                   , min.info.thresh = 0.9)

#> 
#>  ############## TRAIT INFORMATIONS ############## 
#> 
#>  Number of species :  337
#>  Measured traits :  Disp, Light, Height, Palatability
#>  Groups :  C, H, P
#>  Number of species in each group :  48 271 18
#> 
#>  Number of species with traits :  337
#>  Number of species with traits and no overlap information :  69
#>  Number of species with overlap :  270
#>  Number of species with overlap and no traits information :  2
#> 
#>  Comparison of groups' dimensions : 
#> 
#>  Group  1 :
#> Trait distances :  48 48 
#> Overlap distances :  38 38 
#> 
#>  Group  2 :
#> Trait distances :  271 271 
#> Overlap distances :  215 215 
#> 
#>  Group  3 :
#> Trait distances :  18 18 
#> Overlap distances :  15 15 
#> 
#>  Number of species with both trait and overlap distances:  268

str(sp.DIST)

#> List of 3
#>  $ C:Class 'dist'  atomic [1:703] 0.0846 0.2436 0.1439 0.3741 0.038 ...
#>   .. ..- attr(*, "Labels")= chr [1:38] "X10192" "X11256" "X12051" "X12060" ...
#>   .. ..- attr(*, "Size")= int 38
#>   .. ..- attr(*, "call")= language as.dist.default(m = mat)
#>   .. ..- attr(*, "Diag")= logi FALSE
#>   .. ..- attr(*, "Upper")= logi FALSE
#>  $ H:Class 'dist'  atomic [1:23005] 0.232 0.165 0.262 0.273 0.24 ...
#>   .. ..- attr(*, "Labels")= chr [1:215] "X10053" "X10106" "X10113" "X10182" ...
#>   .. ..- attr(*, "Size")= int 215
#>   .. ..- attr(*, "call")= language as.dist.default(m = mat)
#>   .. ..- attr(*, "Diag")= logi FALSE
#>   .. ..- attr(*, "Upper")= logi FALSE
#>  $ P:Class 'dist'  atomic [1:105] 0.382 0.192 0.277 0.252 0.395 ...
#>   .. ..- attr(*, "Labels")= chr [1:15] "X11313" "X11388" "X11726" "X11729" ...
#>   .. ..- attr(*, "Size")= int 15
#>   .. ..- attr(*, "call")= language as.dist.default(m = mat)
#>   .. ..- attr(*, "Diag")= logi FALSE
#>   .. ..- attr(*, "Upper")= logi FALSE