Compare observed and simulated habitat of a FATE simulation to a chosen year.

To compare observations and simulations, this function compute confusion matrix between observation and prediction and then compute the TSS for each habitat.

# S3 method for habitat.validation
do(
  output.path,
  RF.model,
  predict.all.map,
  sim,
  simu_PFG,
  habitat.whole.area.df
)

Arguments

output.path

access path to the for the folder where output files will be created.

RF.model

random forest model trained on CBNA data (train.RF.habitat function)

predict.all.map

Logical. If TRUE, the script will predict habitat for the whole map.

sim

name of the single simulation to validate.

simu_PFG

a data frame with simulated abundance for each PFG and strata (if option selected) and pixel ID, extracted from a FATE simulation (see POST_FATE.temporalEvolution).

habitat.whole.area.df

a data frame which contain habitat names and code for each pixel that needs validation.

Value

Habitat performance file.
If option selected, the function also returns an habitat prediction file with observed and simulated habitat for each pixel of the whole map.

Details

For a given simulation sim, this script takes the evolution abundance for each PFG and all strata (or for each PFG & each strata if option selected) data frame and predicts the habitat for the whole map (if option selected) thanks to the RF model. Finally, the function computes habitat performance based on TSS for each habitat.

Author

Matthieu Combaud & Maxime Delprat