Obtain ellipse coordinates from (PCO) X,Y and a factor value

Obtain ellipse coordinates from (PCO) X,Y and a factor value

util.ellipse(mx, my, vx, cxy, vy, coeff)
util.ELLIPSE(x, y, z)
.getELLIPSE(xy, fac)

Arguments

mx
\(\Sigma x * \frac{z}{\Sigma z}\)

my
\(\Sigma y * \frac{z}{\Sigma z}\)

vx
\(\Sigma (x - mx) * (x - mx) * \frac{z}{\Sigma z}\)

cxy
\(\Sigma (x - mx) * (y - my) * \frac{z}{\Sigma z}\)

vy
\(\Sigma (y - my) * (y - my) * \frac{z}{\Sigma z}\)

coeff
default 1

xy
a data.frame or matrix with 2 columns corresponding to individuals coordinates, extracted from example from dudi.pco analysis

fac
a vector containing group labels for individuals (with length(fac) = nrow(xy))

x
a vector corresponding to abscissa coordinates of individuals (column 1 of xy)

y
a vector corresponding to ordinate coordinates of individuals (column 2 of xy)

z
a data.frame with one column for each level represented in fac and nrow(z) = length(fac) = nrow(xy). Values are corresponding to the relative representation of each level (for level i : \(\frac{1}{N_i}\))