Load a dataset

.loadData(data.name, format.name)

Arguments

data.name
a string corresponding to the name of the dataset that will be loaded

format.name
a string corresponding to the downloading format way (either RData or 7z)

Value

#########################################################
>>> Champsaur_PFG
#########################################################

A list object with 18 elements to help building the Plant Functional Group :

name.dataset

Champsaur

sp.observations

a data.frame of dimension 127257 x 6
containing releves data about plant species in Champsaur
to be used with the PRE_FATE.selectDominant function

sites : sites ID

species : species ID

abundBB : Braun-Blanquet abundance

abund : relative abundance obtained from the abund_BB column with the PRE_FATE.abundBraunBlanquet function

habitat : habitat classes from [CESBIO2018](http://osr-cesbio.ups-tlse.fr/~oso/posts/2019-03-25-carte-s2-2018 (with some classes gathering)

TYPE : type of records, exhaustive (COMMUNITY) or single occurrences (OCCURRENCE)

rules.selectDominant

a vector containing parameters to be given to the PRE_FATE.selectDominant function
(doRuleA, rule.A1, rule.A2_quantile, doRuleB, rule.B1_percentage, rule.B1_number, rule.B2, doRuleC)

sp.SELECT

a list obtained from the PRE_FATE.selectDominant function run with sp.observations and rules.selectDominant

sp.traits

a data.frame of dimension 250 x 23
containing traits for dominant species
to be used with the PRE_FATE.speciesDistance function

species : species ID

GROUP : rough generalization of Raunkier life-forms (Phanerophyte, Chamaephyte, Herbaceous)

MATURITY / LONGEVITY / HEIGHT / LDMC / LNC / SEEDM / SLA : numerical values

LONGEVITY_log / HEIGHT_log / LDMC_log / LNC_log / SEEDM_log / SLA_log : log-transformed numerical values

DISPERSAL / LIGHT / NITROGEN / NITROGEN_TOLERANCE / MOISTURE / GRAZ_MOW_TOLERANCE / HABITAT / STRATEGY : categorical values

sp.traits.P

a data.frame of dimension 19 x 9
a subset of sp.traits (LONGEVITY_log / HEIGHT_log / SEEDM_log / SLA_log / DISPERSAL / LIGHT / NITROGEN) for Phanerophyte dominant species
to be used with the PRE_FATE.speciesDistance function

sp.traits.C

a data.frame of dimension 39 x 10
a subset of sp.traits (LONGEVITY_log / HEIGHT_log / SEEDM_log / SLA_log / LNC_log / DISPERSAL / LIGHT / NITROGEN) for Chamaephyte dominant species
to be used with the PRE_FATE.speciesDistance function

sp.traits.H

a data.frame of dimension 192 x 7
a subset of sp.traits (HEIGHT_log / SLA_log / LDMC_log / LNC_log / LIGHT) for Herbaceous dominant species
to be used with the PRE_FATE.speciesDistance function

mat.habitat

matrix of dimension 240 x 240
containing habitat dissimilarity distance (1 - Schoeners' D, obtained with gowdis) for dominant species
to be used with the PRE_FATE.speciesDistance function

tab.dom.PA

matrix of dimension 13654 x 264
containing dominant species occurrences and absences (obtained from sp.SELECT object and absences corrected with the TYPE information of sp.observations), to build mat.overlap

tab.env

matrix of dimension 13590 x 5
containing environmental values (bio1, bio12, slope, dem, CESBIO2018) for sites, to build mat.overlap

mat.overlap

matrix of dimension 243 x 243
containing niche overlap distance (1 - Schoeners' D, obtained with ecospat.niche.overlap) for dominant species
to be used with the PRE_FATE.speciesDistance function

sp.DIST.P

a list obtained from the PRE_FATE.speciesDistance function run for Phanerophyte dominant species with mat.habitat, mat.overlap and sp.traits.P parameters

sp.DIST.C

a list obtained from the PRE_FATE.speciesDistance function run for Chamaephyte dominant species with mat.habitat, mat.overlap and sp.traits.C parameters

sp.DIST.H

a list obtained from the PRE_FATE.speciesDistance function run for Herbaceous dominant species with mat.habitat, mat.overlap and sp.traits.H parameters

PFG.species

data.frame of dimension 232 x 5
containing dominant species information relative to PFG
obtained from the PRE_FATE.speciesClustering_step2 function

PFG : name of assigned Plant Functional Group

DETERMINANT : is the species kept as determinant species within the PFG ?
(see PRE_FATE.speciesClustering_step2 function for details)

species : species ID

species_name : species name (taxonomic)

species_genus : species genus (taxonomic)

PFG.traits

data.frame of dimension 15 x 12
containing traits for plant functional groups
obtained from the PRE_FATE.speciesClustering_step3 function

PFG : Plant Functional Group short name

no.species : number of species within each group

maturity : MEAN age of maturity

longevity : MEAN age of lifespan

height : MEAN height (cm)

light : MEDIAN Landolt indicator value for light preference (from 1 to 5)

dispersal : MEDIAN classes (from 1 to 7) based on dispersal distances and types (Vittoz & Engler)

soil_contrib : MEAN Landolt indicator value for nitrogen preference (from 1 to 5)

soil_tol_min : MIN tolerance value for nitrogen preference (based on Landolt indicators, from 1 to 5)

soil_tol_max : MAX tolerance value for nitrogen preference (based on Landolt indicators, from 1 to 5)

LDMC : MEAN LDMC

LNC : MEAN LNC

PFG.PA

data.frame of dimension 13654 x 15
containing PFG occurrences and absences obtained from tab.dom.PA object and the PRE_FATE.speciesClustering_step3 function

#########################################################
>>> Champsaur_params
#########################################################

A list object with 12 elements to help building the simulation files and folders to run a FATE simulation :

name.dataset

Champsaur

tab.occ

a data.frame of dimension 13590 x 15
containing presence / absence (NA, 0 or 1) values per site for each PFG

tab.env

a data.frame of dimension 13590 x 5
containing environmental values (bio1, bio12, slope, dem, CESBIO2018) for each site

tab.xy

a data.frame of dimension 13590 x 2
containing coordinates for each site

stk.var

a stack object of dimension 358 x 427 with a resolution of 100m and ETRS89 projection, containing 3 layers with environmental values and to be used to produce PFG SDM

tab.SUCC

a data.frame of dimension 15 x 5
containing PFG characteristics (type, height, maturity, longevity)
to be used with the PRE_FATE.params_PFGsuccession function

tab.DISP

a data.frame of dimension 15 x 4
containing PFG characteristics (d50, d99, ldd)
to be used with the PRE_FATE.params_PFGdispersal function

tab.DIST

a data.frame of dimension 15 x 3
containing PFG characteristics (strategy_tol)
to be used with the PRE_FATE.params_PFGdisturbance function

tab.LIGHT

a data.frame of dimension 15 x 3
containing PFG characteristics (strategy_tol)
to be used with the PRE_FATE.params_PFGlight function

tab.SOIL

a data.frame of dimension 15 x 5
containing PFG characteristics (type, soil_contrib, soil_tol_min, soil_tol_max)
to be used with the PRE_FATE.params_PFGsoil function

stk.wmean

a stack object of dimension 358 x 427 with a resolution of 100m and ETRS89 projection, containing 15 layers with habitat suitability values (between 0 and 1) for each PFG

stk.mask

a stack object of dimension 358 x 427 with a resolution of 100m and ETRS89 projection, containing 3 mask layers with binary values (0 or 1) to be used in a FATE simulation :

Champsaur : simulation map, where occurs succession

noDisturb : perturbation map, when there is none

mowing : perturbation map, where occurs mowing

#########################################################
>>> Champsaur_results
#########################################################

Author

Maya Guéguen